% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPerCellQC.R
\name{addQCPerCell}
\alias{addQCPerCell}
\alias{addQCPerFeature}
\title{Add QC metrics}
\usage{
addQCPerCell(x, ...)

addQCPerFeature(x, ...)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} object or one of its subclasses.}

\item{...}{For \code{addQCPerCell}, further arguments to pass to \code{\link{perCellQCMetrics}}.

For \code{addQCPerFeature}, further arguments to pass to \code{\link{perFeatureQCMetrics}}.}
}
\value{
An object like \code{x} but with the QC metrics added to the row or column metadata.
}
\description{
Add per-feature or per-cell QC metrics to the row and column metadata, respectively, of a SummarizedExperiment object.
}
\details{
These are simply convenient functions that save the user from having to manually assign the QC metrics into SummarizedExperiment metadata.

Any QC metrics are appended onto the existing metadata fields. 
No protection is provided to avoid duplicated column names.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)

example_sce <- addQCPerCell(example_sce)
colData(example_sce)

example_sce <- addQCPerFeature(example_sce)
rowData(example_sce)

}
\seealso{
\code{\link{perCellQCMetrics}} and \code{\link{perFeatureQCMetrics}}, which do the actual work.
}
\author{
Aaron Lun
}
