% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{areSizeFactorsCentred}
\alias{areSizeFactorsCentred}
\title{Check if the size factors are centred at unity}
\usage{
areSizeFactorsCentred(object, centre = 1, tol = 1e-06)
}
\arguments{
\item{object}{an \code{SCESet} object containing multiple sets of
size factors.}

\item{centre}{a numeric scalar, the value around which all sets of
size factors should be centred.}

\item{tol}{a numeric scalar, the tolerance for testing equality of the
mean of each size factor set to \code{centre}.}
}
\value{
a \code{SCESet} object with centred size factors
}
\description{
Checks if each set of size factors is centred at unity, such that
abundances can be reasonably compared between features normalized
with different sets of size factors.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
keep_gene <- rowSums(counts(example_sceset)) > 0
example_sceset <- example_sceset[keep_gene,]

sizeFactors(example_sceset) <- runif(ncol(example_sceset))
areSizeFactorsCentred(example_sceset)
example_sceset <- normalize(example_sceset, centre=TRUE)
areSizeFactorsCentred(example_sceset)

}

