% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{arrange}
\alias{arrange}
\alias{arrange,SCESet-method}
\alias{arrange.SCESet}
\title{Arrange rows of \code{pData(object)} by variables.}
\usage{
arrange(object, ...)

\S4method{arrange}{SCESet}(object, ...)

arrange.SCESet(object, ...)
}
\arguments{
\item{object}{A \code{SCESet} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::arrange} to
act on \code{pData(object)}.}
}
\description{
The \code{SCESet} returned will have cells ordered by the corresponding
variable in \code{pData(object)}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- arrange(example_sceset, Cell_Cycle)
}

