% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calcAverage}
\alias{calcAverage}
\title{Calculate average counts, adjusting for size factors or library size}
\usage{
calcAverage(object, size.factors = NULL)
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object or a matrix of counts}

\item{size.factors}{numeric(), vector of size factors to use to scale library 
size in computation of counts-per-million. Extracted from the 
object if it is a \code{SingleCellExperiment} object; if object is a matrix, then 
if non-NULL, the provided size factors are used. Default is \code{NULL}, in which
case size factors are all set to 1 (i.e. library size adjustment only).}
}
\value{
Vector of average count values with same length as number of features.
}
\description{
Calculate average counts per feature, adjusting them as appropriate to take
into account for size factors for normalization or library sizes (total
counts).
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)

## calculate average counts
ave_counts <- calcAverage(example_sce)

}
