% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calcIsExprs}
\alias{calcIsExprs}
\title{Calculate which features are expressed in which cells using a threshold on
observed counts, transcripts-per-million, counts-per-million, FPKM, or
defined expression levels.}
\usage{
calcIsExprs(object, detection_limit = 0, exprs_values = "counts")
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object with expression 
and/or count data.}

\item{detection_limit}{numeric scalar giving the minimum expression level
for an expression observation in a cell for it to qualify as expressed.}

\item{exprs_values}{character scalar indicating whether the count data
(\code{"counts"}), the log-transformed count data (\code{"logcounts"}),
transcript-per-million (\code{"tpm"}), counts-per-million (\code{"cpm"}) or
FPKM (\code{"fpkm"}) should be used to define if an observation is expressed
or not. Defaults to the first available value of those options in the
order shown.}
}
\value{
a logical matrix indicating whether or not a feature in a particular
cell is expressed.
}
\description{
Calculate which features are expressed in which cells using a threshold on
observed counts, transcripts-per-million, counts-per-million, FPKM, or
defined expression levels.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
assay(example_sce, "is_exprs") <- calcIsExprs(example_sce, 
detection_limit = 1, exprs_values = "counts")
}
