% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calculateTPM}
\alias{calculateTPM}
\title{Calculate transcripts-per-million (TPM)}
\usage{
calculateTPM(object, effective_length = NULL, calc_from = "counts")
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object}

\item{effective_length}{vector of class \code{"numeric"} providing the
effective length for each feature in the \code{SingleCellExperiment} object}

\item{calc_from}{character string indicating whether to compute TPM from
\code{"counts"}, \code{"normcounts"} or \code{"fpkm"}.
Default is to use \code{"counts"}, in which case the \code{effective_length}
argument must be supplied.}
}
\value{
Matrix of TPM values.
}
\description{
Calculate transcripts-per-million (TPM) values for expression from counts for
a set of features.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
tpm(example_sce) <- calculateTPM(example_sce, effective_length = 5e04,
    calc_from = "counts")

## calculate from FPKM
fpkm(example_sce) <- calculateFPKM(example_sce, effective_length = 5e04,
use.size.factors = FALSE)
tpm(example_sce) <- calculateTPM(example_sce, effective_length = 5e04,
                                    calc_from = "fpkm")
}
