% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{cellNames}
\alias{cellNames}
\title{Get cell names from an SCESet object}
\usage{
cellNames(object)
}
\arguments{
\item{object}{An SCESet object.}
}
\value{
A vector of cell names.
}
\description{
Get cell names from an SCESet object
}
\details{
Simply a wrapper to \code{\link[Biobase]{sampleNames}}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
cellNames(example_sceset)

}

