% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{cellPairwiseDistances}
\alias{cellPairwiseDistances}
\alias{cellPairwiseDistances<-}
\alias{cellDist}
\alias{cellDist<-}
\alias{cellPairwiseDistances}
\alias{cellPairwiseDistances.SCESet}
\alias{cellPairwiseDistances,SCESet-method}
\alias{cellPairwiseDistances<-,SCESet,matrix-method}
\alias{cellPairwiseDistances<-,SCESet,dist-method}
\alias{cellDist,SCESet-method}
\alias{cellDist<-,SCESet,matrix-method}
\alias{cellDist<-,SCESet,dist-method}
\alias{cellPairwiseDistances,SCESet-method}
\alias{cellPairwiseDistances}
\alias{cellDistSCESet}
\alias{cellPairwiseDistances}
\alias{cellDist,SCESet-method}
\alias{cellPairwiseDistances}
\alias{cellPairwiseDistances<-}
\alias{cellPairwiseDistances}
\alias{cellPairwiseDistances<-}
\alias{cellPairwiseDistances}
\alias{cellDist<-}
\alias{cellPairwiseDistances}
\alias{cellDist<-}
\alias{cellPairwiseDistances}
\title{cellPairwiseDistances in an SCESet object}
\usage{
cellPairwiseDistances(object)

cellPairwiseDistances(object) <- value

cellDist(object)

cellDist(object) <- value

cellPairwiseDistances.SCESet(object)

\S4method{cellPairwiseDistances}{SCESet}(object)

cellDistSCESet(object)

\S4method{cellDist}{SCESet}(object)

\S4method{cellPairwiseDistances}{SCESet,matrix}(object) <- value

\S4method{cellPairwiseDistances}{SCESet,dist}(object) <- value

\S4method{cellDist}{SCESet,matrix}(object) <- value

\S4method{cellDist}{SCESet,dist}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"} containing cell pairwise
distances}
}
\value{
An SCESet object containing new cell pairwise distances matrix.
}
\description{
SCESet objects can contain a matrix of pairwise distances between cells. These
 functions conveniently access and replace the cell pairwise distances with the value
 supplied, which must be a matrix of the correct size. The function \code{cellDist}
 is simply shorthand for \code{cellPairwiseDistances}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
cellPairwiseDistances(example_sceset)

}
\author{
Davis McCarthy
}
