% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{computeLibSizeFactors}
\alias{computeLibSizeFactors}
\title{Compute library size factors}
\usage{
computeLibSizeFactors(object, exprs_values = "counts")
}
\arguments{
\item{object}{A count matrix or SingleCellExperiment object containing counts.}

\item{exprs_values}{A string indicating the assay of \code{object} containing
the counts, if \code{object} is a SingleCellExperiment.}
}
\value{
A numeric vector of size factors.
}
\description{
Define size factors from the library sizes after centering.
}
\examples{
data("sc_example_counts")
summary(computeLibSizeFactors(sc_example_counts))

}
