% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{counts}
\alias{counts}
\alias{counts,SCESet-method}
\alias{counts.SCESet}
\alias{counts<-}
\alias{counts<-,SCESet,matrix-method}
\title{Accessors for the 'counts' element of an SCESet object.}
\usage{
counts(object)

counts(object) <- value

\S4method{counts}{SCESet}(object)

\S4method{counts}{SCESet,matrix}(object)<-value

\S4method{counts}{SCESet}(object)

\S4method{counts}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{an integer matrix}
}
\value{
a matrix of count data
}
\description{
The counts element holds the count data as a matrix of non-negative integer
count values, one row for each feature (gene, exon, region, etc), and one
column for each cell. It is an element of the assayData slot of the SCESet
object.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
counts(example_sceset)

}
\author{
Davis McCarthy
}

