% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{cpm}
\alias{cpm}
\alias{cpm<-}
\alias{cpm}
\alias{cpmSCESet}
\alias{cpm,SCESet-method}
\alias{cpm<-,SCESet,matrix-method}
\alias{cpm}
\alias{cpm,SCESet-method}
\alias{cpm<-}
\alias{cpm<-,SCESet,matrix-method}
\title{Accessors for the 'cpm' (counts per million) element of an SCESet object.}
\usage{
cpm(object)

cpm(object) <- value

\S4method{cpm}{SCESet}(object)

\S4method{cpm}{SCESet,matrix}(object)<-value

\S4method{cpm}{SCESet}(object)

\S4method{cpm}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"}}
}
\value{
a matrix of counts-per-million values
}
\description{
The \code{cpm} element of the arrayData slot in an SCESet object holds
a matrix containing counts-per-million values. It has the same dimensions
as the 'exprs' and 'counts' elements, which hold the transformed expression
data and count data, respectively.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData=sc_example_counts)
cpm(example_sceset)[1:10, 1:6]
}
\author{
Davis McCarthy
}
