% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/defunct.R, R/runDiffusionMap.R
\name{calculateDiffusionMap}
\alias{calculateDiffusionMap}
\alias{defunct}
\alias{calculateQCMetrics}
\alias{normalize,SingleCellExperiment-method}
\alias{centreSizeFactors}
\alias{calculateDiffusionMap,ANY-method}
\alias{runDiffusionMap}
\title{Defunct functions}
\usage{
calculateDiffusionMap(x, ...)

calculateQCMetrics(...)

\S4method{normalize}{SingleCellExperiment}(object, ...)

centreSizeFactors(...)

\S4method{calculateDiffusionMap}{ANY}(x, ...)

runDiffusionMap(x, ...)
}
\arguments{
\item{object, ...}{Ignored arguments.}
}
\value{
All functions error out with a defunct message pointing towards its descendent (if available).
}
\description{
Functions that have passed on to the function afterlife.
Their successors are also listed.
}
\details{
\code{calculateQCMetrics} is succeeded by \code{\link{perCellQCMetrics}} and \code{\link{perFeatureQCMetrics}}.

\code{normalize} is succeeded by \code{\link{logNormCounts}}.

\code{centreSizeFactors} has no replacement - the \pkg{SingleCellExperiment} is removing support for multiple size factors, so this function is now trivial.

\code{runDiffusionMap} has no replacement. \pkg{destiny} is no longer on 
Bioconductor. You can calculate a diffusion map using \pkg{destiny} yourself,
and add it to a \code{reducedDim} field if you so wish.
}
\examples{
try(calculateQCMetrics())
}
\author{
Aaron Lun
}
