% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{deprecated}
\alias{deprecated}
\alias{getExprs}
\alias{deprecated}
\alias{toCellDataSet}
\alias{deprecated}
\alias{fromCellDataSet}
\title{Retrieve a representation of gene expression}
\usage{
getExprs(object)

toCellDataSet(sce, exprs_values = "exprs")

fromCellDataSet(cds, exprs_values = "tpm", logged = FALSE,
  logExprsOffset = 1)
}
\arguments{
\item{object}{An object of type \code{SCESet}}

\item{sce}{An \code{SCESet} object}

\item{exprs_values}{What should \code{exprs(cds)} be mapped from in the
\code{SCESet}? Should be one of "exprs", "tpm", "fpkm", "counts"}

\item{cds}{A \code{CellDataSet} from the \code{monocle} package}

\item{logged}{logical, if \code{exprs_values="exprs"}, are
the expression values already on the log2 scale, or not?}

\item{logExprsOffset}{numeric, value to add prior to log-transformation.}
}
\value{
A matrix representation of expression values.

An object of class \code{CellDataSet}

An object of class \code{SCESet}
}
\description{
Deprecated from scater version 1.3.29.

Deprecated from scater version 1.5.2.

Deprecated from scater version 1.5.2.
}
\examples{
\dontrun{"Deprecated"}
\dontrun{"Deprecated"}
}
