% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10ximport-wrapper.R
\name{downsampleCounts}
\alias{downsampleCounts}
\title{Downsample a count matrix}
\usage{
downsampleCounts(x, prop)
}
\arguments{
\item{x}{matrix of counts}

\item{prop}{numeric scalar or vector of length \code{ncol(x)} in [0, 1] 
indicating the downsampling proportion}
}
\value{
an integer matrix of downsampled counts
}
\description{
Downsample a count matrix to a desired proportion.
}
\details{
Given multiple 10X batches of very different sequencing depths, it
can be beneficial to downsample the deepest batches to match the coverage of
the shallowest batches. This avoids differences in technical noise that can
drive clustering by batch. 

Downsampling without replacement is performed on the counts in each cell to
generate the output matrix. Each count in the returned matrix is guaranteed
to be smaller than the original value in \code{x}. This provides an 
alternative to downsampling in the CellRanger \code{aggr} function.
}
\examples{
sce10x <- read10xResults(system.file("extdata", package="scater"))
downsampled <- downsampleCounts(counts(sce10x), prop = 0.5)

}
