% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{fData<-}
\alias{fData<-}
\title{Replacement for featureData}
\usage{
fData(x) <- value
}
\arguments{
\item{x}{object containing featureData to replace}

\item{value}{AnnotatedDataFrame to replace existing featureData}
}
\value{
the SCESet with replaced featureData
}
\description{
Replacement for featureData
}

