% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\docType{methods}
\name{fData<-,SCESet,AnnotatedDataFrame-method}
\alias{fData}
\alias{fData,SCESet-method}
\alias{fData<-,SCESet,AnnotatedDataFrame-method}
\alias{fData<-,SCESet,data.frame-method}
\title{Replaces featureData in an SCESet object}
\usage{
\S4method{fData}{SCESet,AnnotatedDataFrame}(x) <- value

\S4method{fData}{SCESet,data.frame}(x) <- value
}
\arguments{
\item{x}{An SCESet object.}

\item{value}{an AnnotatedDataFrame with updated featureData to replace
existing}
}
\value{
A matrix of expression count data, where rows correspond to features
(e.g. genes) and columns correspond to cells.
}
\description{
SCESet objects contain feature information (inherited from the ExpressionSet
class). This function conveniently replaces the feature data with the
value supplied, which must be an AnnotatedDataFrame.
}
\examples{
\dontrun{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
fData(example_sceset)
}
}

