% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{featurePairwiseDistances}
\alias{featDist}
\alias{featDist,SCESet-method}
\alias{featDist<-}
\alias{featDist<-,SCESet,matrix-method}
\alias{featDistSCESet}
\alias{featurePairwiseDistances}
\alias{featurePairwiseDistances,SCESet-method}
\alias{featurePairwiseDistances<-}
\alias{featurePairwiseDistances<-,SCESet,matrix-method}
\alias{featurePairwiseDistancesSCESet}
\title{featurePairwiseDistances in an SCESet object}
\usage{
featurePairwiseDistances(object)

featurePairwiseDistances(object) <- value

featDist(object)

featDist(object) <- value

featurePairwiseDistancesSCESet(object)

\S4method{featurePairwiseDistances}{SCESet}(object)

featDistSCESet(object)

\S4method{featDist}{SCESet}(object)

\S4method{featurePairwiseDistances}{SCESet,matrix}(object) <- value

\S4method{featDist}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"} containing feature pairwise
distances}
}
\value{
An SCESet object containing new feature pairwise distances matrix.
}
\description{
SCESet objects can contain a matrix of pairwise distances between features
(e.g. genes, transcripts). These functions conveniently access and replace
the gene pairwise distances with the value supplied, which must be a matrix
of the correct size. The function \code{featDist} is simply shorthand for
\code{featurePairwiseDistances}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
featurePairwiseDistances(example_sceset)

}
\author{
Davis McCarthy
}

