% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{fpkm}
\alias{fpkm}
\alias{fpkm,SCESet-method}
\alias{fpkm.SCESet}
\alias{fpkm<-}
\alias{fpkm<-,SCESet,matrix-method}
\title{Accessors for the 'fpkm' (fragments per kilobase of exon per million reads mapped) element of an SCESet object.}
\usage{
fpkm(object)

fpkm(object) <- value

\S4method{fpkm}{SCESet}(object)

\S4method{fpkm}{SCESet,matrix}(object)<-value

\S4method{fpkm}{SCESet}(object)

\S4method{fpkm}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"}}
}
\value{
a matrix of FPKM values
}
\description{
The \code{fpkm} element of the arrayData slot in an SCESet object holds
a matrix containing fragments per kilobase of exon per million reads mapped
(FPKM) values. It has the same dimensions as the 'exprs' and 'counts'
elements, which hold the transformed expression data and count data,
respectively.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
fpkm(example_sceset)

}
\author{
Davis McCarthy
}

