% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{isSpike}
\alias{isSpike}
\alias{isSpike}
\alias{isSpike,SCESet-method}
\title{Get spike-in features in an SCESet object}
\usage{
isSpike(object, ...)

\S4method{isSpike}{SCESet}(object, type = NULL, warning = TRUE)
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{...}{arguments passed through generic version of the function.}

\item{type}{a character vector specifying the feature control sets to use. All 
specified spike-in sets in \code{featureControlInfo(object)} are used by default.}

\item{warning}{A logical scalar specifying if a warning should be raised 
if spike-in controls are unavailable.}
}
\value{
A logical vector specifying if each row is a spike-in feature.
}
\description{
Get the features in the SCESet object that are spike-in controls, as 
specified using \code{\link{setSpike}}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset,
                            feature_controls = list(ERCC = 1:40, Mito=41:50))
setSpike(example_sceset) <- "ERCC"
summary(isSpike(example_sceset))
}
\author{
Aaron Lun
}
