% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarySizeFactors.R
\name{librarySizeFactors}
\alias{librarySizeFactors}
\title{Compute library size factors}
\usage{
librarySizeFactors(object, exprs_values = "counts", subset_row = NULL)
}
\arguments{
\item{object}{A count matrix or SingleCellExperiment object containing counts.}

\item{exprs_values}{A string indicating the assay of \code{object} containing the counts, if \code{object} is a SingleCellExperiment.}

\item{subset_row}{A vector specifying whether the rows of \code{object} should be (effectively) subsetted before calculating library sizes.}
}
\value{
A numeric vector of size factors.
}
\description{
Define size factors from the library sizes after centering.
This ensures that the library size adjustment yields values comparable to those generated after normalization with other sets of size factors.
}
\examples{
data("sc_example_counts")
summary(librarySizeFactors(sc_example_counts))

}
