% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logNormCounts.R
\name{logNormCounts}
\alias{logNormCounts}
\title{Compute log-normalized expression values}
\usage{
logNormCounts(x, ...)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} or \linkS4class{SummarizedExperiment} object containing a count matrix.}

\item{size.factors}{A numeric vector of cell-specific size factors.
Alternatively \code{NULL}, in which case the size factors are extracted or computed from \code{x}.}

\item{assay.type}{String or integer scalar indicating which assay contains the count data.}

\item{log}{Logical scalar indicating whether normalized values should be log2-transformed.}

\item{pseudo.count}{Numeric scalar specifying the pseudo-count to add when log-transforming expression values.}

\item{center.sf}{Logical scalar indicating whether size fators should be centred.}

\item{use.alt.exps}{Logical scalar indicating whether normalization should be performed for alternative experiments in \code{x}.
If \code{TRUE}, statistics are computed for all alternative experiments. 

Alternatively, an integer or character vector specifying the alternative Experiments to use to compute QC statistics.

Alternatively, \code{NULL} in which case alternative experiments are not used.}

\item{name}{String containing an assay name for storing the output normalized values.
Defaults to \code{"logcounts"} when \code{log=TRUE} and \code{"normcounts"} otherwise.}
}
\value{
\code{x} is returned containing the (log-)normalized expression values in an additional assay named as \code{name}.

If \code{x} is a \linkS4class{SingleCellExperiment}, the size factors used for normalization are stored in \code{\link{sizeFactors}}.
These are centered if \code{center.sf=TRUE}.

Morevoer, if there are alternative experiments and \code{use.alt.exps} is specified appropriately,
each of the alternative experiments in \code{x} will also contain an additional assay.
}
\description{
Compute log-transformed normalized expression values from a count matrix in a \linkS4class{SingleCellExperiment} object.
}
\details{
This function is a convenience wrapper around \code{\link{normalizeCounts}}.
It returns a \linkS4class{SingleCellExperiment} or \linkS4class{SummarizedExperiment} containing the normalized values in a separate assay.
This makes it easier to perform normalization by avoiding book-keeping errors during a long analysis workflow.

If \code{x} is a \linkS4class{SingleCEllExperiment} that contains alternative experiments, normalized values are computed and stored within each alternative experiment.
If \code{size.factors=NULL}, size factors are obtained separately for each nested experiment following the rules in \code{\link{normalizeCounts}}.
However, if \code{size.factors} is supplied, it will override any size factors available in the alternative experiments.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)

example_sce <- logNormCounts(example_sce)

}
\seealso{
\code{\link{normalizeCounts}}, which is used to compute the normalized expression values.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
