% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{mutate}
\alias{mutate}
\alias{mutate,SingleCellExperiment-method}
\title{Add new variables to \code{colData(object)} (deprecated).}
\usage{
mutate(object, ...)

\S4method{mutate}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::mutate} to
act on \code{colData(object)}.}
}
\value{
An SingleCellExperiment object.
}
\description{
Add new variables to \code{colData(object)} (deprecated).
}
\details{
Refer to \url{https://github.com/sa-lee/plyexperiment} for replacement functionality.
}
