% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{newSCESet}
\alias{newSCESet}
\title{Create a new SCESet object}
\usage{
newSCESet(exprsData = NULL, countData = NULL, tpmData = NULL,
  fpkmData = NULL, cpmData = NULL, phenoData = NULL, featureData = NULL,
  experimentData = NULL, is_exprsData = NULL,
  cellPairwiseDistances = dist(vector()),
  featurePairwiseDistances = dist(vector()), lowerDetectionLimit = NULL,
  logExprsOffset = NULL)
}
\arguments{
\item{exprsData}{expression data matrix for an experiment (features x cells)}

\item{countData}{data matrix containing raw count expression values}

\item{tpmData}{matrix of class \code{"numeric"} containing
transcripts-per-million (TPM) expression values}

\item{fpkmData}{matrix of class \code{"numeric"} containing fragments per
kilobase of exon per million reads mapped (FPKM) expression values}

\item{cpmData}{matrix of class \code{"numeric"} containing counts per
million (CPM) expression values (optional)}

\item{phenoData}{data frame containing attributes of individual cells}

\item{featureData}{data frame containing attributes of features (e.g. genes)}

\item{experimentData}{MIAME class object containing metadata data and details
about the experiment and dataset.}

\item{is_exprsData}{matrix of class \code{"logical"}, indicating whether
or not each observation is above the \code{lowerDetectionLimit}.}

\item{cellPairwiseDistances}{object of class \code{"dist"} (or a class that
extends "dist") containing cell-cell distance or dissimilarity values.}

\item{featurePairwiseDistances}{object of class \code{"dist"} (or a class that
extends "dist") containing feature-feature distance or dissimilarity values.}

\item{lowerDetectionLimit}{the minimum expression level that constitutes true
expression (defaults to zero and uses count data to determine if an
observation is expressed or not).}

\item{logExprsOffset}{numeric scalar, providing the offset used when doing
log2-transformations of expression data to avoid trying to take logs of zero.
Default offset value is \code{1}.}
}
\value{
a \code{\link{SingleCellExperiment}} object
}
\description{
Deprecated from scater version 1.3.29; the package now uses the 
\code{\link{SingleCellExperiment}} class. To convert an SCESet object to 
SingleCellExperiment see the \code{\link{toSingleCellExperiment}} function. 
This function is retained for backwards compatibility.
}
\details{
This function now returns a \code{\link{SingleCellExperiment}} object, 
whereas earlier versions produced an \code{SCESet} object. The \code{scater}
package now uses \code{SingleCellExperiment} as its data structure instead
of \code{SCESet}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
\dontrun{
example_sce <- newSCESet(countData = sc_example_counts, phenoData = pd)
}
}
