% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{norm_counts}
\alias{norm_counts}
\alias{norm_counts,SCESet-method}
\alias{norm_counts.SCESet}
\alias{norm_counts<-}
\alias{norm_counts<-,SCESet,matrix-method}
\title{Accessors for the 'norm_counts' element of an SCESet object.}
\usage{
norm_counts(object)

norm_counts(object) <- value

\S4method{norm_counts}{SCESet}(object)

\S4method{norm_counts}{SCESet,matrix}(object)<-value

\S4method{norm_counts}{SCESet}(object)

\S4method{norm_counts}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{an integer matrix}
}
\value{
a matrix of normalised count data
}
\description{
The norm_counts element holds normalised count data as a matrix of
non-negative values, one row for each feature (gene, exon, region, etc), and one
column for each cell. It is an element of the assayData slot of the SCESet
object.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
norm_counts(example_sceset)

}
\author{
Davis McCarthy
}

