% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{norm_exprs}
\alias{norm_exprs}
\alias{norm_exprs<-}
\alias{norm_exprs}
\alias{norm_exprs.SCESet}
\alias{norm_exprs,SCESet-method}
\alias{norm_exprs<-,SCESet,matrix-method}
\alias{norm_exprs,SCESet-method}
\alias{norm_exprs<-}
\title{Accessors for the 'norm_exprs' (normalised expression) element of an SCESet object.}
\usage{
norm_exprs(object)

norm_exprs(object) <- value

\S4method{norm_exprs}{SCESet}(object)

\S4method{norm_exprs}{SCESet,matrix}(object)<-value

\S4method{norm_exprs}{SCESet}(object)

\S4method{norm_exprs}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{an integer matrix}
}
\value{
a matrix of normalised expression data
}
\description{
The \code{norm_exprs} element of the arrayData slot in an SCESet object holds
a matrix containing normalised expression values. It has the same dimensions
as the 'exprs' and 'counts' elements, which hold the transformed expression
data and count data, respectively.
}
\details{
The default for normalised expression values is mean-centred and
variance-standardised expression data from the \code{exprs} slot of the
\code{SCESet} object. The function \code{normaliseExprs} (or
\code{normalizeExprs}) provides more options and functionality for
normalising expression data.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
norm_exprs(example_sceset)

}
\author{
Davis McCarthy
}
