% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{norm_fpkm}
\alias{norm_fpkm}
\alias{norm_fpkm<-}
\alias{norm_fpkm}
\alias{norm_fpkm.SCESet}
\alias{norm_fpkm,SCESet-method}
\alias{norm_fpkm<-,SCESet,matrix-method}
\alias{norm_fpkm}
\alias{norm_fpkm,SCESet-method}
\alias{norm_fpkm<-}
\alias{norm_fpkm<-,SCESet,matrix-method}
\title{Accessors for the 'norm_fpkm' (normalised fragments per kilobase of exon per million reads mapped) element of an SCESet object.}
\usage{
norm_fpkm(object)

norm_fpkm(object) <- value

\S4method{norm_fpkm}{SCESet}(object)

\S4method{norm_fpkm}{SCESet,matrix}(object)<-value

\S4method{norm_fpkm}{SCESet}(object)

\S4method{norm_fpkm}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"}}
}
\value{
a matrix of normalised FPKM data
}
\description{
The \code{norm_fpkm} element of the arrayData slot in an SCESet object holds
a matrix containing normalised fragments per kilobase of exon per million
reads mapped (FPKM) values. It has the same dimensions as the 'exprs' and
'counts' elements, which hold the transformed expression data and count data,
respectively.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
norm_fpkm(example_sceset)

}
\author{
Davis McCarthy
}
