% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perCellDataFrameFromSCE.R
\name{perCellDataFrameFromSCE}
\alias{perCellDataFrameFromSCE}
\title{Create a per-cell data.frame}
\usage{
perCellDataFrameFromSCE(
  x,
  ...,
  exprs_values = "logcounts",
  include_dimred = NULL,
  ncomponents = 2,
  include_size_factors = FALSE
)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.}

\item{...}{Any number of elements, each following the format of the \code{by} argument in \code{\link{retrieveCellInfo}}.

Briefly, each element can be a single string, in which case it is assumed to refer to a \code{\link{colData}} column,
a row of \code{\link{assay}(x, exprs_values)}, or a row of one of the \code{\link{altExps}(x)}.

Alternatively, each element can be an \link{AsIs}-wrapped vector, in which case it is used directly.
In this case, the element is expected to be named in \code{...}.}

\item{exprs_values}{String or integer scalar specifying the assay from which expression values should be extracted.}

\item{include_dimred}{Character vector containing the names of dimensionality reduction results to include.}

\item{ncomponents}{Integer scalar indicating the number of components to return if \code{include_dimred} is set.}

\item{include_size_factors}{Logical scalar indicating whether size factors should be included in the output.}
}
\value{
A data.frame containing the requested fields,
named according to the names in \code{...} or the values themselves (if the values are strings and no names are supplied).
Each row corresponds to a cell (i.e., column) of \code{x},
}
\description{
Create a per-cell data.frame from a \linkS4class{SingleCellExperiment},
most typically for creating custom \pkg{ggplot2} plots.
}
\details{
This function enables us to conveniently create a data.frame from a \linkS4class{SingleCellExperiment},
ostensibly to put in a \code{ggplot} command.
The user can then use this to create custom plots that are not covered by \code{\link{plotExpression}} and related functions.

If \code{include_dimred} is set,
each set of dimensionality reduction results is named by appending the component number to the name of the result,
e.g., \code{"PCA.1"}, \code{"PCA.2"}.
Note that \code{n_dimred} is applied to all results listed in \code{include_dimred}.

If \code{include_size_factors=TRUE}, an extra numeric \code{size_factor} column is added to the output
if there are any size factors present in \code{x}.

The same \code{exprs_values} is used for all assay-related extractions in \code{...}.
See \code{\link{retrieveCellInfo}} for more details.
}
\examples{
example_sce <- mockSCE()
example_sce <- logNormCounts(example_sce)

df1 <- perCellDataFrameFromSCE(example_sce, "Mutation_Status", "Gene_0001")
head(df1)

df2 <- perCellDataFrameFromSCE(example_sce, "Mutation_Status", 
    stuff="Cell_Cycle", other_stuff="Gene_0002")
head(df2)

example_sce <- runPCA(example_sce)
df3 <- perCellDataFrameFromSCE(example_sce, "Mutation_Status", 
    stuff=I(runif(ncol(example_sce))), include_dimred="PCA")
head(df3)

}
\seealso{
\code{\link{retrieveCellInfo}}, which powers this function.
}
\author{
Aaron Lun
}
