% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/plotting.R
\docType{methods}
\name{plotDiffusionMap}
\alias{plotDiffusionMap}
\alias{plotDiffusionMap,SCESet-method}
\alias{plotDiffusionMapSCESet}
\title{Plot a diffusion map for an SCESet object}
\usage{
plotDiffusionMap(object, ...)

plotDiffusionMapSCESet(object, ntop = 500, ncomponents = 2,
  exprs_values = "exprs", colour_by = NULL, shape_by = NULL,
  size_by = NULL, feature_set = NULL, return_SCESet = FALSE,
  scale_features = TRUE, draw_plot = TRUE, theme_size = 10,
  rand_seed = NULL, sigma = NULL, distance = "euclidean", legend = TRUE,
  ...)

\S4method{plotDiffusionMap}{SCESet}(object, ntop = 500, ncomponents = 2,
  exprs_values = "exprs", colour_by = NULL, shape_by = NULL,
  size_by = NULL, feature_set = NULL, return_SCESet = FALSE,
  scale_features = FALSE, draw_plot = TRUE, theme_size = 10,
  rand_seed = NULL, sigma = NULL, distance = "euclidean", legend = TRUE,
  ...)
}
\arguments{
\item{object}{an \code{SCESet} object}

\item{...}{further arguments passed to \code{\link[destiny]{DiffusionMap}}}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the diffusion map. Default is \code{500}, but any \code{ntop} 
argument is overrided if the \code{feature_set} argument is non-NULL.}

\item{ncomponents}{numeric scalar indicating the number of principal
components to plot, starting from the first diffusion map component. Default 
is 2. If \code{ncomponents} is 2, then a scatterplot of component 1 vs 
component 2 is produced. If \code{ncomponents} is greater than 2, a pairs 
plots for the top components is produced. NB: computing many components for 
the diffusion map can become time consuming.}

\item{exprs_values}{character string indicating which values should be used
as the expression values for this plot. Valid arguments are \code{"tpm"}
(default; transcripts per million), \code{"norm_tpm"} (normalised TPM
values), \code{"fpkm"} (FPKM values), \code{"norm_fpkm"} (normalised FPKM
values), \code{"counts"} (counts for each feature), \code{"norm_counts"},
\code{"cpm"} (counts-per-million), \code{"norm_cpm"} (normalised
counts-per-million), \code{"exprs"} (whatever is in the \code{'exprs'} slot
of the \code{SCESet} object; default), \code{"norm_exprs"} (normalised
expression values) or \code{"stand_exprs"} (standardised expression values)
or any other named element of the \code{assayData} slot of the \code{SCESet}
object that can be accessed with the \code{get_exprs} function.}

\item{colour_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to colour the points in the plot.}

\item{shape_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to define the shape of the points in the plot.}

\item{size_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to define the size of points in the plot.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the diffusion map. If character, entries must all be in
\code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{return_SCESet}{logical, should the function return an \code{SCESet}
object with principal component values for cells in the
\code{reducedDimension} slot. Default is \code{FALSE}, in which case a
\code{ggplot} object is returned.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{draw_plot}{logical, should the plot be drawn on the current graphics
device? Only used if \code{return_SCESet} is \code{TRUE}, otherwise the plot
is always produced.}

\item{theme_size}{numeric scalar giving default font size for plotting theme
(default is 10).}

\item{rand_seed}{(optional) numeric scalar that can be passed to
\code{set.seed} to make plots reproducible.}

\item{sigma}{argument passed to \code{\link[destiny]{DiffusionMap}}}

\item{distance}{argument passed to \code{\link[destiny]{DiffusionMap}}}

\item{legend}{logical, should the legend(s) be shown? Default is \code{TRUE}.}
}
\value{
If \code{return_SCESet} is \code{TRUE}, then the function returns an
\code{SCESet} object, otherwise it returns a \code{ggplot} object.
}
\description{
Produce a diffusion map plot of two components for an \code{SCESet} dataset.
}
\details{
The function \code{\link[destiny]{DiffusionMap}} is used internally 
to compute the diffusion map.
}
\examples{
## Set up an example SCESet
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
drop_genes <- apply(exprs(example_sceset), 1, function(x) {var(x) == 0})
example_sceset <- example_sceset[!drop_genes, ]

## Examples plotting diffusion maps
plotDiffusionMap(example_sceset)
plotDiffusionMap(example_sceset, colour_by = "Cell_Cycle")
plotDiffusionMap(example_sceset, colour_by = "Cell_Cycle",
shape_by = "Treatment")
plotDiffusionMap(example_sceset, colour_by = "Cell_Cycle",
shape_by = "Treatment", size_by = "Mutation_Status")
plotDiffusionMap(example_sceset, shape_by = "Treatment",
size_by = "Mutation_Status")
plotDiffusionMap(example_sceset, feature_set = 1:100, colour_by = "Treatment",
shape_by = "Mutation_Status")

plotDiffusionMap(example_sceset, shape_by = "Treatment",
return_SCESet = TRUE)


}
\references{
Haghverdi L, Buettner F, Theis FJ. Diffusion maps for high-dimensional single-cell analysis of differentiation data. Bioinformatics. 2015; doi:10.1093/bioinformatics/btv325
}
\seealso{
\code{\link[destiny]{destiny}}
}

