% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotMDS}
\alias{plotMDS}
\alias{plotMDS,SingleCellExperiment-method}
\title{Produce a multidimensional scaling plot for a SingleCellExperiment object}
\usage{
plotMDS(object, ..., ncomponents = 2, return_SCE = FALSE, rerun = FALSE,
  draw_plot = TRUE, run_args = list())
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object}

\item{...}{Additional arguments to pass to \code{\link{plotReducedDim}}.}

\item{ncomponents}{numeric scalar indicating the number of principal
components to plot, starting from the first principal component. Default is
2. If \code{ncomponents} is 2, then a scatterplot of PC2 vs PC1 is produced.
If \code{ncomponents} is greater than 2, a pairs plots for the top components
is produced. NB: computing more than two components for t-SNE can become very
time consuming.}

\item{return_SCE}{logical, should the function return a \code{SingleCellExperiment}
object with principal component values for cells in the
\code{reducedDims} slot. Default is \code{FALSE}, in which case a
\code{ggplot} object is returned.}

\item{rerun}{logical, should PCA be recomputed even if \code{object} contains a
\code{"MDS"} element in the \code{reducedDims} slot?}

\item{draw_plot}{logical, should the plot be drawn on the current graphics
device? Only used if \code{return_SCE} is \code{TRUE}, otherwise the plot
is always produced.}

\item{run_args}{Arguments to pass to \code{\link{runMDS}} when \code{rerun=TRUE}
or if there is no existing \code{"MDS"} element in the \code{reducedDims} slot.}
}
\value{
If \code{return_SCE} is \code{TRUE}, then the function returns an
\code{SingleCellExperiment} object, otherwise it returns a \code{ggplot} object.
}
\description{
Produce an MDS plot from the cell pairwise distance data in a
\code{SingleCellExperiment} object.
}
\details{
The function \code{\link{cmdscale}} is used internally to
compute the multidimensional scaling components to plot.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)
drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]

## Examples plotting
plotMDS(example_sce)
plotMDS(example_sce, colour_by = "Cell_Cycle")
plotMDS(example_sce, colour_by = "Cell_Cycle", shape_by = "Treatment")

## define cell-cell distances differently
plotMDS(example_sce, colour_by = "Cell_Cycle",
shape_by = "Treatment", size_by = "Mutation_Status", method = "canberra")

}
\seealso{
\code{\link{runMDS}}
}
