% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPlatePosition}
\alias{plotPlatePosition}
\title{Plot cells in plate positions}
\usage{
plotPlatePosition(object, plate_position = NULL, colour_by = NULL,
  x_position = NULL, y_position = NULL, theme_size = 24,
  legend = "auto")
}
\arguments{
\item{object}{an \code{SCESet} object. If \code{object$plate_position} is not
\code{NULL}, then this will be used to define each cell's position on the 
plate, unless the \code{plate_position} argument is specified.}

\item{plate_position}{optional character vector providing a position on the
plate for each cell (e.g. A01, B12, etc, where letter indicates row and 
number indicates column). Specifying this argument overrides any plate 
position information extracted from the SCESet object.}

\item{colour_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to colour the points in the plot.}

\item{x_position}{numeric vector providing x-axis positions for the cells
(ignored if \code{plate_position} is not \code{NULL})}

\item{y_position}{numeric vector providing y-axis positions for the cells
(ignored if \code{plate_position} is not \code{NULL})}

\item{theme_size}{numeric scalar giving default font size for plotting theme
(default is 10).}

\item{legend}{character, specifying how the legend(s) be shown? Default is 
\code{"auto"}, which hides legends that have only one level and shows others. 
Alternatives are "all" (show all legends) or "none" (hide all legends).}
}
\description{
Plots cells in their position on a plate, coloured by phenotype data or 
feature expression.
}
\details{
This function expects plate positions to be given in a charcter 
format where a letter indicates the row on the plate and a numeric value 
indicates the column. So each cell has a plate position such as "A01", "B12",
"K24" and so on. From these plate positions, the row is extracted as the 
letter, and the column as the numeric part. If \code{object$plate_position} 
or the \code{plate_position} argument are used to define plate positions, 
then positions should be provided in this format. Alternatively, numeric 
values to be used as x- and y-coordinates by supplying both the 
\code{x_position} and \code{y_position} arguments to the function.
}
\examples{
## prepare data
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset)

## define plate positions
example_sceset$plate_position <- paste0(
rep(LETTERS[1:5], each = 8), rep(formatC(1:8, width = 2, flag = "0"), 5))

## plot plate positions
plotPlatePosition(example_sceset, colour_by = "Mutation_Status")

plotPlatePosition(example_sceset, colour_by = "Gene_0004")

}

