% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQC}
\alias{plotQC}
\title{Produce QC diagnostic plots}
\usage{
plotQC(object, type = "highest-expression", ...)
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{type}{character scalar providing type of QC plot to compute: 
"highest-expression" (showing features with highest expression), "find-pcs" (showing
the most important principal components for a given variable), 
"explanatory-variables" (showing a set of explanatory variables plotted 
against each other, ordered by marginal variance explained), or 
"exprs-mean-vs-freq" (plotting the mean expression levels against the 
frequency of expression for a set of features).}

\item{...}{arguments passed to \code{plotHighestExprs}, 
\code{plotImportantPCs}, \code{plotExplanatoryVariables} and 
\code{plotExprsMeanVsFreq} as appropriate.}
}
\value{
a ggplot plot object
}
\description{
Produce QC diagnostic plots
}
\details{
Calculate useful quality control metrics to help with pre-processing
of data and identification of potentially problematic features and cells.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data=sc_example_cell_info)
rownames(pd) <- pd$Cell
example_sceset <- newSCESet(countData=sc_example_counts, phenoData=pd)
drop_genes <- apply(exprs(example_sceset), 1, function(x) {var(x) == 0})
example_sceset <- example_sceset[!drop_genes, ]
example_sceset <- calculateQCMetrics(example_sceset)
plotQC(example_sceset, type="high", col_by_variable="Mutation_Status")
plotQC(example_sceset, type="find", variable="total_features")
vars <- names(pData(example_sceset))[c(2:3, 5:14)]
plotQC(example_sceset, type="expl", variables=vars)

}

