% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_reddim.R
\docType{methods}
\name{plotRDSCE}
\alias{plotRDSCE}
\alias{plotPCASCE}
\alias{plotPCA}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotTSNE}
\alias{plotDiffusionMap}
\alias{plotMDS}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotPCA}
\alias{plotTSNE}
\alias{plotDiffusionMap}
\alias{plotMDS}
\title{Plot specific reduced dimensions for a SingleCellExperiment object}
\usage{
plotPCASCE(object, ..., return_SCE = FALSE, draw_plot = TRUE,
  rerun = FALSE, ncomponents = 2, run_args = list())

\S4method{plotPCA}{SingleCellExperiment}(object, ..., return_SCE = FALSE,
  draw_plot = TRUE, rerun = FALSE, ncomponents = 2, run_args = list())

plotTSNE(object, ..., return_SCE = FALSE, draw_plot = TRUE, rerun = FALSE,
  ncomponents = 2, run_args = list())

plotDiffusionMap(object, ..., return_SCE = FALSE, draw_plot = TRUE,
  rerun = FALSE, ncomponents = 2, run_args = list())

plotMDS(object, ..., ncomponents = 2, return_SCE = FALSE, rerun = FALSE,
  draw_plot = TRUE, run_args = list())
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{...}{Additional arguments to pass to \code{\link{plotReducedDim}}.}

\item{return_SCE}{Logical, should the function return a SingleCellExperiment object with reduced dimension results in the \code{reducedDim} slot?
Default is \code{FALSE}, in which case a \code{ggplot} object is returned. 
This will be deprecated in the next release in favour of directly calling the underlying \code{run*} functions to compute the results.}

\item{draw_plot}{Logical, should the plot be drawn on the current graphics device? 
Only used if \code{return_SCE} is \code{TRUE}, otherwise the plot is always produced.}

\item{rerun}{Logical, should the reduced dimensions be recomputed even if \code{object} contains an appropriately named set of results in the \code{reducedDims} slot?}

\item{ncomponents}{Numeric scalar indicating the number of dimensions components to (calculate and) plot, starting from the first dimension.
Default is 2, which results in a scatter plot of the first two dimensions.
If \code{ncomponents} is greater than 2, a pairs plots for the top components is produced.}

\item{run_args}{Arguments to pass to \code{\link{runPCA}} when \code{rerun=TRUE} or if there is no existing result in the \code{reducedDims} slot.}
}
\value{
A ggplot plot object or an SingleCellExperiment object, depending on \code{return_SCE}.
}
\description{
Convenience functions to create plots for specific types of reduced dimension results in a SingleCellExperiment object,
or, if they are not already present, to calculate those results and then plot them.
}
\details{
Each function will search the \code{\link{reducedDims}} slot for an appropriately named set of results and pass those coordinates onto \code{\link{plotReducedDim}}.
If the results are not present, they will be computed using the relevant \code{run*} function.
The result name and \code{run*} function for each \code{plot*} function are:
\itemize{
\item \code{"PCA"} and \code{\link{runPCA}} for \code{plotPCA}
\item \code{"TSNE"} and \code{\link{runTSNE}} for \code{plotTSNE}
\item \code{"DiffusionMap"} and \code{\link{runDiffusionMap}} for \code{plotDiffusionMap}
\item \code{"MDS"} and \code{\link{runMDS}} for \code{"plotMDS"}
}
Users can specify arguments to \code{\link{run*}} functions in \code{...}.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

## Examples plotting PC1 and PC2
plotPCA(example_sce)
plotPCA(example_sce, colour_by = "Cell_Cycle")
plotPCA(example_sce, colour_by = "Cell_Cycle", shape_by = "Treatment")
plotPCA(example_sce, colour_by = "Cell_Cycle", shape_by = "Treatment",
    size_by = "Mutation_Status")
plotPCA(example_sce, shape_by = "Treatment", size_by = "Mutation_Status")
plotPCA(example_sce, feature_set = 1:100, colour_by = "Treatment",
    shape_by = "Mutation_Status")

## Experiment with legend
example_subset <- example_sce[, example_sce$Treatment == "treat1"]
plotPCA(example_subset, colour_by = "Cell_Cycle", shape_by = "Treatment", legend = "all")

plotPCA(example_sce, shape_by = "Treatment", return_SCE = TRUE)

## Examples plotting more than 2 PCs
plotPCA(example_sce, ncomponents = 8)
plotPCA(example_sce, ncomponents = 4, colour_by = "Treatment",
shape_by = "Mutation_Status")

## Same for TSNE:
plotTSNE(example_sce, perplexity = 10)
plotTSNE(example_sce, colour_by = "Cell_Cycle", perplexity = 10)
plotTSNE(example_sce, colour_by = "Cell_Cycle", shape_by = "Treatment",
    size_by = "Mutation_Status", perplexity = 10)

## Same for DiffusionMaps:
plotDiffusionMap(example_sce)

## Same for MDS plots:
plotMDS(example_sce)
plotMDS(example_sce, colour_by = "Cell_Cycle")
plotMDS(example_sce, colour_by = "Cell_Cycle", shape_by = "Treatment")

}
\seealso{
\code{\link{runPCA}} \code{\link{runDiffusionMap}} \code{\link{runTSNE}} \code{\link{runMDS}} \code{\link{plotReducedDim}}
}
