% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{runTSNE}
\alias{runTSNE}
\alias{plotTSNE}
\alias{plotTSNE,SingleCellExperiment-method}
\title{Plot t-SNE for an SingleCellExperiment object}
\usage{
runTSNE(object, ntop = 500, ncomponents = 2, exprs_values = "logcounts",
  feature_set = NULL, use_dimred = NULL, n_dimred = NULL,
  scale_features = TRUE, rand_seed = NULL,
  perplexity = floor(ncol(object)/5), ...)

plotTSNE(object, colour_by = NULL, shape_by = NULL, size_by = NULL,
  return_SCE = FALSE, draw_plot = TRUE, theme_size = 10,
  legend = "auto", rerun = FALSE, ncomponents = 2, ...)
}
\arguments{
\item{object}{an \code{SingleCellExperiment} object}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the t-SNE Default is \code{500}, but any \code{ntop} argument is
overrided if the \code{feature_set} argument is non-NULL.}

\item{ncomponents}{numeric scalar indicating the number of t-SNE
components to plot, starting from the first t-SNE component. Default is
2. If \code{ncomponents} is 2, then a scatterplot of component 1 vs component
2 is produced. If \code{ncomponents} is greater than 2, a pairs plots for the
top components is produced. NB: computing more than two components for t-SNE
can become very time consuming.}

\item{exprs_values}{character string indicating which values should be used
as the expression values for this plot. Valid arguments are \code{"tpm"}
(transcripts per million), \code{"norm_tpm"} (normalised TPM
values), \code{"fpkm"} (FPKM values), \code{"norm_fpkm"} (normalised FPKM
values), \code{"counts"} (counts for each feature), \code{"norm_counts"},
\code{"cpm"} (counts-per-million), \code{"norm_cpm"} (normalised
counts-per-million), \code{"logcounts"} (log-transformed count data; default),
\code{"norm_exprs"} (normalised
expression values) or \code{"stand_exprs"} (standardised expression values),
or any other named element of the \code{assayData} slot of the \code{SingleCellExperiment}
object that can be accessed with the \code{assay} function.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the t-SNE calculation. If character, entries must all be
in \code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{use_dimred}{character(1), use named reduced dimension representation of cells
stored in \code{SingleCellExperiment} object instead of recomputing (e.g. "PCA").
 Default is \code{NULL}, no reduced dimension values are provided to \code{Rtsne}.}

\item{n_dimred}{integer(1), number of components of the reduced dimension slot
to use. Default is \code{NULL}, in which case (if \code{use_dimred} is not \code{NULL})
all components of the reduced dimension slot are used.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{rand_seed}{(optional) numeric scalar that can be passed to
\code{set.seed} to make plots reproducible.}

\item{perplexity}{numeric scalar value defining the "perplexity parameter"
for the t-SNE plot. Passed to \code{\link[Rtsne]{Rtsne}} - see documentation
for that package for more details.}

\item{...}{further arguments passed to \code{\link[Rtsne]{Rtsne}}}

\item{colour_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to colour the points in the plot. Alternatively,
a data frame with one column containing values to map to colours for all cells.}

\item{shape_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to define the shape of the points in the plot.
Alternatively, a data frame with one column containing values to map to shapes.}

\item{size_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to define the size of points in the plot.
Alternatively, a data frame with one column containing values to map to sizes.}

\item{return_SCE}{logical, should the function return an \code{SingleCellExperiment}
object with principal component values for cells in the
\code{reducedDims} slot. Default is \code{FALSE}, in which case a
\code{ggplot} object is returned.}

\item{draw_plot}{logical, should the plot be drawn on the current graphics
device? Only used if \code{return_SCE} is \code{TRUE}, otherwise the plot
is always produced.}

\item{theme_size}{numeric scalar giving default font size for plotting theme
(default is 10).}

\item{legend}{character, specifying how the legend(s) be shown? Default is
\code{"auto"}, which hides legends that have only one level and shows others.
Alternatives are "all" (show all legends) or "none" (hide all legends).}

\item{rerun}{logical, should PCA be recomputed even if \code{object} contains a
"PCA" element in the \code{reducedDims} slot?}
}
\value{
If \code{return_SCE} is \code{TRUE}, then the function returns a
\code{SingleCellExperiment} object, otherwise it returns a \code{ggplot} object.
}
\description{
Produce a t-distributed stochastic neighbour embedding (t-SNE) plot of two
components for an \code{SingleCellExperiment} dataset.
}
\details{
The function \code{\link[Rtsne]{Rtsne}} is used internally to
compute the t-SNE. Note that the algorithm is not deterministic, so different
runs of the function will produce differing plots (see \code{\link{set.seed}}
to set a random seed for replicable results). The value of the
\code{perplexity} parameter can have a large effect on the resulting plot, so
it can often be worthwhile to try multiple values to find the most appealing
visualisation.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)
drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]

## Examples plotting t-SNE
plotTSNE(example_sce, perplexity = 10)
plotTSNE(example_sce, colour_by = "Cell_Cycle", perplexity = 10)
plotTSNE(example_sce, colour_by = "Cell_Cycle", shape_by = "Treatment",
size_by = "Mutation_Status", perplexity = 10)
plotTSNE(example_sce, shape_by = "Treatment", size_by = "Mutation_Status",
perplexity = 5)
plotTSNE(example_sce, feature_set = 1:100, colour_by = "Treatment",
shape_by = "Mutation_Status", perplexity = 5)

plotTSNE(example_sce, shape_by = "Treatment", return_SCE = TRUE,
perplexity = 10)


}
\references{
L.J.P. van der Maaten. Barnes-Hut-SNE. In Proceedings of the International
Conference on Learning Representations, 2013.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}
}
