% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCellQC.R
\name{quickCellQC}
\alias{quickCellQC}
\title{Quick cell-level QC}
\usage{
quickCellQC(df, lib_size = "sum", n_features = "detected",
  percent_subsets = NULL, ...)
}
\arguments{
\item{df}{A \linkS4class{DataFrame} containing per-cell QC statistics, as computed by \code{\link{perCellQCMetrics}}.}

\item{lib_size}{String specifying the column of \code{df} containing the library size for each cell.}

\item{n_features}{String specifying the column of \code{df} containing the number of detected features per cell.}

\item{percent_subsets}{String specifying the column(s) of \code{df} containing the percentage of counts in subsets of \dQuote{control features}.}

\item{...}{Further arguments to pass to \code{\link{isOutlier}}.}
}
\value{
A \linkS4class{DataFrame} with one row per cell and containing columns of logical vectors.
Each column specifies a reason for why a cell was considered to be low quality,
with the final \code{discard} column indicating whether the cell should be discarded.
}
\description{
A convenient utility that identifies low-quality cells based on frequently used QC metrics.
}
\details{
This function simply calls \code{\link{isOutlier}} on the various QC metrics in \code{df}.
\itemize{
\item For \code{lib_size}, small outliers are detected on the log-scale to remove cells with low library sizes.
\item For \code{n_features}, small outliers are detected on the log-scale to remove cells with few detected features.
\item For each field in \code{percent_subsets}, large outliers are detected on the original scale.
This aims to remove cells with high spike-in or mitochondrial content.
}

Users can change the number of MADs used to define an outlier or specify batches by passing appropriate arguments to \code{...}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)

df <- perCellQCMetrics(example_sce, subsets=list(Mito=1:100))

discarded <- quickCellQC(df, percent_subsets="subsets_Mito_percent")
colSums(as.data.frame(discarded))

}
\seealso{
\code{\link{perCellQCMetrics}}, for calculation of these metrics.

\code{\link{isOutlier}}, to identify outliers with a MAD-based approach.
}
\author{
Aaron Lun
}
