% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10ximport-wrapper.R
\name{read10XResults}
\alias{read10XResults}
\title{Load in data from 10X experiment}
\usage{
read10XResults(data_dir = NULL)
}
\arguments{
\item{data_dir}{Directory containing the matrix.mtx, genes.tsv, and barcodes.tsv
files provided by 10X. A vector or named vector can be given in order to load 
several data directories. If a named vector is given, the cell barcode names 
will be prefixed with the name.}
}
\value{
Returns a full matrix with rows and columns labeled
}
\description{
Creates a full matrix from a sparse data matrix provided by 10X genomics.
}
\details{
This function is identical to the \code{Read10X} from \code{Seurat} package.
}

