% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kallisto-wrapper.R
\name{readKallistoResults}
\alias{readKallistoResults}
\title{Read kallisto results from a batch of jobs}
\usage{
readKallistoResults(kallisto_log = NULL, samples = NULL,
  directories = NULL, read_h5 = FALSE, kallisto_version = "current",
  logExprsOffset = 1, verbose = TRUE)
}
\arguments{
\item{kallisto_log}{list, generated by \code{runKallisto}. If provided, then
\code{samples} and \code{directories} arguments are ignored.}

\item{samples}{character vector providing a set of sample names to use for
the abundance results.}

\item{directories}{character vector providing a set of directories containing
kallisto abundance results to be read in.}

\item{read_h5}{logical, should the bootstrap results be read in from the HDF5
objects produced by kallisto?}

\item{kallisto_version}{character string indicating whether or not the
version of kallisto to be used is \code{"pre-0.42.2"} or \code{"current"}.
This is required because the kallisto developers changed the output file
extensions and added features in version 0.42.2.}

\item{logExprsOffset}{numeric scalar, providing the offset used when doing
log2-transformations of expression data to avoid trying to take logs of zero.
Default offset value is \code{1}.}

\item{verbose}{logical, should function provide output about progress?}
}
\value{
an SCESet object
}
\description{
After generating transcript/feature abundance results using kallisto for a
batch of samples, read these abundance values into an \code{SCESet} object.
}
\details{
This function expects to find only one set of kallisto abundance
results per directory; multiple adundance results in a given directory will
be problematic.
}
\examples{
\dontrun{
kallisto_log <- runKallisto("targets.txt", "transcripts.idx", single_end=FALSE,
         output_prefix="output", verbose=TRUE, n_bootstrap_samples=10)
sceset <- readKallistoResults(kallisto_log)
}

}

