% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon-wrapper.R
\name{readSalmonResults}
\alias{readSalmonResults}
\title{Read Salmon results from a batch of jobs}
\usage{
readSalmonResults(Salmon_log = NULL, samples = NULL, directories = NULL,
  logExprsOffset = 1, verbose = TRUE)
}
\arguments{
\item{Salmon_log}{list, generated by \code{runSalmon}. If provided, then 
\code{samples} and \code{directories} arguments are ignored.}

\item{samples}{character vector providing a set of sample names to use for 
the abundance results.}

\item{directories}{character vector providing a set of directories containing
Salmon abundance results to be read in.}

\item{logExprsOffset}{numeric scalar, providing the offset used when doing
log2-transformations of expression data to avoid trying to take logs of zero.
Default offset value is \code{1}.}

\item{verbose}{logical, should function provide output about progress?}
}
\value{
an SCESet object
}
\description{
After generating transcript/feature abundance results using Salmon for a 
batch of samples, read these abundance values into an \code{SCESet} object.
}
\details{
This function expects to find only one set of Salmon abundance 
results per directory; multiple adundance results in a given directory will 
be problematic.
}
\examples{
\dontrun{
## Define output directories in a vector called here "Salmon_dirs"
## and sample names as "Salmon_samples"
sceset <- readSalmonResults(samples = Salmon_samples, 
directories = Salmon_dirs)
}

}

