% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{reducedDimension}
\alias{redDim}
\alias{redDim,SCESet-method}
\alias{redDim.SCESet}
\alias{redDim<-}
\alias{redDim<-,SCESet,matrix-method}
\alias{reducedDimension}
\alias{reducedDimension,SCESet-method}
\alias{reducedDimension.SCESet}
\alias{reducedDimension<-}
\alias{reducedDimension<-,SCESet,matrix-method}
\title{Reduced dimension representation for cells in an SCESet object}
\usage{
reducedDimension(object)

reducedDimension(object) <- value

redDim(object)

redDim(object) <- value

reducedDimension.SCESet(object)

\S4method{reducedDimension}{SCESet}(object)

redDim.SCESet(object)

\S4method{redDim}{SCESet}(object)

\S4method{reducedDimension}{SCESet,matrix}(object) <- value

\S4method{redDim}{SCESet,matrix}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a matrix of class \code{"numeric"} containing reduced dimension
coordinates for cells.}
}
\value{
If accessing the \code{reducedDimension} slot, then the matrix of
reduced dimension coordinates. If replacing the \code{reducedDimension} slot
then the new matrix is added to the \code{SCESet} object.
}
\description{
SCESet objects can contain a matrix of reduced dimension coordinates for
cells. These functions conveniently access and replace the reduced dimension
coordinates with the value supplied, which must be a matrix of the correct
size. The function \code{redDim} is simply shorthand for
\code{reducedDimension}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
reducedDimension(example_sceset)

}
\author{
Davis McCarthy
}

