% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runDiffusionMap.R
\docType{methods}
\name{runDiffusionMap}
\alias{runDiffusionMap}
\alias{runDiffusionMap,ANY-method}
\alias{runDiffusionMap,SingleCellExperiment-method}
\title{Create a diffusion map from cell-level data}
\usage{
runDiffusionMap(x, ...)

\S4method{runDiffusionMap}{ANY}(x, ncomponents = 2, ntop = 500,
  subset.row = NULL, feature_set = NULL, scale = FALSE,
  scale_features = NULL, transposed = FALSE, ...)

\S4method{runDiffusionMap}{SingleCellExperiment}(x, ...,
  use.dimred = NULL, use_dimred = NULL, n.dimred = NULL,
  n_dimred = NULL, assay.type = "logcounts", exprs_values = NULL,
  alt.exp = NULL, name = "DiffusionMap")
}
\arguments{
\item{x}{A numeric matrix of log-expression values where rows are features and columns are cells.

Alternatively, a \linkS4class{SingleCellExperiment} object containing such a matrix.

Alternatively, if \code{transposed=TRUE}, a numeric matrix where rows are cells and columns are dimensions.}

\item{...}{For the generic, additional arguments to pass to specific methods.

For the ANY method, additional arguments to pass to \code{\link[destiny]{DiffusionMap}}.

For the SingleCellExperiment method, additional arguments to pass to the ANY method.}

\item{ncomponents}{Numeric scalar indicating the number of UMAP dimensions to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{subset.row}{Vector specifying the subset of features to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{feature_set}{Deprecated, same as \code{subset.row}.}

\item{scale}{Logical scalar, should the expression values be standardised? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{scale_features}{Deprecated, same as \code{scale} but with a different default.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{use.dimred}{String or integer scalar specifying the existing dimensionality reduction results to use, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_dimred}{Deprecated, same as \code{use.dimred}.}

\item{n.dimred}{Integer scalar or vector specifying the dimensions to use if \code{use.dimred} is specified, see \code{?"\link{scater-red-dim-args}"}.}

\item{n_dimred}{Deprecated, same as \code{n.dimred}.}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the expression values, see \code{?"\link{scater-red-dim-args}"}.}

\item{exprs_values}{Deprecated, same as \code{assay.type}.}

\item{alt.exp}{String or integer scalar specifying an alternative experiment to use to compute the PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{name}{String specifying the name to be used to store the result in the \code{reducedDims} of the output.}

\item{feature_set}{Character vector of row names, a logical vector or a numeric vector of indices indicating a set of features to use for UMAP.
This will override any \code{ntop} argument if specified.}
}
\value{
For the ANY method, a matrix is returned containing the diffusion map coordinates for each cell (row) and dimension (column).

For the \linkS4class{SingleCellExperiment} method, a modified version of \code{x} is returned that contains the diffusion map coordinates in the \code{"DiffusionMap"} entry of the \code{\link{reducedDims}}.
}
\description{
Produce a diffusion map for the cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link[destiny]{DiffusionMap}} is used internally to compute the diffusion map.
The behaviour of \code{\link[destiny]{DiffusionMap}} seems to be non-deterministic, in a manner that is not responsive to any \code{\link{set.seed}} call.
The reason for this is unknown.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runDiffusionMap(example_sce, scale_features=NULL)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\references{
Haghverdi L, Buettner F, Theis FJ (2015).
Diffusion maps for high-dimensional single-cell analysis of differentiation data. 
\emph{Bioinformatics} 31(18), 2989-2998.
}
\seealso{
\code{\link[destiny]{DiffusionMap}}, to perform the underlying calculations.

\code{\link[scater]{plotDiffusionMap}}, to quickly visualize the results.

\code{?"\link{scater-red-dim-args}"}, for a full description of various options.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
