% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmon-wrapper.R
\name{runSalmon}
\alias{runSalmon}
\title{Run Salmon on FASTQ files to quantify feature abundance}
\usage{
runSalmon(targets_file, transcript_index, single_end = FALSE,
  output_prefix = "output", lib_type = "A", n_processes = 2,
  n_thread_per_process = 4, n_bootstrap_samples = 0, seqBias = TRUE,
  gcBias = TRUE, posBias = FALSE, allowOrphans = FALSE,
  advanced_opts = NULL, verbose = TRUE, dry_run = FALSE,
  salmon_cmd = "salmon")
}
\arguments{
\item{targets_file}{character string giving the path to a tab-delimited text
file with either 2 columns (single-end reads) or 3 columns (paired-end reads)
that gives the sample names (first column) and FastQ file names (column 2 and
if applicable 3). The file is assumed to have column headers, although these
are not used.}

\item{transcript_index}{character string giving the path to the Salmon
index to be used for the feature abundance quantification.}

\item{single_end}{logical, are single-end reads used, or paired-end reads?}

\item{output_prefix}{character string giving the prefix for the output folder
that will contain the Salmon results. The default is \code{"output"} and
the sample name (column 1 of \code{targets_file}) is appended (preceded by an
underscore).}

\item{lib_type}{scalar, indicating RNA-seq library type. See Salmon 
documentation for details. Default is "A", for automatic detection.}

\item{n_processes}{integer giving the number of processes to use for
parallel Salmon jobs across samples. The package \code{parallel} is used. 
Default is 2 concurrent processes.}

\item{n_thread_per_process}{integer giving the number of threads for Salmon
to use per process (to parallelize Salmon for a given sample). Default is 4.}

\item{n_bootstrap_samples}{integer giving the number of bootstrap samples
that Salmon should use (default is 0). With bootstrap samples, uncertainty
in abundance can be quantified.}

\item{seqBias}{logical, should Salmon's option be used to model and correct
abundances for sequence specific bias? Default is \code{TRUE}.}

\item{gcBias}{logical, should Salmon's option be used to model and correct
abundances for GC content bias? Requires Salmon version 0.7.2 or higher. 
Default is \code{TRUE}.}

\item{posBias}{logical, should Salmon's option be used to model and correct
abundances for positional biases? Requires Salmon version 0.7.3 or higher.
Default is \code{FALSE}.}

\item{allowOrphans}{logical, Consider orphaned reads as valid hits when 
performing lightweight-alignment. This option will increase sensitivity 
(allow more reads to map and more transcripts to be detected), but may 
decrease specificity as orphaned alignments are more likely to be spurious.
For more details see Salmon documentation.}

\item{advanced_opts}{character scalar supplying list of advanced option 
arguments to apply to each Salmon call. For details see Salmon documentation
or type \code{salmon quant --help-reads} at the command line.}

\item{verbose}{logical, should timings for the run be printed?}

\item{dry_run}{logical, if \code{TRUE} then a list containing the Salmon
commands that would be run and the output directories is returned. Can be
used to read in results if Salmon is run outside an R session or to produce
 a script to run outside of an R session.}

\item{salmon_cmd}{(optional) string giving full command to use to call
Salmon, if simply typing "salmon" at the command line does not give the
required version of Salmon or does not work. Default is simply "salmon".
If used, this argument should give the full path to the desired Salmon
binary.}
}
\value{
A list containing three elements for each sample for which feature
abundance has been quantified: (1) \code{salmon_call}, the call used for
Salmon, (2) \code{salmon_log} the log generated by Salmon, and (3)
\code{output_dir} the directory in which the Salmon results can be found.
}
\description{
Run the abundance quantification tool \code{Salmon} on a set of FASTQ
files. Requires \code{Salmon} (\url{https://combine-lab.github.io/salmon/})
to be installed and a Salmon transcript index must have been generated prior
to using this function. See the Salmon website for installation and basic
usage instructions.
}
\details{
A Salmon transcript index can be built from a FASTA file:
\code{salmon index [arguments] FASTA-file}. See the Salmon documentation
for further details. This simple wrapper does not give access to all nuances
of Salmon usage. For finer-grained usage of Salmon please run it at the 
command line - results can still be read into R with 
\code{\link{readSalmonResults}}.
}
\examples{
\dontrun{
## If in Salmon's 'test' directory, then try these calls:
## Generate 'targets.txt' file:
write.table(data.frame(Sample="sample1", File1="reads_1.fastq.gz", File2="reads_1.fastq.gz"),
 file="targets.txt", quote=FALSE, row.names=FALSE, sep="\\t")
Salmon_log <- runSalmon("targets.txt", "transcripts.idx", single_end=FALSE,
         output_prefix="output", verbose=TRUE, n_bootstrap_samples=10,
         dry_run = FALSE)
}
}

