% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_example_cell_info-data.R
\docType{data}
\name{sc_example_cell_info}
\alias{sc_example_cell_info}
\title{Cell information for the small example single-cell counts dataset to 
demonstrate capabilities of scater}
\format{a data.frame instance, 1 row per cell.}
\source{
Wellcome Trust Centre for Human Genetics, Oxford
}
\usage{
sc_example_cell_info
}
\value{
NULL, but makes aavailable a data frame with cell metadata
}
\description{
This data.frame contains cell metadata information for the 40 
cells included in the example \code{counts} dataset included in the package.
}
\author{
Davis McCarthy, 2015-03-05
}
