% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_example_counts-data.R
\docType{data}
\name{sc_example_counts}
\alias{sc_example_counts}
\title{A small example of single-cell counts dataset to demonstrate 
capabilities of scater}
\format{a matrix instance, 1 row per gene.}
\source{
Wellcome Trust Centre for Human Genetics, Oxford
}
\usage{
sc_example_counts
}
\value{
NULL, but makes aavailable a matrix of count data
}
\description{
This data set contains counts for 2000 genes for 40 cells. They 
are from a real experiment, but details have been anonymised.
}
\author{
Davis McCarthy, 2015-03-05
}
