% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater-package.R
\docType{package}
\name{scater-package}
\alias{scater-package}
\title{Single-cell analysis toolkit for expression in R}
\description{
\pkg{scater} provides a class and numerous functions for the quality control,
normalisation and visualisation of single-cell RNA-seq expression data.
}
\details{
In particular, \pkg{scater} provides easy generation of quality control 
metrics and simple functions to visualise quality control metrics and their 
relationships.
}

