% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scater-pkg.R
\name{scater-pkg}
\alias{scater-pkg}
\title{The \pkg{scater} package}
\description{
Provides functions for convenient visualization of single-cell data,
mostly via \pkg{ggplot2}.
It also used to provide utilities for data transformation and quality control,
but these have largely been moved to the \pkg{scuttle} package.
}
\author{
Davis McCarthy, Aaron Lun
}
