% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reddim.R
\name{scater-red-dim-args}
\alias{scater-red-dim-args}
\title{Dimensionality reduction options}
\description{
An overview of the common options for dimensionality reduction methods in \pkg{scater}.
The following sections consider an input \code{x} to the various \code{run*} methods,
where \code{x} can be a numeric matrix or a \linkS4class{SingleCellExperiment}.
}
\section{Feature selection}{

This section is relevant if \code{x} is a numeric matrix of (log-)expression values with features in rows and cells in columns;
or if \code{x} is a \linkS4class{SingleCellExperiment} and \code{dimred=NULL}.
In the latter, the expression values are obtained from the assay specified by \code{exprs_values}.

The \code{subset_row} argument specifies the features to use in a dimensionality reduction algorithm.
This can be set to any user-defined vector containing, e.g., highly variable features or genes in a pathway of interest.
It can be a character vector of row names, an integer vector of row indices or a logical vector.

If \code{subset_row=NULL}, the \code{ntop} features with the largest variances are used instead.
This literally computes the variances from the expression values without considering any mean-variance trend.
Note that the value of \code{ntop} is ignored if \code{subset_row} is specified.

If \code{scale=TRUE}, the expression values for each feature are standardized so that their variance is unity.
This will also remove features with standard deviations below 1e-8.
}

\section{Using reduced dimensions}{

This section is relevant if \code{x} is a \linkS4class{SingleCellExperiment} and \code{dimred} is not \code{NULL}.

All dimensionality reduction methods can be applied on existing dimensionality reduction results in \code{x} by setting \code{dimred}.
This is typically used to run non-linear algorithms like t-SNE or UMAP on the PCA results.
It may also be desirable in cases where the existing reduced dimensions are computed from \emph{a priori} knowledge (e.g., gene set scores).
In such cases, further reduction with PCA could be used to compress the data.

The matrix of existing reduced dimensions is taken from \code{\link{reducedDims}(x, dimred)}.
By default, all dimensions are used to compute the second set of reduced dimensions.
If \code{n_dimred} is also specified, only the first \code{n_dimred} columns are used.
Alternatively, \code{n_dimred} can be an integer vector specifying the column indices of the dimensions to use.

When \code{dimred} is specified, no additional feature selection or standardization is performed.
This means that any settings of \code{ntop}, \code{subset_row} and \code{scale} are ignored.
}

\section{Transposed inputs}{

This section is relevant if \code{x} is a numeric matrix and \code{transposed=TRUE},
such that cells are the rows and the various dimensions are the columns.

Here, the aim is to allow users to manually pass in dimensionality reduction results without needing to wrap them in a \linkS4class{SingleCellExperiment}.
As such, no feature selection or standardization is performed, i.e., \code{ntop}, \code{subset_row} and \code{scale} are ignored.
}

\section{Alternative experiments}{

This section is relevant if \code{x} is a \linkS4class{SingleCellExperiment} and \code{altexp} is not \code{NULL}.

If \code{altexp} is specified, the method is run on data from an alternative \linkS4class{SummarizedExperiment} nested within \code{x}.
This is useful for performing dimensionality reduction on other features stored in \code{\link{altExp}(x, altexp)}, e.g., antibody tags. 

Setting \code{altexp} with \code{exprs_values} will use the specified assay from the alternative SummarizedExperiment.
If the alternative is a SingleCellExperiment, setting \code{dimred} will use the specified dimensionality reduction results from the alternative. 
This option will also interact as expected with \code{n_dimred}.

Note that the output is still stored in the \code{\link{reducedDims}} of the output SingleCellExperiment.
It is advisable to use a different \code{name} to distinguish this from PCA results obtained from the main experiment's assay values.
}

\seealso{
These arguments are used throughout  
\code{\link[scater]{runPCA}}, \code{\link{runTSNE}}, \code{\link{runUMAP}}, \code{\link{runMDS}} and \code{\link{runDiffusionMap}}.
}
\author{
Aaron Lun
}
