% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_parallel.R
\docType{class}
\name{scater-utils}
\alias{scater-utils}
\alias{.splitRowsByWorkers}
\alias{.splitColsByWorkers}
\alias{.splitVectorByWorkers}
\alias{.assignIndicesToWorkers}
\alias{.subset2index}
\alias{.bpNotSharedOrUp}
\title{Developer utilities}
\description{
Various utilities for re-use in packages that happen to depend on \pkg{scater}.
These are exported simply to avoid re-writing them in downstream packages, and should not be touched by end-users.
}
\author{
Aaron Lun
}
