% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{scater-vis-var}
\alias{scater-vis-var}
\title{Variable selection for visualization}
\description{
A number of \pkg{scater} functions accept a SingleCellExperiment object and extract (meta)data from it for use in a plot.
These values are then used on the x- or y-axes (e.g., \code{\link{plotColData}}) or for tuning visual parameters, e.g.,
\code{colour_by}, \code{shape_by}, \code{size_by}.
This page describes how the selection of these values can be controlled by the user,
by passing appropriate values to the arguments of the desired plotting function.
}
\section{When plotting by cells}{

Here, we assume that each visual feature of interest (e.g., point or line) corresponds to a cell in the SingleCellExperiment object \code{sce}.
We will also assume that the user wants to change the colour of each feature according to the cell (meta)data.
To do so, the user can pass as an argument:
\itemize{
\item An unnamed character vector of length 1, i.e., a string. 
This is initially assumed to be the name of a column-level metadata field.
The function will first search the column names of \code{colData(sce)}, and extract metadata for all cells if a matching field is found.
If no match is found, the function will assume that the string represents a gene name.
It will search \code{rownames(sce)} and extract gene expression values for any matching row across all cells.
Otherwise, an error is raised.
\item A named character vector of length 1, where the name is either \code{"exprs"} or \code{"metadata"}.
This forces the function to only search for the string in \code{rownames(sce)} or \code{colnames(colData(sce))}, respectively.
Adding an explicit name is useful when the same field exists in both the row names and column metadata names.
\item A character vector of length greater than 1.
This will search for nested fields in \code{colData(sce)}.
For example, supplying a character vector \code{c("A", "B", "C")} will retrieve \code{colData(sce)$A$B$C}, where both \code{A} and \code{B} contain nested DataFrames.
See \code{\link{calculateQCMetrics}} with \code{compact=TRUE} for an example of how these can be constructed. 
The concatenated name \code{"A:B:C"} will be used in the legend.
\item A character vector of length greater than 1 and the first element set to \code{NA}.
This will search for nested fields in the internal column data of a \linkS4class{SingleCellExperiment}, i.e., in \code{\link{int_colData}}.
For example, \code{c(NA, "size_factor")} would retrieve the values corresponding to \code{sizeFactors(object)}.
The concatenated name without the \code{NA} is used in the legend.
Note that internal fields are \emph{only} searched when \code{NA} is the first element.
\item A data frame with one column and number of rows equal to the number of cells.
This should contain values to use for visualization, e.g., for plotting on the x-/y-axis, or for colouring by.
In this manner, the user can use new information without manually adding it to the SingleCellExperiment object.
The column name of the data frame will be used in the legend.
}

The same logic applies for other visualization parameters such as \code{shape_by} and \code{size_by}.
Other arguments may also use the same scheme, but this depends on the context; see the documentation for each function for details.
In particular, if an argument explicitly refers to a metadata field, any names for the character string will be ignored.
Similarly, a character vector of length > 1 is not allowed for an argument that explicitly refers to expression values.
}

\section{When plotting by features}{

Here, we assume that each visual feature of interest (e.g., point or line) corresponds to a feature in the SingleCellExperiment object \code{sce}.
The scheme is mostly the same as described above, with a few differences:
\itemize{
\item \code{rowData} is searched instead of \code{colData}, as we are extracting metadata for each feature.
\item When extracting expression values, the name of a single cell must be specified.
Visualization will then use the expression profile for all features in that cell.
(This tends to be a rather unusual choice for colouring.)
\item Character strings named with \code{"exprs"} will search for the string in \code{colnames(sce)}.
\item A data frame input should have number of rows equal to the number of features.
}
}

\section{Miscellaneous details}{

Most functions will have a \code{by_exprs_values} parameter. 
This defines the assay of the SingleCellExperiment object from which expression values are extracted for use in colouring, shaping or sizing the points.
The setting of \code{by_exprs_values} will usually default to \code{"logcounts"}, or to the value of \code{exprs_values} in functions such as \code{\link{plotExpression}}.
However, it can be specified separately from \code{exprs_values}, which is useful for visualizing two different types of expression values on the same plot.

Most functions will also have a \code{by_show_single} parameter.
If \code{FALSE}, variables with only one level are not used for visualization, i.e., the visual aspect (colour or shape or size) is set to the default for all points.
No guide is created for this aspect, avoiding clutter in the legend when that aspect provides no information.
If \code{TRUE}, all supplied variables are used for visualization, regardless of how many levels they have.
}

\seealso{
\code{\link{plotColData}}, 
\code{\link{plotRowData}}, 
\code{\link{plotReducedDim}}, 
\code{\link{plotExpression}}, 
\code{\link{plotPlatePosition}},
and most other plotting functions.
}
