% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{scater_gui}
\alias{scater_gui}
\title{scater GUI function}
\usage{
scater_gui(object)
}
\arguments{
\item{object}{SinglCellExperiment object after running \code{\link{calculateQCMetrics}} 
on it}
}
\value{
Opens a browser window with an interactive shiny app and visualize
all possible plots included in the scater
}
\description{
scater shiny app GUI for workflow for less programmatically inclined users or
those who would like a quick and easy way to view multiple plots.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)
drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]
example_sce <- calculateQCMetrics(example_sce, 
feature_controls = list(set1 = 1:40))
\dontrun{
scater_gui(example_sce)
}
}
\author{
Davis McCarthy and Vladimir Kiselev
}
