% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\docType{methods}
\name{SCESet-subset}
\alias{SCESet-subset}
\alias{[,SCESet,ANY,ANY,ANY-method}
\alias{[,SCESet,ANY,ANY-method}
\alias{[,SCESet,ANY-method}
\title{Subsetting SCESet Objects}
\usage{
\S4method{[}{SCESet,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{
    object from which to extract element(s) or in which to replace element(s).
  }

\item{i, j, ...}{indices specifying elements to extract or replace. Indices
are numeric or character vectors or empty (missing) or \code{NULL}. Numeric
values are coerced to integer as by \code{\link[base]{as.integer}} (and hence
truncated towards zero). Character vectors will be matched to the names of
the object (or for matrices/arrays, the dimnames): see
\code{\link[base]{Extract}} for further details.

For \code{[}-indexing only: \code{i, j, ...} can be logical vectors, indicating
elements/slices to select. Such vectors are recycled if necessary to match
the corresponding extent. \code{i, j, ...} can also be negative integers,
indicating elements/slices to leave out of the selection. When indexing
arrays by \code{[} a single argument i can be a matrix with as many columns
as there are dimensions of \code{x}; the result is then a vector with
elements corresponding to the sets of indices in each row of \code{i}. An
index value of \code{NULL} is treated as if it were \code{integer(0)}.}

\item{drop}{For matrices and arrays.  If \code{TRUE} the result is
    coerced to the lowest possible dimension (see the examples).  This
    only works for extracting elements, not for the replacement.  See
    \code{\link{drop}} for further details.
  }
}
\value{
an SCESet object
}
\description{
Subset method for SCESet objects, which subsets both the expression data,
phenotype data, feature data and other slots in the object.
}
\seealso{
\code{\link[base]{Extract}}
}

