% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{spikes}
\alias{spikes}
\alias{spikes,SCESet-method}
\title{Extract expression values for spike-in features in an SCESet object}
\usage{
spikes(object, ...)

\S4method{spikes}{SCESet}(object, exprs_values = "counts", type = NULL)
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{exprs_values}{a string specifying the type of expression values to extract.}

\item{type}{a character vector containing the names of the spike-in control sets to extract. By default,
expression values for features in all spike-in control sets are extracted.}
}
\value{
A matrix of expression values for features in the specified spike-in control sets.
}
\description{
Extract a matrix of expression values for features in spike-in control sets.
}
\details{
If \code{exprs_values="exprs"}, users should have run \code{normalize(object)} first,
so that spike-in features are normalized with spike-in size factors.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset,
                            feature_controls = list(ERCC = 1:40, Mito=41:50))
setSpike(example_sceset) <- "ERCC"
head(spikes(example_scesets))
}
\author{
Aaron Lun
}

