% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{toCellDataSet}
\alias{toCellDataSet}
\title{Convert an \code{SCESet} to a \code{CellDataSet}.}
\usage{
toCellDataSet(sce, exprs_values = "exprs")
}
\arguments{
\item{sce}{An \code{SCESet} object}

\item{exprs_values}{What should \code{exprs(cds)} be mapped from in the
\code{SCESet}? Should be one of "exprs", "tpm", "fpkm", "counts"}
}
\value{
An object of class \code{CellDataSet}
}
\description{
Deprecated from scater version 1.5.2.
}
\examples{
\dontrun{"Deprecated"}
}
