% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{updateSCESet}
\alias{updateSCESet}
\alias{toSingleCellExperiment}
\title{Convert an SCESet object to a SingleCellExperiment object}
\usage{
updateSCESet(object)

toSingleCellExperiment(object)
}
\arguments{
\item{object}{an \code{\link{SCESet}} object to be updated}
}
\value{
a \code{\link[SingleCellExperiment]{SingleCellExperiment}} object
}
\description{
Convert an SCESet object produced with an older version of the
package to a SingleCellExperiment object compatible with the current version.
}
\examples{
\dontrun{
updateSCESet(example_sceset)
}
\dontrun{
toSingleCellExperiment(example_sceset)
}
}
