% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{updateSCESet}
\alias{updateSCESet}
\title{Update an SCESet object to the current version}
\usage{
updateSCESet(object)
}
\arguments{
\item{object}{an \code{\link{SCESet}} object to be updated}
}
\value{
an updated \code{\link{SCESet}} object
}
\description{
It can be necessary to update an SCESet produced with an older version of the
package to be compatible with the current version of the package.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
updateSCESet(example_sceset)
}

