% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{whichSpike}
\alias{whichSpike}
\alias{whichSpike,SCESet-method}
\title{Identify spike-in feature control sets in an SCESet object}
\usage{
whichSpike(object)

\S4method{whichSpike}{SCESet}(object)
}
\arguments{
\item{object}{a \code{SCESet} object.}
}
\value{
A character vector containing the names of feature control sets
that are spike-in sets.
}
\description{
Get the names of the feature control sets that are spike-ins.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset,
                            feature_controls = list(ERCC = 1:40, Mito=41:50))
setSpike(example_sceset) <- "ERCC"
whichSpike(example_sceset)
}
\author{
Aaron Lun
}

