% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{writeSCESet}
\alias{writeSCESet}
\title{Write an SCESet object to an HDF5 file}
\usage{
writeSCESet(object, file_path, type = "HDF5", overwrite_existing = FALSE)
}
\arguments{
\item{object}{\code{\link{SCESet}} object to be writted to file}

\item{file_path}{path to written file containing data from SCESet object}

\item{type}{character string indicating type of output file. Default is "HDF5".}

\item{overwrite_existing}{logical, if a file of the same name already exists
should it be overwritten? Default is \code{FALSE}.}
}
\value{
Return is \code{NULL}, having written the \code{SCESet} object to file.
}
\description{
Write an SCESet object to an HDF5 file
}
\details{
Currently writing to HDF5 files is supported. The \pkg{\link{rhdf5}}
package is used to write data to file and can be used to read data from HDF5
files into R. For further details about the HDF5 data format see 
\url{https://support.hdfgroup.org/HDF5/}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)

\dontrun{
writeSCESet(example_sceset, "test.h5")
file.remove("test.h5")
}

}

