% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPerCellQC.R
\name{addQCPerCell}
\alias{addQCPerCell}
\alias{addQCPerFeature}
\title{Add QC to an SE}
\usage{
addQCPerCell(x, ...)

addQCPerFeature(x, ...)
}
\arguments{
\item{x}{A \linkS4class{SummarizedExperiment} object or one of its subclasses.}

\item{...}{For \code{addQCPerCell}, further arguments to pass to \code{\link{perCellQCMetrics}}.

For \code{addQCPerFeature}, further arguments to pass to \code{\link{perFeatureQCMetrics}}.}
}
\value{
An object like \code{x} but with the QC metrics added to the row or column metadata.
}
\description{
Convenient utilities to compute QC metrics and add them to a \linkS4class{SummarizedExperiment}'s metadata.
}
\details{
These functions are simply wrappers around \code{\link{perCellQCMetrics}} and \code{\link{perFeatureQCMetrics}}, respectively.
The computed QC metrics are automatically appended onto the existing \code{\link{colData}} or \code{\link{rowData}}.
No protection is provided to avoid duplicated column names.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)

example_sce <- addQCPerCell(example_sce)
colData(example_sce)

example_sce <- addQCPerFeature(example_sce)
rowData(example_sce)

}
\seealso{
\code{\link{perCellQCMetrics}} and \code{\link{perFeatureQCMetrics}}, which do the actual work.
}
\author{
Aaron Lun
}
