% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{arrange}
\alias{arrange}
\alias{arrange,SingleCellExperiment-method}
\title{Arrange columns (cells) of a SingleCellExperiment object (deprecated).}
\usage{
arrange(object, ...)

\S4method{arrange}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::arrange} to
act on \code{colData(object)}.}
}
\value{
An SingleCellExperiment object.
}
\description{
The \code{SingleCellExperiment} returned will have cells ordered by the corresponding
variable in \code{colData(object)}.
}
\details{
Refer to \url{https://github.com/sa-lee/plyexperiment} for replacement functionality.
}
