% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{arrange}
\alias{arrange}
\alias{arrange,SingleCellExperiment-method}
\title{Arrange columns (cells) of a SingleCellExperiment object}
\usage{
arrange(object, ...)

\S4method{arrange}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::arrange} to
act on \code{colData(object)}.}
}
\value{
An SingleCellExperiment object.
}
\description{
The \code{SingleCellExperiment} returned will have cells ordered by the corresponding
variable in \code{colData(object)}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), 
colData = sc_example_cell_info)
example_sce <- arrange(example_sce, Cell_Cycle)

}
