% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{bootstraps}
\alias{bootstraps}
\alias{bootstraps,SCESet-method}
\alias{bootstraps.SCESet}
\alias{bootstraps<-}
\alias{bootstraps<-,SCESet,array-method}
\title{Accessor and replacement for bootstrap results in an SCESet object}
\usage{
bootstraps(object)

bootstraps(object) <- value

bootstraps.SCESet(object)

\S4method{bootstraps}{SCESet}(object)

\S4method{bootstraps}{SCESet,array}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{an array of class \code{"numeric"} containing bootstrap
expression values}
}
\value{
If accessing bootstraps slot of an \code{SCESet}, then an array with
the bootstrap values, otherwise an \code{SCESet} object containing new
bootstrap values.
}
\description{
SCESet objects can contain an of bootstrap expression values (for example, as
generated by the kallisto software for quantifying feature abundance). These
 functions conveniently access and replace the 'bootstrap' slot with the value
 supplied, which must be an matrix of the correct size, namely the same
 number of rows and columns as the \code{SCEset} object as a whole.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
bootstraps(example_sceset)

}
\author{
Davis McCarthy
}

