% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calcAverage}
\alias{calcAverage}
\title{Calculate average counts, adjusting for size factors or library size}
\usage{
calcAverage(object, size_factors = NULL)
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object or a matrix of counts}

\item{size_factors}{numeric vector of size factors to use to downscale the counts.
prior to computing averages. Extracted from the 
object if it is a \code{SingleCellExperiment} object; if object is a matrix, then 
if non-NULL, the provided size factors are used. Default is \code{NULL}, where
size factors are defined to be proportional to the library sizes.}
}
\value{
Vector of average count values with same length as number of features.
}
\description{
Calculate average counts per feature, adjusting them as appropriate to take
into account for size factors for normalization or library sizes (total
counts).
}
\details{
For spike-in controls in a \code{SingleCellExperiment}, control-specific 
size factors will be used if available (see \code{\link{normalizeSCE}}). If
no size factors are available for any gene, the library sizes are used instead.
All library sizes are scaled so that the mean is 1 across all cells, to ensure
that the averages are interpretable on the scale of the raw counts.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)

## calculate average counts
ave_counts <- calcAverage(example_sce)

}
