% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calcAverage}
\alias{calcAverage}
\title{Calculate average counts, adjusting for size factors or library size}
\usage{
calcAverage(object)
}
\arguments{
\item{object}{an \code{SCESet} object}
}
\value{
Vector of average count values with same length as number of features.
}
\description{
Calculate average counts per feature, adjusting them as appropriate to take
into account for size factors for normalization or library sizes (total
counts).
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)

## calculate average counts
ave_counts <- calcAverage(example_sceset)

}

