% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAverage.R
\name{calcAverage}
\alias{calcAverage}
\title{Calculate average counts, adjusting for size factors or library size}
\usage{
calcAverage(object, exprs_values = "counts", use_size_factors = TRUE,
  size_factors = NULL, subset_row = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object or count matrix.}

\item{exprs_values}{A string specifying the assay of \code{object} containing the count matrix, if \code{object} is a SingleCellExperiment.}

\item{use_size_factors}{a logical scalar specifying whetherthe size factors in \code{object} should be used to construct effective library sizes.}

\item{size_factors}{A numeric vector containing size factors to use for all non-spike-in features.}

\item{subset_row}{A vector specifying whether the rows of \code{object} should be (effectively) subsetted before calcaulting feature averages.}
}
\value{
Vector of average count values with same length as number of features, or the number of features in \code{subset_row} if supplied.
}
\description{
Calculate average counts per feature, adjusting them as appropriate to take
into account for size factors for normalization or library sizes (total
counts).
}
\details{
The size-adjusted average count is defined by dividing each count by the size factor and taking the average across cells.
All sizes factors are scaled so that the mean is 1 across all cells, to ensure that the averages are interpretable on the scale of the raw counts. 

If \code{use_size_factors=TRUE} and \code{object} is a SingleCellExperiment, size factors are automatically extracted from the object.
For spike-in controls, control-specific size factors will be used if available (see \code{\link{normalizeSCE}}). 
If \code{use_size_factors=FALSE} or \code{object} is a matrix, the library size for each cell is used as the size factor via \code{\link{librarySizeFactors}}.

If \code{size_factors} is supplied, it will override the any size factors for non-spike-in features in \code{object} (if it is a SingleCellExperiment).
The spike-in size factors will still be used. 
If \code{object} is a matrix, \code{size_factors} will be used instead of the library size.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)

## calculate average counts
ave_counts <- calcAverage(example_sce)

}
