% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFPKM.R
\name{calculateFPKM}
\alias{calculateFPKM}
\title{Calculate FPKMs}
\usage{
calculateFPKM(x, effective.length, effective_length = NULL, ...,
  subset_row = NULL)
}
\arguments{
\item{x}{A numeric matrix of counts where features are rows and cells are columns.

Alternatively, a \linkS4class{SummarizedExperiment} or a \linkS4class{SingleCellExperiment} containing such counts.}

\item{effective.length}{Numeric vector providing the effective length for each feature in \code{x}.}

\item{effective_length}{Deprecated, same as \code{effective.length}.}

\item{...}{Further arguments to pass to \code{\link{calculateCPM}}.}
}
\value{
A numeric matrix of FPKM values.
}
\description{
Calculate fragments per kilobase of exon per million reads mapped (FPKM) values from the feature-level counts.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    list(counts = sc_example_counts), 
    colData = sc_example_cell_info)

eff_len <- runif(nrow(example_sce), 500, 2000)
fout <- calculateFPKM(example_sce, eff_len)
str(fout)
}
\seealso{
\code{\link{calculateCPM}}, for the initial calculation of CPM values.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
