% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-expression.R
\name{calculateFPKM}
\alias{calculateFPKM}
\title{Calculate fragments per kilobase of exon per million reads mapped (FPKM)}
\usage{
calculateFPKM(object, effective_length)
}
\arguments{
\item{object}{an \code{SCESet} object}

\item{effective_length}{vector of class \code{"numeric"} providing the 
effective length for each feature in the \code{SCESet} object}
}
\value{
Matrix of FPKM values.
}
\description{
Calculate fragments per kilobase of exon per million reads mapped (FPKM) 
values for expression from counts for a set of features.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
effective_length <- rep(1000, 2000)
fpkm(example_sceset) <- calculateFPKM(example_sceset, effective_length)

}

