% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateFPKM.R
\name{calculateFPKM}
\alias{calculateFPKM}
\title{Calculate fragments per kilobase of exon per million reads mapped (FPKM)}
\usage{
calculateFPKM(object, effective_length, ..., subset_row = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object or a numeric matrix of counts.}

\item{effective_length}{Numeric vector providing the effective length for each feature in \code{object}.}

\item{...}{Further arguments to pass to \code{\link{calculateCPM}}.}

\item{subset_row}{A vector specifying the subset of rows of \code{object} for which to return a result.}
}
\value{
A numeric matrix of FPKM values.
}
\description{
Calculate fragments per kilobase of exon per million reads mapped (FPKM) values for expression from counts for a set of features.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    list(counts = sc_example_counts), 
    colData = sc_example_cell_info)

eff_len <- runif(nrow(example_sce), 500, 2000)
fout <- calculateFPKM(example_sce, eff_len, use_size_factors = FALSE)

}
