% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/SCESet-methods.R
\docType{methods}
\name{cellNames<-}
\alias{cellNames}
\alias{cellNames<-}
\alias{cellNames<-,SCESet,vector-method}
\title{Get or set cell names from an SCESet object}
\usage{
cellNames(object) <- value

cellNames(object)

\S4method{cellNames}{SCESet,vector}(object)<-value
}
\arguments{
\item{object}{An \code{\link{SCESet}} object.}

\item{value}{a vector of cell names to apply to the \code{SCESet} object.}
}
\value{
A vector of cell names.
}
\description{
Get or set cell names from an SCESet object
}
\details{
Simply a wrapper to \code{\link[Biobase]{sampleNames}}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
cellNames(example_sceset)

data("sc_example_counts")
data("sc_example_cell_info")
example_sceset <- newSCESet(countData = sc_example_counts)
cellNames(example_sceset) <- 1:ncol(example_sceset)

}
\author{
Davis McCarthy
}

