% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\docType{methods}
\name{defunct}
\alias{defunct}
\alias{calculateQCMetrics}
\alias{normalize,SingleCellExperiment-method}
\alias{centreSizeFactors}
\title{Defunct functions}
\usage{
calculateQCMetrics(...)

\S4method{normalize}{SingleCellExperiment}(object, ...)

centreSizeFactors(...)
}
\arguments{
\item{object, ...}{Ignored arguments.}
}
\value{
All functions error out with a defunct message pointing towards its descendent (if available).
}
\description{
Functions that have passed on to the function afterlife.
Their successors are also listed.
}
\details{
\code{calculateQCMetrics} is succeeded by \code{\link{perCellQCMetrics}} and \code{\link{perFeatureQCMetrics}}.

\code{normalize} is succeeded by \code{\link{logNormCounts}}.

\code{centreSizeFactors} has no replacement - the \pkg{SingleCellExperiment} is removing support for multiple size factors, so this function is now trivial.
}
\examples{
try(calculateQCMetrics())
}
\author{
Aaron Lun
}
