% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{calculateQCMetrics}
\alias{normalize,SingleCellExperiment-method}
\alias{centreSizeFactors}
\alias{calculateDiffusionMap}
\alias{calculateDiffusionMap,ANY-method}
\alias{runDiffusionMap}
\alias{multiplot}
\title{Defunct functions}
\usage{
calculateQCMetrics(...)

\S4method{normalize}{SingleCellExperiment}(object, ...)

centreSizeFactors(...)

calculateDiffusionMap(x, ...)

\S4method{calculateDiffusionMap}{ANY}(x, ...)

runDiffusionMap(...)

multiplot(...)
}
\arguments{
\item{object, x, ...}{Ignored arguments.}
}
\value{
All functions error out with a defunct message pointing towards its descendent (if available).
}
\description{
Functions that have passed on to the function afterlife.
Their successors are also listed.
}
\details{
\code{calculateQCMetrics} is succeeded by \code{\link{perCellQCMetrics}} and \code{\link{perFeatureQCMetrics}}.

\code{normalize} is succeeded by \code{\link{logNormCounts}}.

\code{centreSizeFactors} has no replacement - the \pkg{SingleCellExperiment} is removing support for multiple size factors, so this function is now trivial.

\code{runDiffusionMap} and \code{calculateDiffusionMap} have no replacement.
\pkg{destiny} is no longer on Bioconductor. You can calculate a diffusion map
yourself, and add it to a \code{reducedDim} field, if you so wish.
}
\examples{
try(calculateQCMetrics())
}
\author{
Aaron Lun
}
