% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10ximport-wrapper.R
\name{downsampleCounts}
\alias{downsampleCounts}
\title{Downsample a count matrix}
\usage{
downsampleCounts(x, prop)
}
\arguments{
\item{x}{matrix of counts}

\item{prop}{numeric scalar or vector of length \code{ncol(x)} in [0, 1] 
indicating the downsampling proportion}
}
\value{
an integer matrix of downsampled counts
}
\description{
Downsample a count matrix to a desired proportion.
}
\details{
This function calls \code{\link[DropletUtils]{downsampleMatrix}}.
from the \pkg{DropletUtils} package. It is deprecated and will be removed
in the next release.
}
\examples{
sce10x <- read10xResults(system.file("extdata", package="scater"))
downsampled <- downsampleCounts(counts(sce10x), prop = 0.5)

}
