% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{filter}
\alias{filter}
\alias{filter,SingleCellExperiment-method}
\title{Return \code{SingleCellExperiment} with cells matching conditions (deprecated).}
\usage{
filter(object, ...)

\S4method{filter}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::filter} to
act on \code{colData(object)}.}
}
\value{
An SingleCellExperiment object.
}
\description{
Subsets the columns (cells) of a \code{SingleCellExperiment} based on
matching conditions in the rows of \code{colData(object)}.
}
\details{
Refer to \url{https://github.com/sa-lee/plyexperiment} for replacement functionality.
}
