% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-preprocessing.R
\name{getBMFeatureAnnos}
\alias{getBMFeatureAnnos}
\title{Get feature annotation information from Biomart}
\usage{
getBMFeatureAnnos(object, filters = "ensembl_transcript_id",
  attributes = c("ensembl_transcript_id", "ensembl_gene_id", feature_symbol,
  "chromosome_name", "transcript_biotype", "transcript_start", "transcript_end",
  "transcript_count"), feature_symbol = "mgi_symbol",
  feature_id = "ensembl_gene_id", biomart = "ENSEMBL_MART_ENSEMBL",
  dataset = "mmusculus_gene_ensembl", host = "www.ensembl.org")
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{filters}{Character vector defining the filters to pass to the \code{\link[biomaRt]{getBM}} function.}

\item{attributes}{Character vector defining the attributes to pass to \code{\link[biomaRt]{getBM}}.}

\item{feature_symbol}{String specifying the attribute to be used to define the symbol to be used for each feature 
Default is \code{"mgi_symbol"}, using gene symbols for mouse - this should be changed if the organism is not \emph{Mus musculus}.}

\item{feature_id}{String specifying the attribute to be used to define the ID to be used for each feature.
Default is \code{"ensembl_gene_id"}, using the Ensembl gene IDs.}

\item{biomart}{String defining the biomaRt to be used, to be passed to \code{\link[biomaRt]{useMart}}.
Default is \code{"ENSEMBL_MART_ENSEMBL"}.}

\item{dataset}{String defining the dataset to use, to be passed to \code{\link[biomaRt]{useMart}}.
Default is \code{"mmusculus_gene_ensembl"}, which should be changed if the organism is not mouse.}

\item{host}{Character string argument which can be used to select a particular \code{"host"} to pass to \code{\link[biomaRt]{useMart}}.
Useful for accessing archived versions of biomaRt data. 
Default is \code{"www.ensembl.org"}, in which case the current version of the biomaRt (now hosted by Ensembl) is used.}
}
\value{
A SingleCellExperiment object containing feature annotation.
The input \code{feature_symbol} appears as the \code{feature_symbol} field in the \code{rowData} of the output object.
}
\description{
Use the \pkg{biomaRt} package to add feature annotation information to an \code{\link{SingleCellExperiment}}.
}
\examples{
\dontrun{
object <- getBMFeatureAnnos(object)
}

}
