% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-preprocessing.R
\name{getBMFeatureAnnos}
\alias{getBMFeatureAnnos}
\title{Get feature annotation information from Biomart}
\usage{
getBMFeatureAnnos(object, filters = "ensembl_transcript_id",
  attributes = c("ensembl_transcript_id", "ensembl_gene_id", "mgi_symbol",
  "chromosome_name", "transcript_biotype", "transcript_start", "transcript_end",
  "transcript_count"), feature_symbol = "mgi_symbol",
  feature_id = "ensembl_gene_id", biomart = "ENSEMBL_MART_ENSEMBL",
  dataset = "mmusculus_gene_ensembl", host = "www.ensembl.org")
}
\arguments{
\item{object}{an \code{SCESet} object}

\item{filters}{character vector defining the "filters" terms to pass to the
biomaRt::getBM function.}

\item{attributes}{character vector defining the biomaRt attributes to pass to
the \code{attributes} argument of \code{\link[biomaRt]{getBM}}.}

\item{feature_symbol}{character string defining the biomaRt attribute to be 
used to define the symbol to be used for each feature (which appears as the 
\code{feature_symbol} in fData(object), subsequently). Default is 
\code{"mgi_symbol"}, gene symbols for mouse. This should be changed if the 
organism is not Mus musculus!}

\item{feature_id}{character string defining the biomaRt attribute to be used 
to define the ID to be used for each feature (which appears as the 
\code{feature_id} in fData(object), subsequently). Default is 
\code{"ensembl_gene_id"}, Ensembl gene IDs for mouse. This should be changed 
if the organism is not Mus musculus!}

\item{biomart}{character string defining the biomaRt to be used. Default is 
\code{"ENSEMBL_MART_ENSEMBL"}.}

\item{dataset}{character string defining the biomaRt dataset to use. Default
is \code{"mmusculus_gene_ensembl"}, which should be changed if the organism 
is not the mouse!}

\item{host}{optional character string argument which can be used to select a
particular \code{"host"} from biomaRt to use. Useful for accessing archived
versions of biomaRt data. Default is \code{"www.ensembl.org"}, in which case the current 
version of the biomaRt (now hosted by Ensembl) is used.}
}
\value{
an SCESet object
}
\description{
Use the \code{biomaRt} package to add feature annotation information to an 
\code{SCESet}.
}
\details{
See the documentation for the biomaRt package, specifically for the
functions \code{useMart} and \code{getBM}, for information on what are 
permitted values for the filters, attributes, biomart, dataset and host 
arguments.
}
\examples{
\dontrun{
object <- getBMFeatureAnnos(object)
}

}

