% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBMFeatureAnnos.R
\name{getBMFeatureAnnos}
\alias{getBMFeatureAnnos}
\title{Get feature annotation information from Biomart}
\usage{
getBMFeatureAnnos(object, ids = rownames(object),
  filters = "ensembl_gene_id", attributes = c(filters, "mgi_symbol",
  "chromosome_name", "gene_biotype", "start_position", "end_position"),
  biomart = "ENSEMBL_MART_ENSEMBL", dataset = "mmusculus_gene_ensembl",
  host = "www.ensembl.org")
}
\arguments{
\item{object}{A \linkS4class{SingleCellExperiment} object.}

\item{ids}{A character vector containing the identifiers for all rows of \code{object}, of the same type specified by \code{filters}.}

\item{filters}{Character vector defining the filters to pass to the \code{\link[biomaRt]{getBM}} function.}

\item{attributes}{Character vector defining the attributes to pass to \code{\link[biomaRt]{getBM}}.}

\item{biomart}{String defining the biomaRt to be used, to be passed to \code{\link[biomaRt]{useMart}}.
Default is \code{"ENSEMBL_MART_ENSEMBL"}.}

\item{dataset}{String defining the dataset to use, to be passed to \code{\link[biomaRt]{useMart}}.
Default is \code{"mmusculus_gene_ensembl"}, which should be changed if the organism is not mouse.}

\item{host}{Character string argument which can be used to select a particular \code{"host"} to pass to \code{\link[biomaRt]{useMart}}.
Useful for accessing archived versions of biomaRt data. 
Default is \code{"www.ensembl.org"}, in which case the current version of the biomaRt (now hosted by Ensembl) is used.}
}
\value{
A SingleCellExperiment object containing feature annotation.
The input \code{feature_symbol} appears as the \code{feature_symbol} field in the \code{rowData} of the output object.
}
\description{
Use the \pkg{biomaRt} package to add feature annotation information to an \code{\link{SingleCellExperiment}}.
}
\examples{
\dontrun{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)

mock_id <- paste0("ENSMUSG", sprintf("\%011d", seq_len(nrow(example_sce))))
example_sce <- getBMFeatureAnnos(example_sce, ids=mock_id)
}

}
