% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{getExprs}
\alias{getExprs}
\title{Retrieve a representation of gene expression}
\usage{
getExprs(object)
}
\arguments{
\item{object}{An object of type \code{SCESet}}
}
\value{
A matrix representation of expression corresponding to \code{object@useForExprs}.
}
\description{
Deprecated from scater version 1.3.29.
}

