% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarianceExplained.R
\name{getVarianceExplained}
\alias{getVarianceExplained}
\title{Estimate the percentage of variance explained for each gene.}
\usage{
getVarianceExplained(object, exprs_values = "logcounts",
  variables = NULL, chunk = 1000)
}
\arguments{
\item{object}{A SingleCellExperiment object containing expression values and per-cell experimental information.}

\item{exprs_values}{String specifying the expression values for which to compute the variance.}

\item{variables}{Character vector specifying the explanatory factors in \code{colData(object)} to use.
Default is \code{NULL}, in which case all variables in \code{colData(object)} are considered.}

\item{chunk}{Integer scalar specifying the chunk size for chunk-wise processing.
Only affects the speed/memory usage trade-off.}
}
\value{
A matrix containing the percentage of variance explained by each factor (column) and for each gene (row).
}
\description{
Estimate the percentage of variance explained for each gene.
}
\details{
This function computes the percentage of variance in gene expression that is explained by variables in the sample-level metadata.
It allows problematic factors to be quickly identified, as well as the genes that are most affected.
}
