% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsce.R
\name{ggcells}
\alias{ggcells}
\alias{ggfeatures}
\title{Create a ggplot from a SingleCellExperiment}
\usage{
ggcells(
  x,
  features = NULL,
  exprs_values = "logcounts",
  use_dimred = TRUE,
  use_altexps = TRUE,
  prefix_altexps = TRUE,
  check_names = TRUE,
  ...
)

ggfeatures(
  x,
  cells = NULL,
  exprs_values = "logcounts",
  check_names = TRUE,
  ...
)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.
This is expected to have row names for \code{ggcells} and column names for \code{ggfeatures}.}

\item{features}{Character vector specifying the features for which to extract expression profiles across cells.
May also include features in alternative Experiments if permitted by \code{use_altexps}.}

\item{exprs_values}{String or integer scalar indicating the assay to use to obtain expression values.
Must refer to a matrix-like object with integer or numeric values.}

\item{use_dimred}{Logical scalar indicating whether data should be extracted for dimensionality reduction results in \code{x}.
Alternatively, a character or integer vector specifying the dimensionality reduction results to use.}

\item{use_altexps}{Logical scalar indicating whether (meta)data should be extracted for alternative experiments in \code{x}.
Alternatively, a character or integer vector specifying the alternative experiments to use.}

\item{prefix_altexps}{Logical scalar indicating whether \code{\link{altExp}}-derived fields should be prefixed with the name of the alternative Experiment.}

\item{check_names}{Logical scalar indicating whether the column names of the output data.frame should be made syntactically valid and unique.}

\item{...}{Further arguments to pass to \link{ggplot}.}
}
\value{
A \link{ggplot} object containing the specified contents of \code{x}.
}
\description{
Create a base \link{ggplot} object from a \linkS4class{SingleCellExperiment},
the contents of which can be directly referenced in subsequent layers without prior specification.
}
\details{
These functions generate a data.frame from the contents of a \linkS4class{SingleCellExperiment} and pass it to \code{\link{ggplot}}.
Rows, columns or metadata fields in the \code{x} can then be referenced in subsequent \pkg{ggplot2} commands.

\code{ggcells} treats cells as the data values,
so users can reference row names of \code{x} (if provided in \code{features}), column metadata variables and dimensionality reduction results.
They can also reference row names and metadata variables for alternative Experiments.

\code{ggfeatures} treats features as the data values,
so users can reference column names of \code{x} (if provided in \code{cells}) and row metadata variables.
}
\examples{
example_sce <- mockSCE()
example_sce <- logNormCounts(example_sce)
example_sce <- runPCA(example_sce)

ggcells(example_sce, features="Gene_0001") + 
    geom_point(aes(x=PCA.1, y=PCA.2, color=Gene_0001))

ggcells(example_sce, features="Gene_0001") + 
    geom_violin(aes(x=Mutation_Status, y=Gene_0001)) +
    facet_wrap(~Cell_Cycle)

rowData(example_sce)$GC <- runif(nrow(example_sce))
ggfeatures(example_sce, cells="Cell_001") +
    geom_point(mapping=aes(x=GC, y=Cell_001)) +
    stat_smooth()

}
\seealso{
\code{\link{makePerCellDF}} and \code{\link{makePerFeatureDF}}, for the construction of the data.frame.
}
\author{
Aaron Lun
}
