% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsce.R
\name{ggsce}
\alias{ggsce}
\title{Create a ggplot from a SingleCellExperiment}
\usage{
ggsce(x, exprs_values = "logcounts", use_altexps = FALSE, ...)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.}

\item{exprs_values}{String or integer scalar indicating the assay to use to obtain expression values.
Must refer to a matrix-like object with integer or numeric values.}

\item{use_altexps}{Logical scalar indicating whether to extract assay/metadata values from \code{\link{altExps}(x)}.}

\item{...}{Further arguments to pass to \link{ggplot}.}
}
\value{
A \link{ggplot} object.
}
\description{
Create a base \link{ggplot} object from a \linkS4class{SingleCellExperiment},
the contents of which can be directly referenced in subsequent layers without prior specification.
}
\details{
This function generates a data.frame from the contents of a \linkS4class{SingleCellExperiment} and passes it to \code{\link{ggplot}}.
Almost any field in the \code{x} can be referenced in subsequent \pkg{ggplot2} commands.
}
\examples{
example_sce <- mockSCE()
example_sce <- logNormCounts(example_sce)
example_sce <- runPCA(example_sce)

ggsce(example_sce) + geom_point(aes(x=PCA.1, y=PCA.2, color=Gene_001))

}
\seealso{
\code{\link{makePerCellDF}}, for the construction of the data.frame.
}
\author{
Aaron Lun
}
