% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{isOutlier}
\alias{isOutlier}
\title{Identify if a cell is an outlier based on a metric}
\usage{
isOutlier(metric, nmads = 5, type = c("both", "lower", "higher"),
  log = FALSE, subset = NULL, batch = NULL, min.diff = NA)
}
\arguments{
\item{metric}{numeric or integer vector of values for a metric}

\item{nmads}{scalar, number of median-absolute-deviations away from median
required for a value to be called an outlier}

\item{type}{character scalar, choice indicate whether outliers should be 
looked for at both tails (default: "both") or only at the lower end ("lower") 
or the higher end ("higher")}

\item{log}{logical, should the values of the metric be transformed to the 
log10 scale before computing median-absolute-deviation for outlier detection?}

\item{subset}{logical or integer vector, which subset of values should be 
used to calculate the median/MAD? If \code{NULL}, all values are used.
Missing values will trigger a warning and will be automatically ignored.}

\item{batch}{factor of length equal to \code{metric}, specifying the batch
to which each observation belongs. A median/MAD is calculated for each batch,
and outliers are then identified within each batch.}

\item{min.diff}{numeric scalar indicating the minimum difference from the 
median to consider as an outlier. The outlier threshold is defined from the 
larger of \code{nmads} MADs and \code{min.diff}, to avoid calling many 
outliers when the MAD is very small. If \code{NA}, it is ignored.}
}
\value{
a logical vector of the same length as the \code{metric} argument
}
\description{
Convenience function to determine which values for a metric are
outliers based on median-absolute-deviation (MAD).
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- calculateQCMetrics(example_sce)

## with a set of feature controls defined
example_sce <- calculateQCMetrics(example_sce, 
feature_controls = list(set1 = 1:40))
isOutlier(example_sce$total_counts, nmads = 3)

}
