% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/librarySizeFactors.R
\name{librarySizeFactors}
\alias{librarySizeFactors}
\alias{librarySizeFactors,ANY-method}
\alias{librarySizeFactors,SummarizedExperiment-method}
\alias{computeLibraryFactors}
\title{Compute library size factors}
\usage{
librarySizeFactors(x, ...)

\S4method{librarySizeFactors}{ANY}(
  x,
  subset_row = NULL,
  geometric = FALSE,
  pseudo_count = 1,
  BPPARAM = SerialParam()
)

\S4method{librarySizeFactors}{SummarizedExperiment}(x, exprs_values = "counts", ...)

computeLibraryFactors(x, ...)
}
\arguments{
\item{x}{For \code{librarySizeFactors}, a numeric matrix of counts with one row per feature and column per cell.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such counts.

For \code{computeLibraryFactors}, only a \linkS4class{SingleCellExperiment} is accepted.}

\item{...}{For the \code{librarySizeFactors} generic, arguments to pass to specific methods.
For the SummarizedExperiment method, further arguments to pass to the ANY method.

For \code{computeLibraryFactors}, further arguments to pass to \code{librarySizeFactors}.}

\item{subset_row}{A vector specifying whether the size factors should be computed from a subset of rows of \code{x}.}

\item{geometric}{Logical scalar indicating whether the size factor should be defined using the geometric mean.}

\item{pseudo_count}{Numeric scalar specifying the pseudo-count to add during log-transformation when \code{geometric=TRUE}.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object indicating how calculations are to be parallelized.
Only relevant when \code{x} is a \linkS4class{DelayedArray} object.}

\item{exprs_values}{String or integer scalar indicating the assay of \code{x} containing the counts.}
}
\value{
For \code{librarySizeFactors}, a numeric vector of size factors is returned for all methods.

For \code{computeLibraryFactors}, a numeric vector is also returned for the ANY and SummarizedExperiment methods.
For the SingleCellExperiment method, \code{x} is returned containing the size factors in \code{\link{sizeFactors}(x)}.
}
\description{
Define per-cell size factors from the library sizes (i.e., total sum of counts per cell).
}
\details{
Library sizes are converted into size factors by scaling them so that their mean across cells is unity.
This ensures that the normalized values are still on the same scale as the raw counts.
Preserving the scale is useful for interpretation of operations on the normalized values,
e.g., the pseudo-count used in \code{\link{logNormCounts}} can actually be considered an additional read/UMI.
This is important for ensuring that the effect of the pseudo-count decreases with increasing sequencing depth.

When using the library size-derived size factor, we implicitly assume that sequencing coverage is the only difference between cells.
This is reasonable for homogeneous cell populations but is compromised by composition biases introduced by DE genes between cell types.
In such cases, normalization by library size factors will not be entirely correct though the effect on downstream conclusions will vary, e.g., clustering is usually unaffected by composition biases but log-fold change estimates will be less accurate.

A closely related alternative approach involves using the geometric mean of counts within each cell to define the size factor,
instead of the library size (which is proportional to the arithmetic mean).
This is enabled with \code{geometric=TRUE} with addition of \code{pseudo_count} to avoid undefined values with zero counts.
The geometric mean is more robust to composition biases from upregulated features but is a poor estimator of the relative bias when there are many zero counts, and thus is best suited for deeply sequenced features, e.g., antibody-derived tags.
}
\examples{
example_sce <- mockSCE()
summary(librarySizeFactors(example_sce))
}
\seealso{
\code{\link{logNormCounts}}, where these size factors are used by default.
}
\author{
Aaron Lun
}
