% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/librarySizeFactors.R
\docType{methods}
\name{librarySizeFactors}
\alias{librarySizeFactors}
\alias{librarySizeFactors,ANY-method}
\alias{librarySizeFactors,SummarizedExperiment-method}
\alias{librarySizeFactors,SingleCellExperiment-method}
\alias{computeLibraryFactors}
\title{Compute library size factors}
\usage{
librarySizeFactors(x, ...)

\S4method{librarySizeFactors}{ANY}(x, subset.row = NULL,
  subset_row = NULL)

\S4method{librarySizeFactors}{SummarizedExperiment}(x,
  assay.type = "counts", exprs_values = NULL, ...)

\S4method{librarySizeFactors}{SingleCellExperiment}(x, ...,
  alt.exp = NULL)

computeLibraryFactors(x, ...)
}
\arguments{
\item{x}{For \code{libarySizeFactors}, a numeric matrix of counts with one row per feature and column per cell.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such counts.

For \code{computeLibraryFactors}, only a \linkS4class{SingleCellExperiment} is accepted.}

\item{...}{For the \code{librarySizeFactors} generic, arguments to pass to specific methods.

For the SummarizedExperiment \code{librarySizeFactors} method, further arguments to pass to the ANY method.

For the SingleCellExperiment \code{librarySizeFactors} method, further arguments to pass to the SummarizedExperiment method.

For \code{computeLibraryFactors}, further arguments to pass to \code{librarySizeFactors}.}

\item{subset.row}{A vector specifying whether the size factors should be computed from a subset of rows of \code{x}.}

\item{subset_row}{Deprecated, same as \code{subset_row}.}

\item{assay.type}{String or integer scalar indicating the assay of \code{x} containing the counts.}

\item{exprs_values}{Deprecated, same as \code{assay.type}.}

\item{alt.exp}{String or integer scalar indicating which (if any) alternative experiment should be used
to provide the counts to compute the size factors.}
}
\value{
For \code{librarySizeFactors}, a numeric vector of size factors is returned for all methods.

For \code{computeLibraryFactors}, a numeric vector is also returned for the ANY and SummarizedExperiment methods.
For the SingleCellExperiment method, \code{x} is returned containing the size factors in \code{\link{sizeFactors}(x)}.
}
\description{
Define per-cell size factors from the library sizes (i.e., total sum of counts per cell).
}
\details{
Library sizes are converted into size factors by scaling them so that their mean across cells is unity.
This ensures that the normalized values are still on the same scale as the raw counts.

Preserving the scale is useful for interpretation of operations on the normalized values,
e.g., the pseudo-count used in \code{\link{logNormCounts}} can actually be considered an additional read/UMI.
This is important for ensuring that the effect of the pseudo-count decreases with increasing sequencing depth.

Setting \code{alt.exp} is occasionally useful for computing size factors from spike-in transcripts
and using them on the count matrix for endogenous genes (stored in the main experiment).
}
\examples{
data("sc_example_counts")
summary(librarySizeFactors(sc_example_counts))

sce <- SingleCellExperiment(list(counts=sc_example_counts))
sce <- computeLibraryFactors(sce)
summary(sizeFactors(sce))
}
\seealso{
\code{\link{logNormCounts}}, where these size factors are used by default.
}
\author{
Aaron Lun
}
