% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logNormCounts.R
\docType{methods}
\name{logNormCounts}
\alias{logNormCounts}
\alias{logNormCounts,DelayedMatrix-method}
\alias{logNormCounts,ANY-method}
\alias{logNormCounts,SummarizedExperiment-method}
\alias{logNormCounts,SingleCellExperiment-method}
\title{Compute log-normalized expression values}
\usage{
logNormCounts(x, ...)

\S4method{logNormCounts}{DelayedMatrix}(x, size.factors = NULL,
  log = TRUE, pseudo.count = 1, center.sf = TRUE)

\S4method{logNormCounts}{ANY}(x, size.factors = NULL, log = TRUE,
  pseudo.count = 1, center.sf = TRUE)

\S4method{logNormCounts}{SummarizedExperiment}(x, log = TRUE, ...,
  assay.type = "counts")

\S4method{logNormCounts}{SingleCellExperiment}(x, size.factors = NULL,
  log = TRUE, pseudo.count = 1, center.sf = TRUE,
  assay.type = "counts", use.alt.exps = TRUE)
}
\arguments{
\item{x}{A numeric matrix-like object of counts for each feature (row) and cell (column).

Alternatively, a \linkS4class{SingleCellExperiment} or \linkS4class{SummarizedExperiment} object containing such a count matrix.}

\item{...}{For the generic, arguments to pass to specific methods.

For the SummarizedExperiment method, further arguments to pass to the ANY method.

For the SingleCellExperiment method, further arguments to pass to the SummarizedExperiment method.}

\item{log}{Logical scalar indicating whether normalized values should be log2-transformed.}

\item{pseudo.count}{Numeric scalar specifying the pseudo-count to add when log-transforming expression values.}

\item{center.sf}{Logical scalar indicating whether size fators should be centred.}

\item{assay.type}{String or integer scalar indicating which assay contains the count data.}

\item{use.alt.exps}{Logical scalar indicating whether QC statistics should be computed for alternative Experiments in \code{x}.
If \code{TRUE}, statistics are computed for all alternative experiments. 

Alternatively, an integer or character vector specifying the alternative Experiments to use to compute QC statistics.

Alternatively, \code{NULL} in which case alternative experiments are not used.}
}
\value{
For the ANY method, a numeric matrix-like object is returned containing (log-)normalized expression values, depending on \code{log}.

For the \linkS4class{DelayedMatrix} method, a DelayedMatrix object is returned with delayed division and log-transformation operations.
This avoids explicitly calculating normalized expression values across a very large (possibly file-backed) matrix.

For the SummarizedExperiment method, \code{x} is returned containing the (log-)normalized expression values in an additional assay.
This is named \code{"logcounts"} if \code{log=TRUE} and \code{"normcounts"} otherwise.

For the SingleCellExperiment method, \code{x} is returned containing an additional assay like the output of the SummarizedExperiment method.
The output will also contain the size factors used in \code{\link{sizeFactors}}, which will be centered if \code{center.sf=TRUE}.
}
\description{
Compute log-transformed normalized expression values from a count matrix based on scaling with size factors.
}
\details{
Normalized expression values are computed by dividing the counts for each cell by the size factor for that cell.
This aims to remove cell-specific scaling biases, e.g., due to differences in sequencing coverage or capture efficiency.
If \code{log=TRUE}, log-normalized values are calculated by adding \code{pseudo.count} to the normalized count and performing a log2 transformation.

If no size factors are supplied, they are determined automatically:
\itemize{
\item For count matrices and \linkS4class{SummarizedExperiment} inputs,
the sum of counts for each cell is used to compute a size factor via the \code{\link{librarySizeFactors}} function.
\item For \linkS4class{SingleCellExperiment} instances, the function attempt to extract \code{\link{sizeFactors}} from \code{x}.
If none are available, it defaults to library size-derived size factors.
}
If \code{size.factors} are supplied, they will override any size factors present in \code{x}.

If \code{center.sf=TRUE}, all sets of size factors will be centered to have the same mean prior to calculation of normalized expression values.
This ensures that abundances are roughly comparable between features normalized with different sets of size factors.
By default, the centre mean is unity, which means that the computed \code{exprs} can be interpreted as being on the same scale as log-counts.
It also means that the added \code{log_exprs_offset} can be interpreted as a pseudo-count (i.e., on the same scale as the counts).

If \code{x} is a SingleCellExperiment and contains alternative experiments, the function will be recursively applied on each nested alternative experiment.
Thus, each nested experiment will contain its own log-normalized expression values.
If \code{size.factors=NULL}, size factors are obtained for each nested experiment following the rules described above.
However, if \code{size.factors} is supplied, it will override any size factors available in the alternative experiments.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)

example_sce <- logNormCounts(example_sce)
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
