% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/logNormCounts.R
\name{logNormCounts}
\alias{logNormCounts}
\alias{logNormCounts,SummarizedExperiment-method}
\alias{logNormCounts,SingleCellExperiment-method}
\title{Compute log-normalized expression values}
\usage{
logNormCounts(x, ...)

\S4method{logNormCounts}{SummarizedExperiment}(
  x,
  size_factors = NULL,
  log = TRUE,
  pseudo_count = 1,
  center_size_factors = TRUE,
  ...,
  exprs_values = "counts",
  name = NULL,
  BPPARAM = SerialParam()
)

\S4method{logNormCounts}{SingleCellExperiment}(
  x,
  size_factors = NULL,
  log = TRUE,
  pseudo_count = 1,
  center_size_factors = TRUE,
  ...,
  exprs_values = "counts",
  use_altexps = FALSE,
  name = NULL,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} or \linkS4class{SummarizedExperiment} object containing a count matrix.}

\item{...}{For the generic, additional arguments passed to specific methods. 

For the methods, additional arguments passed to \code{\link{normalizeCounts}}.}

\item{size_factors}{A numeric vector of cell-specific size factors.
Alternatively \code{NULL}, in which case the size factors are extracted or computed from \code{x}.}

\item{log}{Logical scalar indicating whether normalized values should be log2-transformed.}

\item{pseudo_count}{Numeric scalar specifying the pseudo_count to add when log-transforming expression values.}

\item{center_size_factors}{Logical scalar indicating whether size factors should be centered at unity before being used.}

\item{exprs_values}{A string or integer scalar specifying the assay of \code{x} containing the count matrix.}

\item{name}{String containing an assay name for storing the output normalized values.
Defaults to \code{"logcounts"} when \code{log=TRUE} and \code{"normcounts"} otherwise.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying how library size factor calculations should be parallelized.
Only used if \code{size_factors} is not specified.}

\item{use_altexps}{Logical scalar indicating whether normalization should be performed for alternative experiments in \code{x}.

Alternatively, a character vector specifying the names of the alternative experiments to be normalized.

Alternatively, \code{NULL} in which case alternative experiments are not used.}
}
\value{
\code{x} is returned containing the (log-)normalized expression values in an additional assay named as \code{name}.

If \code{x} is a \linkS4class{SingleCellExperiment}, the size factors used for normalization are stored in \code{\link{sizeFactors}}.
These are centered if \code{center_size_factors=TRUE}.

If \code{x} contains alternative experiments and \code{use_altexps=TRUE},  each of the alternative experiments in \code{x} will also contain an additional assay.
This can be limited to particular \code{\link{altExps}} entries by specifying them in \code{use_altexps}.
}
\description{
Compute log-transformed normalized expression values from a count matrix in a \linkS4class{SingleCellExperiment} object.
}
\details{
This function is a convenience wrapper around \code{\link{normalizeCounts}}.
It returns a \linkS4class{SingleCellExperiment} or \linkS4class{SummarizedExperiment} containing the normalized values in a separate assay.
This makes it easier to perform normalization by avoiding book-keeping errors during a long analysis workflow.

If \code{x} is a \linkS4class{SingleCellExperiment} that contains alternative Experiments, normalized values can be computed and stored within each alternative experiment by setting \code{use_altexps} appropriately.
By default, \code{use_altexps=FALSE} to avoid problems from attempting to library size-normalize alternative experiments that have zero total counts for some cells.

If \code{size_factors=NULL}, size factors are obtained following the rules in \code{\link{normalizeCounts}}.
This is done independently for the main and alternative Experiments when \code{use_altexps} is specified,
i.e. no information is shared between Experiments by default.
However, if \code{size_factors} is supplied, it will override any size factors available in any Experiment.
}
\examples{
example_sce <- mockSCE()
example_sce <- logNormCounts(example_sce)
assayNames(example_sce)
}
\seealso{
\code{\link{normalizeCounts}}, which is used to compute the normalized expression values.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
