% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makePerFeatureDF.R
\name{makePerFeatureDF}
\alias{makePerFeatureDF}
\title{Create a per-feature data.frame from a SingleCellDataFrame}
\usage{
makePerFeatureDF(
  x,
  cells = NULL,
  exprs_values = "logcounts",
  check_names = FALSE
)
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.
This is expected to have non-\code{NULL} row names.}

\item{cells}{Character vector specifying the features for which to extract expression profiles across cells.}

\item{exprs_values}{String or integer scalar indicating the assay to use to obtain expression values.
Must refer to a matrix-like object with integer or numeric values.}

\item{check_names}{Logical scalar indicating whether the column names of the output data.frame should be made syntactically valid and unique.}
}
\value{
A data.frame containing one field per aspect of data in \code{x} - see Details.
Each row corresponds to a feature (i.e., row) of \code{x}.
}
\description{
Create a per-feature data.frame (i.e., where each row represents a feature) from a \linkS4class{SingleCellExperiment},
most typically for creating custom \pkg{ggplot2} plots.
}
\details{
This function enables us to conveniently create a per-feature data.frame from a \linkS4class{SingleCellExperiment}.
Each row of the returned data.frame corresponds to a row in \code{x},
while each column of the data.frame corresponds to one aspect of the (meta)data in \code{x}.
Columns are provided in the following order:
\enumerate{
\item Columns named according to values in \code{cells} represent the expression values across features for the specified cell in the \code{exprs_values} assay.
\item Columns named according to the columns of \code{rowData(x)} represent the row metadata variables.
}

By default, nothing is done to resolve syntactically invalid or duplicated column names;
this will often lead (correctly) to an error in downstream functions like \code{\link{ggplot}}.
If \code{check_names=TRUE}, this is resolved by passing the column names through \code{\link{make.names}}.
Of course, as a result, some columns may not have the same names as the original fields in \code{x}.
}
\examples{
example_sce <- mockSCE()
example_sce <- logNormCounts(example_sce)
rowData(example_sce)$Length <- runif(nrow(example_sce))

df <- makePerFeatureDF(example_sce, cells="Cell_001")
head(colnames(df))
tail(colnames(df))

head(df$Cell_001)
head(df$Length)

}
\seealso{
\code{\link{ggfeatures}}, which uses this function under the hood.
}
\author{
Aaron Lun
}
