% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{mergeSCESet}
\alias{mergeSCESet}
\title{Merge SCESet objects}
\usage{
mergeSCESet(x, y, fdata_cols_x = 1:ncol(fData(x)),
  fdata_cols_y = fdata_cols_x, pdata_cols_x = NULL, pdata_cols_y = NULL)
}
\arguments{
\item{x}{an \code{\link{SCESet}} object}

\item{y}{an \code{\link{SCESet}} object}

\item{fdata_cols_x}{a logical or numeric vector indicating which columns of featureData
for \code{x} are shared between \code{x} and \code{y} and should feature in the
returned merged \code{SCESet}. Default is all columns of \code{fData(x)}.}

\item{fdata_cols_y}{a logical or numeric vector indicating which columns of featureData
for \code{y} are shared between \code{x} and \code{y} and should feature in the
returned merged \code{SCESet}. Default is \code{fdata_cols_x}.}

\item{pdata_cols_x}{a logical or numeric vector indicating which columns of phenoData
of \code{x} should be retained.}

\item{pdata_cols_y}{a logical or numeric vector indicating which columns of phenoData
of \code{y} should be retained.}
}
\value{
a merged \code{SCESet} object combining data and metadata from \code{x} and \code{y}
}
\description{
Merge two SCESet objects that have the same features but contain different cells/samples.
}
\details{
Existing cell-cell pairwise distances and feature-feature pairwise distances will not be valid for a merged SCESet so these are set to \code{NULL} in the returned object. Similarly \code{experimentData} will need to be added anew to the merged SCESet returned.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
mergeSCESet(example_sceset[, 1:20], example_sceset[, 21:40])

## with specification of columns of fData
example_sceset <- calculateQCMetrics(example_sceset)
mergeSCESet(example_sceset[, 1:20], example_sceset[, 21:40], fdata_cols_x = c(1, 7))

## with specification of columns of pData
mergeSCESet(example_sceset[, 1:20], example_sceset[, 21:40], pdata_cols_x = 1:6)


}

