% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{mutate}
\alias{mutate}
\alias{mutate,SCESet-method}
\alias{mutate.SCESet}
\title{Add new variables to \code{pData(object)}.}
\usage{
mutate(object, ...)

\S4method{mutate}{SCESet}(object, ...)

mutate.SCESet(object, ...)
}
\arguments{
\item{object}{A \code{SCESet} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::mutate} to
act on \code{pData(object)}.}
}
\description{
Adds new columns to \code{pData(object)} preserving existing variables.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- mutate(example_sceset, is_quiescent = Cell_Cycle == "G0")
}

