% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{mutate}
\alias{mutate}
\alias{mutate}
\alias{mutate,SingleCellExperiment-method}
\title{Add new variables to \code{colData(object)}.}
\usage{
mutate(object, ...)

\S4method{mutate}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::mutate} to
act on \code{colData(object)}.}
}
\value{
An SingleCellExperiment object.
}
\description{
Adds ne
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), 
colData = sc_example_cell_info)
example_sce <- mutate(example_sce, is_quiescent = Cell_Cycle == "G0")
}
