% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\name{newSCESet}
\alias{newSCESet}
\title{Create a new SCESet object.}
\usage{
newSCESet(exprsData = NULL, countData = NULL, tpmData = NULL,
  fpkmData = NULL, cpmData = NULL, phenoData = NULL, featureData = NULL,
  experimentData = NULL, is_exprsData = NULL,
  cellPairwiseDistances = dist(vector()),
  featurePairwiseDistances = dist(vector()), lowerDetectionLimit = 0,
  logExprsOffset = 1, logged = FALSE, useForExprs = "exprs")
}
\arguments{
\item{exprsData}{expression data matrix for an experiment}

\item{countData}{data matrix containing raw count expression values}

\item{tpmData}{matrix of class \code{"numeric"} containing
transcripts-per-million (TPM) expression values}

\item{fpkmData}{matrix of class \code{"numeric"} containing fragments per
kilobase of exon per million reads mapped (FPKM) expression values}

\item{cpmData}{matrix of class \code{"numeric"} containing counts per
million (CPM) expression values (optional)}

\item{phenoData}{data frame containing attributes of individual cells}

\item{featureData}{data frame containing attributes of features (e.g. genes)}

\item{experimentData}{MIAME class object containing metadata data and details
about the experiment and dataset.}

\item{is_exprsData}{matrix of class \code{"logical"}, indicating whether
or not each observation is above the \code{lowerDetectionLimit}.}

\item{cellPairwiseDistances}{object of class \code{"dist"} (or a class that 
extends "dist") containing cell-cell distance or dissimilarity values.}

\item{featurePairwiseDistances}{object of class \code{"dist"} (or a class that 
extends "dist") containing feature-feature distance or dissimilarity values.}

\item{lowerDetectionLimit}{the minimum expression level that constitutes true
expression (defaults to zero and uses count data to determine if an
observation is expressed or not).}

\item{logExprsOffset}{numeric scalar, providing the offset used when doing
log2-transformations of expression data to avoid trying to take logs of zero.
Default offset value is \code{1}.}

\item{logged}{logical, if a value is supplied for the exprsData argument, are
the expression values already on the log2 scale, or not?}

\item{useForExprs}{character string, either 'exprs' (default),'tpm','counts' or
'fpkm' indicating which expression representation both internal methods and
external packages should use when performing analyses.}
}
\value{
a new SCESet object
}
\description{
Scater requires that all data be housed in SCESet objects. SCESet extends
Bioconductor's ExpressionSet class, and the same basic interface is
supported. newSCESet() expects a matrix of expression values as its first
argument, with rows as features (usually genes) and columns as cells.
Per-feature and per-cell metadata can be supplied with the featureData and
phenoData arguments, respectively. Use of these optional arguments is
strongly encouraged. The SCESet also includes a slot 'counts' to store an
object containing raw count data.
}
\details{
SCESet objects store a matrix of expression values. These values are
typically transcripts-per-million (tpm), counts-per-million (cpm), fragments
per kilobase per million mapped (FPKM) or some other output from a program
that calculates expression values from RNA-Seq reads. We recommend that
expression values on the log2 scale are used for the 'exprs' slot in the
SCESet. For example, you may wish to store raw tpm values in the 'tpm' slot
and \code{log2(tpm + 1)} values in the 'exprs' slot. However, expression
values could also be values from a single cell qPCR run or some other type of
 assay. The newSCESet function can also accept raw count values. In this case
 see \code{\link{calculateTPM}} and \code{\link{calculateFPKM}} for computing
 TPM and FPKM expression values, respectively, from counts. The function
 \code{\link[edgeR]{cpm}} from the package edgeR to can be used to compute
 log2(counts-per-million), if desired.

 An \code{SCESet} object has to have the \code{'exprs'} slot defined, so if
 the \code{exprsData} argument is \code{NULL}, then this function will define
 \code{'exprs'} with the following order of precedence: log2(TPM +
 logExprsOffset), if \code{tpmData} is defined; log2(FPKM + logExprsOffset)
 if \code{fpkmData} is defined; otherwise log2(counts-per-million +
 logExprsOffset) are used. The \code{\link[edgeR]{cpm}} function from the
 edgeR package is used to compte \code{cpm}. Note that for many analyses
 counts-per-million are not recommended, and if possible
 transcripts-per-million should be used.

 In many downstream functions you will likely find it most convenient if the
 \code{'exprs'} values are on the log2-scale, so this is recommended.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset
}

