% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{nexprs}
\alias{nexprs}
\title{Count the number of expressed genes per cell}
\usage{
nexprs(object, threshold = NULL, subset.row = NULL, byrow = FALSE)
}
\arguments{
\item{object}{an \code{SCESet} object}

\item{threshold}{numeric scalar providing the value above which observations
are deemed to be expressed. Defaults to \code{object@lowerDetectionLimit}.}

\item{subset.row}{logical or character vector indicating which rows 
(i.e. features/genes) to subset and calculate 'is_exprs_mat' for.}

\item{byrow}{logical scalar indicating if \code{TRUE} to count expressing 
cells per feature (i.e. gene) and if \code{FALSE} to count expressing 
features (i.e. genes) per cell.}
}
\value{
a numeric vector of the same length as the number of features if
\code{byrow} argument is \code{TRUE} and the same length as the number of 
cells if \code{byrow} is \code{FALSE}
}
\description{
An efficient internal function that avoids the need to construct 
'is_exprs_mat' by counting the number of expressed genes per cell on the fly.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data=sc_example_cell_info)
rownames(pd) <- pd$Cell
example_sceset <- newSCESet(countData=sc_example_counts, phenoData=pd)
nexprs(example_sceset)[1:10]
nexprs(example_sceset, byrow = TRUE)[1:10]

}

