% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nexprs.R
\name{nexprs}
\alias{nexprs}
\title{Count the number of non-zero counts per cell or feature}
\usage{
nexprs(object, detection_limit = 0, exprs_values = "counts",
  byrow = FALSE, subset_row = NULL, subset_col = NULL)
}
\arguments{
\item{object}{A SingleCellExperiment object or a numeric matrix of expression values.}

\item{detection_limit}{Numeric scalar providing the value above which  observations are deemed to be expressed.}

\item{exprs_values}{String or integer specifying the assay of \code{object} to obtain the count matrix from, if \code{object} is a SingleCellExperiment.}

\item{byrow}{Logical scalar indicating whether to count the number of detected cells per feature.
If \code{FALSE}, the function will count the number of detected features per cell.}

\item{subset_row}{Logical, integer or character vector indicating which rows (i.e. features) to use.}

\item{subset_col}{Logical, integer or character vector indicating which columns (i.e., cells) to use.}
}
\value{
An integer vector containing counts per gene or cell, depending on the provided arguments.
}
\description{
An efficient internal function that counts the number of non-zero counts in each row (per feature) or column (per cell).
This avoids the need to construct an intermediate logical matrix.
}
\details{
Setting \code{subset_row} or \code{subset_col} is equivalent to subsetting \code{object} before calling \code{nexprs}, 
but more efficient as a new copy of the matrix is not constructed.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info)

nexprs(example_sce)[1:10]
nexprs(example_sce, byrow = TRUE)[1:10]

}
