% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeCounts.R
\name{normalizeCounts}
\alias{normalizeCounts}
\title{Divide columns of a count matrix by the size factors}
\usage{
normalizeCounts(x, size_factors, return_log = TRUE,
  log_exprs_offset = 1, centre_size_factors = FALSE,
  subset_row = NULL)
}
\arguments{
\item{x}{A count matrix, with cells in the columns and genes in the rows.}

\item{size_factors}{A numeric vector of size factors for all cells.}

\item{return_log}{Logical scalar, should normalized values be returned on the log2 scale?}

\item{log_exprs_offset}{Numeric scalar specifying the offset to add when log-transforming expression values.}

\item{centre_size_factors}{Logical scalar indicating whether size fators should be centred.}

\item{subset_row}{A vector specifying the subset of rows of \code{x} for which to return a result.}
}
\value{
A matrix of (log-)normalized expression values.
}
\description{
Compute (log-)normalized expression values by dividing counts for each cell by the corresponding size factor.
}
\details{
This function is more memory-efficient than \code{t(t(x)/size_factors)},
and more generally applicable to different matrix classes than \code{sweep(x, 2, size_factors, "*")}.

Note that the default \code{centre_size_factors} differs from that in \code{\link{normalizeSCE}}.
Users of this function are assumed to know what they're doing with respect to normalization.
}
\examples{
data("sc_example_counts")
normed <- normalizeCounts(sc_example_counts, 
    librarySizeFactors(sc_example_counts))
}
\author{
Aaron Lun
}
