% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/numDetectedAcrossFeatures.R
\name{numDetectedAcrossFeatures}
\alias{numDetectedAcrossFeatures}
\alias{numDetectedAcrossFeatures,ANY-method}
\alias{numDetectedAcrossFeatures,SummarizedExperiment-method}
\title{Number of detected expression values per group of features}
\usage{
numDetectedAcrossFeatures(x, ...)

\S4method{numDetectedAcrossFeatures}{ANY}(
  x,
  ids,
  average = FALSE,
  subset_row = NULL,
  subset_col = NULL,
  ...,
  BPPARAM = SerialParam()
)

\S4method{numDetectedAcrossFeatures}{SummarizedExperiment}(x, ..., exprs_values = "counts")
}
\arguments{
\item{x}{A numeric matrix of counts where features are rows and cells are columns.

Alternatively, a \linkS4class{SummarizedExperiment} containing such counts.}

\item{...}{For the generic, further arguments to pass to specific methods.

For the SummarizedExperiment method, further arguments to pass to the ANY method.

For the ANY method, further arguments to pass to the \code{\link{nexprs}} function.}

\item{ids}{A vector of length \code{nrow(x)}, specifying the group assignment for each feature.}

\item{average}{Logical scalar indicating whether the proportion of non-zero counts in each group should be computed instead.}

\item{subset_row}{A vector specifying the rows to use.
Defaults to all rows with non-\code{NA} entries of \code{ids}.}

\item{subset_col}{A vector specifying the columns to use.
Defaults to all columns.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether the calculations should be parallelized.}

\item{exprs_values}{String or integer specifying the assay of \code{x} to obtain the count matrix from.}
}
\value{
An integer or numeric matrix containing the number of detected expression values in each group of features (row) and cell (column).
}
\description{
Computes the number of detected expression values (default defined as non-zero counts) for each cell in each group of features.
}
\examples{
example_sce <- mockSCE()

ids <- sample(paste0("GENE_", 1:100), nrow(example_sce), replace=TRUE)
byrow <- numDetectedAcrossFeatures(example_sce, ids)
head(byrow[,1:10])

}
\seealso{
\code{\link{nexprs}}, on which this function is based.
}
\author{
Aaron Lun
}
