% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plotColData}
\alias{plotColData}
\alias{plotPhenoData}
\alias{plotCellData}
\title{Plot cell phenotype data from an SingleCellExperiment object}
\usage{
plotColData(object, y, x = NULL, colour_by = NULL, shape_by = NULL,
  size_by = NULL, exprs_values = "logcounts", ...)

plotPhenoData(...)

plotCellData(...)
}
\arguments{
\item{object}{A SingleCellExperiment object containing expression values and experimental information.}

\item{y}{Specification of the column-level metadata to show on the y-axis, see \code{?"\link{scater-vis-var}"} for possible values.
Note that only metadata fields will be searched, \code{assays} will not be used.}

\item{x}{Specification of the column-level metadata to show on the x-axis, see \code{?"\link{scater-vis-var}"} for possible values.
Again, only metadata fields will be searched, \code{assays} will not be used.}

\item{colour_by}{Specification of a column metadata field or a feature to colour by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{shape_by}{Specification of a column metadata field or a feature to shape by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{size_by}{Specification of a column metadata field or a feature to size by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{...}{Additional arguments for visualization, see \code{?"\link{scater-plot-args}"} for details.}
}
\value{
A ggplot object.
}
\description{
Plot cell phenotype data from an SingleCellExperiment object
}
\details{
If \code{y} is continuous and \code{x=NULL}, a violin plot is generated.
If \code{x} is categorical, a grouped violin plot will be generated, with one violin for each level of \code{x}.
If \code{x} is continuous, a scatter plot will be generated.

If \code{y} is categorical and \code{x} is continuous, horizontal violin plots will be generated.
If \code{x} is missing or categorical, rectangule plots will be generated where the area of a rectangle is proportional to the number of points for a combination of factors.

Note that \code{plotPhenoData} and \code{plotCellData} are synonyms for \code{plotColData}.
These are artifacts of the transition from the old SCESet class, and will be deprecated in future releases.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- calculateQCMetrics(example_sce)
example_sce <- normalize(example_sce)

plotColData(example_sce, y = "total_features_by_counts", 
   x = "log10_total_counts", colour_by = "Mutation_Status")

plotColData(example_sce, y = "total_features_by_counts", 
   x = "log10_total_counts", colour_by = "Mutation_Status",
   size_by = "Gene_0001", shape_by = "Treatment")

plotColData(example_sce, y = "Treatment", 
   x = "log10_total_counts", colour_by = "Mutation_Status")

plotColData(example_sce, y = "total_features_by_counts", 
   x = "Cell_Cycle", colour_by = "Mutation_Status")

plotColData(example_sce, y = "Mutation_Status", 
   x = "Cell_Cycle", colour_by = "Mutation_Status")

plotColData(example_sce, y = "Mutation_Status", 
   x = "Cell_Cycle", colour_by = "Mutation_Status",
   size_by = "Gene_0001", shape_by = "Treatment")
}
