% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotDiffusionMap}
\alias{plotDiffusionMap}
\alias{plotDiffusionMap,SingleCellExperiment-method}
\title{Plot a diffusion map for a SingleCellExperiment object}
\usage{
plotDiffusionMap(object, ..., return_SCE = FALSE, draw_plot = TRUE,
  rerun = FALSE, ncomponents = 2, run_args = list())
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object}

\item{...}{Additional arguments to pass to \code{\link{plotReducedDim}}.}

\item{return_SCE}{logical, should the function return a \code{SingleCellExperiment}
object with principal component values for cells in the
\code{reducedDims} slot. Default is \code{FALSE}, in which case a
\code{ggplot} object is returned. This will be deprecated in the next 
development cycle in favour of directly calling \code{\link{runDiffusionMap}}.}

\item{draw_plot}{logical, should the plot be drawn on the current graphics
device? Only used if \code{return_SCE} is \code{TRUE}, otherwise the plot
is always produced.}

\item{rerun}{logical, should PCA be recomputed even if \code{object} contains a
\code{"DiffusionMap"} element in the \code{reducedDims} slot?}

\item{ncomponents}{numeric scalar indicating the number of diffusion 
components to plot, starting from the first diffusion component. Default
is 2. If \code{ncomponents} is 2, then a scatterplot of component 1 vs
component 2 is produced. If \code{ncomponents} is greater than 2, a pairs
plots for the top components is produced. NB: computing many components for
the diffusion map can become time consuming.}

\item{run_args}{Arguments to pass to \code{\link{runDiffusionMap}} when \code{rerun=TRUE}
or if there is no existing \code{"DiffusionMap"} element in the \code{reducedDims} slot.}
}
\value{
If \code{return_SCE} is \code{TRUE}, then the function returns an
\code{SingleCellExperiment} object, otherwise it returns a \code{ggplot} object.
}
\description{
Produce a diffusion map plot of two components for a \code{SingleCellExperiment} object.
}
\details{
For back-compatabibility purposes, users can specify arguments to 
\code{\link{runDiffusionMap}} in \code{...}. This will trigger a warning as it will be
deprecated in the next development cycle.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)
drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]

\dontrun{
## Examples plotting diffusion maps
plotDiffusionMap(example_sce)
plotDiffusionMap(example_sce, colour_by = "Cell_Cycle")
plotDiffusionMap(example_sce, colour_by = "Cell_Cycle",
shape_by = "Treatment")
plotDiffusionMap(example_sce, colour_by = "Cell_Cycle",
shape_by = "Treatment", size_by = "Mutation_Status")
plotDiffusionMap(example_sce, shape_by = "Treatment",
size_by = "Mutation_Status")
plotDiffusionMap(example_sce, feature_set = 1:100, colour_by = "Treatment",
shape_by = "Mutation_Status")

plotDiffusionMap(example_sce, shape_by = "Treatment",
return_SCE = TRUE)
}

}
\references{
Haghverdi L, Buettner F, Theis FJ. Diffusion maps for high-dimensional single-cell analysis of differentiation data. Bioinformatics. 2015; doi:10.1093/bioinformatics/btv325
}
\seealso{
\code{\link{runDiffusionMap}}
}
