% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{runDiffusionMap}
\alias{runDiffusionMap}
\alias{plotDiffusionMap}
\alias{plotDiffusionMap,SingleCellExperiment-method}
\title{Plot a diffusion map for a SingleCellExperiment object}
\usage{
runDiffusionMap(object, ntop = 500, ncomponents = 2, feature_set = NULL,
  exprs_values = "logcounts", scale_features = TRUE, use_dimred = NULL,
  n_dimred = NULL, rand_seed = NULL, sigma = NULL,
  distance = "euclidean", ...)

plotDiffusionMap(object, colour_by = NULL, shape_by = NULL,
  size_by = NULL, return_SCE = FALSE, draw_plot = TRUE, theme_size = 10,
  legend = "auto", rerun = FALSE, ncomponents = 2, ...)
}
\arguments{
\item{object}{an \code{SingleCellExperiment} object}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the diffusion map. Default is \code{500}, but any \code{ntop}
argument is overrided if the \code{feature_set} argument is non-NULL.}

\item{ncomponents}{numeric scalar indicating the number of principal
components to plot, starting from the first diffusion map component. Default
is 2. If \code{ncomponents} is 2, then a scatterplot of component 1 vs
component 2 is produced. If \code{ncomponents} is greater than 2, a pairs
plots for the top components is produced. NB: computing many components for
the diffusion map can become time consuming.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the diffusion map. If character, entries must all be in
\code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{exprs_values}{character string indicating which values should be used
as the expression values for this plot. Valid arguments are \code{"tpm"}
(transcripts per million), \code{"norm_tpm"} (normalised TPM
values), \code{"fpkm"} (FPKM values), \code{"norm_fpkm"} (normalised FPKM
values), \code{"counts"} (counts for each feature), \code{"norm_counts"},
\code{"cpm"} (counts-per-million), \code{"norm_cpm"} (normalised
counts-per-million), \code{"logcounts"} (log-transformed count data; default),
\code{"norm_exprs"} (normalised
expression values) or \code{"stand_exprs"} (standardised expression values)
or any other named element of the \code{assayData} slot of the \code{SingleCellExperiment}
object that can be accessed with the \code{assay} function.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{use_dimred}{character(1), use named reduced dimension representation of cells
stored in \code{SingleCellExperiment} object instead of recomputing (e.g. "PCA").
 Default is \code{NULL}, no reduced dimension values are provided to \code{Rtsne}.}

\item{n_dimred}{integer(1), number of components of the reduced dimension slot
to use. Default is \code{NULL}, in which case (if \code{use_dimred} is not \code{NULL})
all components of the reduced dimension slot are used.}

\item{rand_seed}{(optional) numeric scalar that can be passed to
\code{set.seed} to make plots reproducible.}

\item{sigma}{argument passed to \code{\link[destiny]{DiffusionMap}}}

\item{distance}{argument passed to \code{\link[destiny]{DiffusionMap}}}

\item{...}{further arguments passed to \code{\link[destiny]{DiffusionMap}}}

\item{colour_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to colour the points in the plot. Alternatively, a
data frame with one column containing values to map to colours for all cells.}

\item{shape_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to define the shape of the points in the plot.}

\item{size_by}{character string defining the column of \code{pData(object)} to
be used as a factor by which to define the size of points in the plot.}

\item{return_SCE}{logical, should the function return an \code{SingleCellExperiment}
object with principal component values for cells in the
\code{reducedDims} slot. Default is \code{FALSE}, in which case a
\code{ggplot} object is returned.}

\item{draw_plot}{logical, should the plot be drawn on the current graphics
device? Only used if \code{return_SCE} is \code{TRUE}, otherwise the plot
is always produced.}

\item{theme_size}{numeric scalar giving default font size for plotting theme
(default is 10).}

\item{legend}{character, specifying how the legend(s) be shown? Default is
\code{"auto"}, which hides legends that have only one level and shows others.
Alternatives are "all" (show all legends) or "none" (hide all legends).}

\item{rerun}{logical, should PCA be recomputed even if \code{object} contains a
"PCA" element in the \code{reducedDims} slot?}
}
\value{
If \code{return_SCE} is \code{TRUE}, then the function returns an
\code{SingleCellExperiment} object, otherwise it returns a \code{ggplot} object.
}
\description{
Produce a diffusion map plot of two components for an \code{SingleCellExperiment} dataset.
}
\details{
The function \code{\link[destiny]{DiffusionMap}} is used internally
to compute the diffusion map.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)
drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]

\dontrun{
## Examples plotting diffusion maps
plotDiffusionMap(example_sce)
plotDiffusionMap(example_sce, colour_by = "Cell_Cycle")
plotDiffusionMap(example_sce, colour_by = "Cell_Cycle",
shape_by = "Treatment")
plotDiffusionMap(example_sce, colour_by = "Cell_Cycle",
shape_by = "Treatment", size_by = "Mutation_Status")
plotDiffusionMap(example_sce, shape_by = "Treatment",
size_by = "Mutation_Status")
plotDiffusionMap(example_sce, feature_set = 1:100, colour_by = "Treatment",
shape_by = "Mutation_Status")

plotDiffusionMap(example_sce, shape_by = "Treatment",
return_SCE = TRUE)
}

}
\references{
Haghverdi L, Buettner F, Theis FJ. Diffusion maps for high-dimensional single-cell analysis of differentiation data. Bioinformatics. 2015; doi:10.1093/bioinformatics/btv325
}
\seealso{
\code{\link[destiny]{destiny}}
}
