% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotExplanatoryVariables}
\alias{plotExplanatoryVariables}
\title{Plot explanatory variables ordered by percentage of phenotypic variance explained}
\usage{
plotExplanatoryVariables(object, method = "density", exprs_values = "exprs",
  nvars_to_plot = 10, min_marginal_r2 = 0, variables = NULL,
  return_object = FALSE, theme_size = 10, ...)
}
\arguments{
\item{object}{an SCESet object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{method}{character scalar indicating the type of plot to produce. If
"density", the function produces a density plot of R-squared values for each
variable when fitted as the only explanatory variable in a linear model. If
"pairs", then the function produces a pairs plot of the explanatory variables
ordered by the percentage of feature expression variance (as measured by
R-squared in a marginal linear model) explained.}

\item{exprs_values}{which slot of the \code{assayData} in the \code{object}
should be used to define expression? Valid options are "exprs" (default),
"tpm", "fpkm", "cpm", and "counts".}

\item{nvars_to_plot}{integer, the number of variables to plot in the pairs
plot. Default value is 10.}

\item{min_marginal_r2}{numeric scalar giving the minimal value required for
median marginal R-squared for a variable to be plotted. Only variables with a
median marginal R-squared strictly larger than this value will be plotted.}

\item{variables}{optional character vector giving the variables to be plotted.
Default is \code{NULL}, in which case all variables in \code{pData(object)}
are considered and the \code{nvars_to_plot} variables with the highest median
marginal R-squared are plotted.}

\item{return_object}{logical, should an \code{SCESet} object with median
marginal R-squared values added to \code{varMetadata(object)} be returned?}

\item{theme_size}{numeric scalar giving font size to use for the plotting
theme}

\item{...}{parameters to be passed to \code{\link{pairs}}.}
}
\value{
A ggplot object
}
\description{
Plot explanatory variables ordered by percentage of phenotypic variance explained
}
\details{
If the \code{method} argument is "pairs", then the function produces
a pairs plot of the explanatory variables ordered by the percentage of
feature expression variance (as measured by R-squared in a marginal linear
model) explained by variable. Median percentage R-squared is reported on the
plot for each variable. Discrete variables are coerced to a factor and
plotted as integers with jittering. Variables with only one unique value are
quietly ignored.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
rownames(pd) <- pd$Cell
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
drop_genes <- apply(exprs(example_sceset), 1, function(x) {var(x) == 0})
example_sceset <- example_sceset[!drop_genes, ]
example_sceset <- calculateQCMetrics(example_sceset)
vars <- names(pData(example_sceset))[c(2:3, 5:14)]
plotExplanatoryVariables(example_sceset, variables=vars)

}
