% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exprs.R
\name{plotExpression}
\alias{plotExpression}
\title{Plot expression values for a set of features (e.g. genes or transcripts)}
\usage{
plotExpression(object, features, x = NULL, exprs_values = "logcounts",
  log2_values = FALSE, colour_by = NULL, shape_by = NULL,
  size_by = NULL, ncol = 2, xlab = NULL, show_median = FALSE,
  show_violin = TRUE, show_smooth = FALSE, theme_size = 10, alpha = 0.6,
  size = NULL, legend = "auto", scales = "fixed", one_facet = FALSE,
  se = TRUE, jitter = "swarm")
}
\arguments{
\item{object}{A SingleCellExperiment object containing expression values and
experimental information.}

\item{features}{a character vector of feature names or Boolean
vector or numeric vector of indices indicating which features should have
their expression values plotted}

\item{x}{character string providing a column name of \code{pData(object)} or
a feature name (i.e. gene or transcript) to plot on the x-axis in the
expression plot(s). If a feature name, then expression values for the feature
will be plotted on the x-axis for each subplot.}

\item{exprs_values}{character string indicating which values should be used
as the expression values for this plot. Valid arguments are \code{"tpm"}
(transcripts per million), \code{"norm_tpm"} (normalised TPM
values), \code{"fpkm"} (FPKM values), \code{"norm_fpkm"} (normalised FPKM
values), \code{"counts"} (counts for each feature), \code{"norm_counts"},
\code{"cpm"} (counts-per-million), \code{"norm_cpm"} (normalised
counts-per-million), \code{"logcounts"} (log-transformed count data; default),
\code{"norm_exprs"} (normalised
expression values) or \code{"stand_exprs"} (standardised expression values)
or any other slots that have been added to the \code{"assayData"} slot by
the user.}

\item{log2_values}{should the expression values be transformed to the
log2-scale for plotting (with an offset of 1 to avoid logging zeroes)?}

\item{colour_by}{optional character string supplying name of a column of
\code{pData(object)} which will be used as a variable by which to colour
expression values on the plot. Alternatively, a data frame with one column,
containing a value for each cell that will be mapped to a colour.}

\item{shape_by}{optional character string supplying name of a column of
\code{pData(object)} which will be used as a variable to define the shape of
points for expression values on the plot. Alternatively, a data frame with
one column containing values to map to shapes.}

\item{size_by}{optional character string supplying name of a column of
\code{pData(object)} which will be used as a variable to define the size of
points for expression values on the plot. Alternatively, a data frame with
one column containing values to map to sizes.}

\item{ncol}{number of columns to be used for the panels of the plot}

\item{xlab}{label for x-axis; if \code{NULL} (default), then \code{x} will be
used as the x-axis label}

\item{show_median}{logical, show the median for each group on the plot}

\item{show_violin}{logical, show a violin plot for the distribution
for each group on the plot}

\item{show_smooth}{logical, show a smoothed fit through the expression values
on the plot}

\item{theme_size}{numeric scalar giving default font size for plotting theme
(default is 10)}

\item{alpha}{numeric value between 0 (completely transparent) and 1 (completely
solid) defining how transparent plotted points (cells) should be.
Points are jittered horizontally if the x-axis value is categorical rather
than numeric to avoid overplotting.}

\item{size}{numeric scalar optionally providing size for points if
\code{size_by} argument is not given. Default is \code{NULL}, in which case
\pkg{ggplot2} default is used.}

\item{legend}{character, specifying how the legend(s) be shown? Default is
\code{"auto"}, which hides legends that have only one level and shows others.
Alternatives are "all" (show all legends) or "none" (hide all legends).}

\item{scales}{character scalar, should scales be fixed ("fixed"),
free ("free"), or free in one dimension ("free_x"; "free_y", the default).
Passed to the \code{scales} argument in the \code{\link[ggplot2]{facet_wrap}} function
from the \code{ggplot2} package.}

\item{se}{logical, should standard errors be shown (default \code{TRUE}) for
the smoothed fit through the cells. (Ignored if \code{show_smooth} is \code{FALSE}).}

\item{jitter}{character scalar to define whether points are to be jittered
(\code{"jitter"}) or presented in a "beeswarm" style (if \code{"swarm"}; default).
"Beeswarm" style usually looks more attractive, but for datasets with a large
number of cells, or for dense plots, the jitter option may work better.}
}
\value{
a ggplot plot object
}
\description{
Plot expression values for a set of features (e.g. genes or transcripts)
}
\details{
Plot expression values (default log2(counts-per-million +
1), if available) for a set of features.
}
\examples{
## prepare data
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- calculateQCMetrics(example_sce)
sizeFactors(example_sce) <- colSums(counts(example_sce))
example_sce <- normalize(example_sce)

## default plot
plotExpression(example_sce, 1:15)
plotExpression(example_sce, 1:15, jitter = "jitter")

## plot expression against an x-axis value
plotExpression(example_sce, 1:6, colour_by = "Mutation_Status")
plotExpression(example_sce, 1:6, colour_by = "Mutation_Status",
     shape_by = "Treatment", size_by = "Gene_0010")

## explore options
plotExpression(example_sce, 1:6, x = "Mutation_Status", exprs_values = "logcounts",
    colour_by = "Cell_Cycle", show_violin = TRUE, show_median = TRUE)
plotExpression(example_sce, 1:6, x = "Mutation_Status", exprs_values = "counts",
    colour_by = "Cell_Cycle", show_violin = TRUE, show_median = TRUE)

plotExpression(example_sce, "Gene_0001", x = "Mutation_Status")
plotExpression(example_sce, c("Gene_0001", "Gene_0004"), x="Mutation_Status")
plotExpression(example_sce, "Gene_0001", x = "Gene_0002")
plotExpression(example_sce, c("Gene_0001", "Gene_0004"), x="Gene_0002")
## plot expression against expression values for Gene_0004
plotExpression(example_sce, 1:4, "Gene_0004")
plotExpression(example_sce, 1:4, "Gene_0004", show_smooth = TRUE)
plotExpression(example_sce, 1:4, "Gene_0004", show_smooth = TRUE, se = FALSE)

}
