% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExprsFreqVsMean.R
\name{plotExprsFreqVsMean}
\alias{plotExprsFreqVsMean}
\title{Plot frequency against mean for each feature}
\usage{
plotExprsFreqVsMean(object, freq_exprs, mean_exprs, controls,
  exprs_values = "counts", by_show_single = FALSE,
  show_smooth = TRUE, show_se = TRUE, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{freq_exprs}{String specifying the column-level metadata field containing the number of expressing cells per feature.
Alternatively, an \link{AsIs} vector or data.frame, see \code{?\link{retrieveFeatureInfo}}.}

\item{mean_exprs}{String specifying the column-level metadata fielcontaining the mean expression of each feature.
Alternatively, an \link{AsIs} vector or data.frame, see \code{?\link{retrieveFeatureInfo}}.}

\item{controls}{Deprecated and ignored.}

\item{exprs_values}{String specifying the assay used for the default \code{freq_exprs} and \code{mean_exprs}.
This can be set to, e.g., \code{"logcounts"} so that \code{freq_exprs} defaults to \code{"n_cells_by_logcounts"}.}

\item{by_show_single}{Deprecated and ignored.}

\item{show_smooth}{Logical scalar, should a smoothed fit be shown on the plot? 
See \code{\link[ggplot2]{geom_smooth}} for details.}

\item{show_se}{Logical scalar, should the standard error be shown for a smoothed fit?}

\item{...}{Further arguments passed to \code{\link{plotRowData}}.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot the frequency of expression (i.e., percentage of expressing cells) against the mean expression level for each feature in a SingleCellExperiment object.
This is deprecated in favour of directly using \code{\link{plotRowData}}.
}
\details{
This function plots gene expression frequency versus mean expression level, which can be useful to assess the effects of technical dropout in the dataset. 
We fit a non-linear least squares curve for the relationship between expression frequency and mean expression.
We use this curve to define the number of genes above high technical dropout and the numbers of genes that are expressed in at least 50\% and at least 25\% of cells.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- calculateQCMetrics(example_sce, 
    feature_controls = list(set1 = 1:500))
plotExprsFreqVsMean(example_sce)

plotExprsFreqVsMean(example_sce, size_by = "is_feature_control")

}
\seealso{
\code{\link{plotRowData}}
}
