% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotFeatureData}
\alias{plotFeatureData}
\alias{plotRowData}
\title{Plot feature (gene) data from a SingleCellExperiment object}
\usage{
plotFeatureData(object, aesth = aes_string(x = "n_cells_counts", y =
  "log10_total_counts"), ...)

plotRowData(...)
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object containing
expression values and experimental information. Must have been appropriately prepared.}

\item{aesth}{aesthetics function call to pass to ggplot. This function
expects at least x and y variables to be supplied. The default is to produce
a density plot of number of cells expressing the feature (requires
\code{calculateQCMetrics} to have been run on the \code{SingleCellExperiment} object prior).}

\item{...}{arguments passed to \code{\link{plotMetadata}}, e.g.
\code{theme_size}, \code{size}, \code{alpha}, \code{shape}.}
}
\value{
a ggplot plot object
}
\description{
\code{plotFeatureData} and \code{plotRowData} are synonymous.
}
\details{
Plot feature (gene) data from an SingleCellExperiment object. If one variable is
supplied then a density plot will be returned. If both variables are
continuous (numeric) then a scatter plot will be returned. If one variable is
discrete and one continuous then a violin plot with jittered points overlaid
will be returned. If both variables are discrete then a jitter plot will be
produced. The object returned is a ggplot object, so further layers and
plotting options (titles, facets, themes etc) can be added.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- calculateQCMetrics(example_sce)
plotFeatureData(example_sce, aesth = aes(x = n_cells_counts, y = log10_total_counts))

plotRowData(example_sce, aesth = aes(x = n_cells_counts, y = log10_total_counts))

}
