% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap of gene expression values}
\usage{
plotHeatmap(object, features, columns = NULL, exprs_values = "logcounts",
  center = FALSE, zlim = NULL, symmetric = FALSE, color = NULL,
  colour_columns_by = NULL, ...)
}
\arguments{
\item{object}{A \code{\link{SingleCellExperiment}} object.}

\item{features}{A character vector containing row names of \code{object} 
to be shown as rows in the heatmap.}

\item{columns}{A vector specifying the subset of columns in \code{object}
to show as columns in the heatmp. All columns are used by default.}

\item{exprs_values}{A string indicating which assay of \code{object}
should be used as expression values for colouring in the heatmap.}

\item{center}{A logical scalar indicating whether each row should be 
centered at zero prior to plotting. Setting this to \code{TRUE}}

\item{zlim}{A numeric vector of length 2, specifying the upper and lower
bounds for the expression values. This winsorizes the expression matrix
prior to plotting (but after centering, if \code{center=TRUE}. If 
\code{NULL}, it defaults to the range of the expression matrix.}

\item{symmetric}{A logical scalar specifying whether the default \code{zlim}
should be symmetric around zero. If \code{TRUE}, the maximum absolute
expression value will be computed and multiplied by \code{c(-1, 1)} to 
define \code{zlim}.}

\item{color}{A vector of colours specifying the palette to use for mapping 
expression values to colours. This defaults to the default setting in 
\code{\link[pheatmap]{pheatmap}}.}

\item{colour_columns_by}{A character vector of column metadata names 
in \code{colData(object)} or of gene names, to use to annotate the columns
by colour.}

\item{...}{Additional arguments to pass to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
A heatmap is produced on the current graphics device. The output
of \code{\link[pheatmap]{pheatmap}} is invisibly returned.
}
\description{
Plot heatmap of gene expression values
}
\details{
Setting \code{center=TRUE} is useful for examining log-fold 
changes of each cell's expression profile from the average across all cells.
This avoids issues with the entire row appearing a certain colour because
the gene is highly/lowly expressed across all cells.

Setting \code{zlim} preserves the dynamic range of colours in the presence 
of outliers. Otherwise, the plot may be dominated by a few genes,
which will \dQuote{flatten} the observed colours for the rest of the heatmap.
}
\examples{
example(normalizeSCE) # borrowing the example objects in here.
plotHeatmap(example_sce, features=rownames(example_sce)[1:10])
plotHeatmap(example_sce, features=rownames(example_sce)[1:10],
    center=TRUE, symmetric=TRUE)

plotHeatmap(example_sce, features=rownames(example_sce)[1:10],
    colour_columns_by=c("Mutation_Status", "Cell_Cycle"))

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\author{
Aaron Lun
}
