% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot heatmap of gene expression values}
\usage{
plotHeatmap(object, features, columns = NULL,
  exprs_values = "logcounts", center = FALSE, zlim = NULL,
  symmetric = FALSE, color = NULL, colour_columns_by = NULL,
  by_exprs_values = exprs_values, by_show_single = FALSE,
  show_colnames = TRUE, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{features}{A character vector of row names, a logical vector of integer vector of indices specifying rows of \code{object} to show in the heatmap.}

\item{columns}{A vector specifying the subset of columns in \code{object} to show as columns in the heatmp. 
By default, all columns are used in their original order.}

\item{exprs_values}{A string or integer scalar indicating which assay of \code{object} should be used as expression values for colouring in the heatmap.}

\item{center}{A logical scalar indicating whether each row should have its mean expression centered at zero prior to plotting.}

\item{zlim}{A numeric vector of length 2, specifying the upper and lower bounds for the expression values. 
This winsorizes the expression matrix prior to plotting (but after centering, if \code{center=TRUE}). 
If \code{NULL}, it defaults to the range of the expression matrix.}

\item{symmetric}{A logical scalar specifying whether the default \code{zlim} should be symmetric around zero. 
If \code{TRUE}, the maximum absolute value of \code{zlim} will be computed and multiplied by \code{c(-1, 1)} to redefine \code{zlim}.}

\item{color}{A vector of colours specifying the palette to use for mapping expression values to colours. 
This defaults to the default setting in \code{\link[pheatmap]{pheatmap}}.}

\item{colour_columns_by}{A list of values specifying how the columns should be annotated with colours.
Each entry of the list can be any acceptable input to the \code{by} argument in \code{?\link{retrieveCellInfo}}.
A character vector can also be supplied and will be treated as a list of strings.}

\item{by_exprs_values}{A string or integer scalar specifying which assay to obtain expression values from, 
for colouring of column-level data - see the \code{exprs_values} argument in \code{?\link{retrieveCellInfo}}.}

\item{by_show_single}{Deprecated and ignored.}

\item{show_colnames}{Logical scalar specifying whether column names should be shown, if available in \code{object}.}

\item{...}{Additional arguments to pass to \code{\link[pheatmap]{pheatmap}}.}
}
\value{
A heatmap is produced on the current graphics device. 
The output of \code{\link[pheatmap]{pheatmap}} is invisibly returned.
}
\description{
Create a heatmap of expression values for each cell and specified features in a SingleCellExperiment object.
}
\details{
Setting \code{center=TRUE} is useful for examining log-fold changes of each cell's expression profile from the average across all cells.
This avoids issues with the entire row appearing a certain colour because the gene is highly/lowly expressed across all cells.

Setting \code{zlim} preserves the dynamic range of colours in the presence  of outliers. 
Otherwise, the plot may be dominated by a few genes, which will \dQuote{flatten} the observed colours for the rest of the heatmap.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

plotHeatmap(example_sce, features=rownames(example_sce)[1:10])

plotHeatmap(example_sce, features=rownames(example_sce)[1:10],
    center=TRUE, symmetric=TRUE)

plotHeatmap(example_sce, features=rownames(example_sce)[1:10],
    colour_columns_by=c("Mutation_Status", "Cell_Cycle"))

}
\seealso{
\code{\link[pheatmap]{pheatmap}}
}
\author{
Aaron Lun
}
