% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlatePosition.R
\name{plotPlatePosition}
\alias{plotPlatePosition}
\title{Plot cells in plate positions}
\usage{
plotPlatePosition(object, plate_position = NULL, colour_by = NULL,
  size_by = NULL, shape_by = NULL, legend = "auto",
  exprs_values = "logcounts", theme_size = 24, alpha = 0.6, size = 24)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{plate_position}{A character vector specifying the plate position for each cell (e.g., A01, B12, and so on, where letter indicates row and number indicates column).
If \code{NULL}, the function will attempt to extract this from \code{object$plate_position}.
Alternatively, a list of two factors (\code{"row"} and \code{"column"}) can be supplied, specifying the row and column for each cell in \code{object}.}

\item{colour_by}{Specification of a column metadata field or a feature to colour by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{size_by}{Specification of a column metadata field or a feature to size by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{shape_by}{Specification of a column metadata field or a feature to shape by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{legend}{String specifying how the legend(s) be shown, see \code{?"\link{scater-plot-args}"} for details.}

\item{exprs_values}{A string or integer scalar specifying which assay in \code{assays(object)} to obtain expression values from, for use in colouring, shaping or sizing.}

\item{theme_size}{Numeric scalar, see \code{?"\link{scater-plot-args}"} for details.}

\item{alpha}{Numeric scalar specifying the transparency of the points, see \code{?"\link{scater-plot-args}"} for details.}

\item{size}{Numeric scalar specifying the size of the points, see \code{?"\link{scater-plot-args}"} for details.}
}
\value{
A ggplot object.
}
\description{
Plots cells in their position on a plate, coloured by metadata variables or feature expression values from a SingleCellExperiment object.
}
\details{
This function expects plate positions to be given in a charcter format where a letter indicates the row on the plate and a numeric value  indicates the column. 
Each cell has a plate position such as "A01", "B12", "K24" and so on. 
From these plate positions, the row is extracted as the letter, and the column as the numeric part. 
Alternatively, the row and column identities can be directly supplied by setting \code{plate_position} as a list of two factors.
}
\examples{
## prepare data
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)
example_sce <- calculateQCMetrics(example_sce)

## define plate positions
example_sce$plate_position <- paste0(
    rep(LETTERS[1:5], each = 8), 
    rep(formatC(1:8, width = 2, flag = "0"), 5)
)

## plot plate positions
plotPlatePosition(example_sce, colour_by = "Mutation_Status")

plotPlatePosition(example_sce, shape_by = "Treatment", colour_by = "Gene_0004")

plotPlatePosition(example_sce, shape_by = "Treatment", size_by = "Gene_0001",
    colour_by = "Cell_Cycle")

}
\author{
Davis McCarthy, with modifications by Aaron Lun
}
