% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQC.R
\name{plotQC}
\alias{plotQC}
\title{Produce QC diagnostic plots}
\usage{
plotQC(object, type = c("highest-expression", "find-pcs",
  "explanatory-variables", "exprs-freq-vs-mean"), ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{type}{String specifying the type of QC plot to compute.}

\item{...}{Arguments passed to further plotting functions.}
}
\value{
A ggplot plot object.
}
\description{
Produce QC diagnostic plots
}
\details{
This is a wrapper function to call a variety of different plotting functions.
It has been deprecated in favour of calling the target functions specifically, which involves less typing anyway!
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
  assays = list(counts = sc_example_counts), 
  colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

example_sce <- calculateQCMetrics(example_sce)
plotQC(example_sce, type="high", colour_cells_by="Mutation_Status")
}
