% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.R
\name{plotQC}
\alias{plotQC}
\title{Produce QC diagnostic plots}
\usage{
plotQC(object, type = "highest-expression", ...)
}
\arguments{
\item{object}{an SingleCellExperiment object containing expression values and
experimental information. Must have been appropriately prepared.}

\item{type}{character scalar providing type of QC plot to compute:
"highest-expression" (showing features with highest expression), "find-pcs" (showing
the most important principal components for a given variable),
"explanatory-variables" (showing a set of explanatory variables plotted
against each other, ordered by marginal variance explained), or
"exprs-mean-vs-freq" (plotting the mean expression levels against the
frequency of expression for a set of features).}

\item{...}{arguments passed to \code{\link{plotHighestExprs}},
\code{\link{findImportantPCs}}, \code{\link{plotExplanatoryVariables}} and
\code{{plotExprsMeanVsFreq}} as appropriate.}
}
\value{
a ggplot plot object
}
\description{
Produce QC diagnostic plots
}
\details{
Display useful quality control plots to help with pre-processing
of data and identification of potentially problematic features and cells.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), 
colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]
example_sce <- calculateQCMetrics(example_sce)
plotQC(example_sce, type="high", colour_cells_by="Mutation_Status")
plotQC(example_sce, type="find", variable="total_features")
vars <- names(colData(example_sce))[c(2:3, 5:14)]
plotQC(example_sce, type="expl", variables=vars)

}
