% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotReducedDim.R
\name{plotReducedDim}
\alias{plotReducedDim}
\title{Plot reduced dimensions}
\usage{
plotReducedDim(object, use_dimred, ncomponents = 2, percentVar = NULL,
  colour_by = NULL, shape_by = NULL, size_by = NULL,
  by_exprs_values = "logcounts", by_show_single = FALSE, ...,
  add_ticks = TRUE)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{use_dimred}{A string or integer scalar indicating the reduced dimension result in \code{reducedDims(object)} to plot.}

\item{ncomponents}{A numeric scalar indicating the number of dimensions to plot, starting from the first dimension.}

\item{percentVar}{A numeric vector giving the proportion of variance in expression explained by each reduced dimension. 
Only expected to be used in PCA settings, e.g., in the \code{\link[scater]{plotPCA}} function.}

\item{colour_by}{Specification of a column metadata field or a feature to colour by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{shape_by}{Specification of a column metadata field or a feature to shape by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{size_by}{Specification of a column metadata field or a feature to size by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{by_exprs_values}{A string or integer scalar specifying which assay to obtain expression values from, 
for use in point aesthetics - see \code{?"\link{scater-vis-var}"} for details.}

\item{by_show_single}{Logical scalar specifying whether single-level factors should be used for point aesthetics, see \code{?"\link{scater-vis-var}"} for details.}

\item{...}{Additional arguments for visualization, see \code{?"\link{scater-plot-args}"} for details.}

\item{add_ticks}{Logical scalar indicating whether ticks should be drawn on the axes corresponding to the location of each point.}
}
\value{
A ggplot object
}
\description{
Plot cell-level reduced dimension results stored in a SingleCellExperiment object.
}
\details{
If \code{ncomponents} is 2, a scatterplot of the first two dimensions is produced. 
If \code{ncomponents} is greater than 2, a pairs plots for the top dimensions is produced.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)
drop_genes <- apply(exprs(example_sce), 1, function(x) {var(x) == 0})
example_sce <- example_sce[!drop_genes, ]

example_sce <- runPCA(example_sce, ncomponents=5)
plotReducedDim(example_sce, "PCA")
plotReducedDim(example_sce, "PCA", colour_by="Cell_Cycle")
plotReducedDim(example_sce, "PCA", colour_by="Cell_Cycle", shape_by="Treatment")

plotReducedDim(example_sce, "PCA", colour_by="Gene_0001")
plotReducedDim(example_sce, "PCA", colour_by="Cell_Cycle", 
   shape_by="Mutation_Status", size_by="Gene_0001")

plotReducedDim(example_sce, "PCA", ncomponents=5)
plotReducedDim(example_sce, "PCA", ncomponents=5, colour_by="Cell_Cycle", 
    shape_by="Treatment")

}
\author{
Davis McCarthy, with modifications by Aaron Lun
}
