% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plotRowData}
\alias{plotRowData}
\alias{plotFeatureData}
\title{Plot row metadata}
\usage{
plotRowData(object, y, x = NULL, colour_by = NULL, shape_by = NULL,
  size_by = NULL, by_exprs_values = "logcounts",
  by_show_single = FALSE, ...)

plotFeatureData(...)
}
\arguments{
\item{object}{A SingleCellExperiment object containing expression values and experimental information.}

\item{y}{Specification of the row-level metadata to show on the y-axis, see \code{?"\link{scater-vis-var}"} for possible values.
Note that only metadata fields will be searched, \code{assays} will not be used.}

\item{x}{Specification of the row-level metadata to show on the x-axis, see \code{?"\link{scater-vis-var}"} for possible values.
Again, only metadata fields will be searched, \code{assays} will not be used.}

\item{colour_by}{Specification of a row metadata field or a cell to colour by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{shape_by}{Specification of a row metadata field or a cell to shape by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{size_by}{Specification of a row metadata field or a cell to size by, see \code{?"\link{scater-vis-var}"} for possible values.}

\item{by_exprs_values}{A string or integer scalar specifying which assay to obtain expression values from, 
for use in point aesthetics - see \code{?"\link{scater-vis-var}"} for details.}

\item{by_show_single}{Logical scalar specifying whether single-level factors should be used for point aesthetics, see \code{?"\link{scater-vis-var}"} for details.}

\item{...}{Additional arguments for visualization, see \code{?"\link{scater-plot-args}"} for details.}
}
\value{
A ggplot object.
}
\description{
Plot row-level (i.e., gene) metadata from a SingleCellExperiment object.
}
\details{
If \code{y} is continuous and \code{x=NULL}, a violin plot is generated.
If \code{x} is categorical, a grouped violin plot will be generated, with one violin for each level of \code{x}.
If \code{x} is continuous, a scatter plot will be generated.

If \code{y} is categorical and \code{x} is continuous, horizontal violin plots will be generated.
If \code{x} is missing or categorical, rectangule plots will be generated where the area of a rectangle is proportional to the number of points for a combination of factors.

Note that \code{plotFeatureData} is a synonym for \code{plotRowData}.
This is an artifact of the transition from the old SCESet class, and will be deprecated in future releases.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- calculateQCMetrics(example_sce,
    feature_controls = list(ERCC=1:40))
example_sce <- normalize(example_sce)

plotRowData(example_sce, y="n_cells_by_counts", x="log10_total_counts")
plotRowData(example_sce, y="n_cells_by_counts", 
   size_by ="log10_total_counts",
   colour_by = "is_feature_control")

}
