% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10ximport-wrapper.R
\name{read10xResults}
\alias{read10xResults}
\alias{read10XResults}
\alias{read10XResults}
\title{Load in data from 10x experiment}
\usage{
read10xResults(data_dir, min_total_cell_counts = NULL,
  min_mean_gene_counts = NULL)

read10XResults(...)
}
\arguments{
\item{data_dir}{Directory containing the matrix.mtx, genes.tsv, and 
barcodes.tsv
files provided by 10x. A vector or named vector can be given in order to load 
several data directories. If a named vector is given, the cell barcode names 
will be prefixed with the name.}

\item{min_total_cell_counts}{integer(1) threshold such that cells (barcodes)
with total counts below the threshold are filtered out}

\item{min_mean_gene_counts}{numeric(1) threshold such that genes with mean 
counts below the threshold are filtered out.}

\item{...}{passed arguments}
}
\value{
Returns an SingleCellExperiment object with counts data stored as a
sparse matrix with rows and columns labeled.
}
\description{
Creates a full or sparse matrix from a sparse data matrix provided by 10X 
genomics.
}
\details{
This function was developed from the \code{Read10X} function from 
the \code{Seurat} package.
}
\examples{
sce10x <- read10xResults(system.file("extdata", package="scater"))

}
