% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tximport-wrapper.R
\name{readTxResults}
\alias{readTxResults}
\title{Read transcript quantification data with tximport package}
\usage{
readTxResults(samples = NULL, files = NULL, log = NULL,
  type = "kallisto", txOut = TRUE, logExprsOffset = 1, verbose = TRUE,
  ...)
}
\arguments{
\item{samples}{character vector providing a set of sample names to use for
the abundance results.}

\item{files}{character vector providing a set of filenames containing
kallisto abundance results to be read in.}

\item{log}{list (optional), generated by \code{runKallisto}. If provided, then
\code{samples} and \code{files} arguments are ignored.}

\item{type}{character, the type of software used to generate the abundances.
Options are "kallisto", "salmon", "sailfish", "rsem". This argument is passed
to \code{\link[tximport]{tximport}}.}

\item{txOut}{logical, whether the function should just output transcript-level (default TRUE)}

\item{logExprsOffset}{numeric scalar, providing the offset used when doing
log2-transformations of expression data to avoid trying to take logs of zero.
Default offset value is \code{1}.}

\item{verbose}{logical, should function provide output about progress?}

\item{...}{optional parameters passed to \code{\link[tximport]{tximport}}. 
See documentation for \code{\link[tximport]{tximport}} for options and 
details.}
}
\value{
an \code{SCESet} object containing the abundance, count and feature
length data from the supplied samples.
}
\description{
After generating transcript/feature abundance results using kallisto, Salmon,
Sailfish or RSEM for a batch of samples, read these abundance values into an
\code{SCESet} object.
}
\details{
Note: tximport does not import bootstrap estimates from kallisto, 
Salmon, or Sailfish. If you want bootstrap estimates use the 
\code{\link{readKallistoResults}} or \code{\link{readSalmonResults}} 
functions.
}
\examples{
\dontrun{
## this example requires installation of the tximportData package from 
## Bioconductor 
library(tximportData)
dir <- system.file("extdata", package = "tximportData")
list.files(dir)
samples <- read.table(file.path(dir, "samples.txt"), header = TRUE)
samples
directories <- file.path(dir, "kallisto", samples$run)
names(directories) <- paste0("sample", 1:6)
files <- file.path(directories, "abundance.tsv")
sce_example <- readTxResults(samples = names(directories), 
files = files, type = "kallisto")

## for faster reading of results use the read_tsv function from the readr pkg
library(readr)
sce_example <- readTxResults(samples = names(directories), 
files = files, type = "kallisto", reader = read_tsv)
}

}
\references{
Soneson C, Love MI, Robinson MD. Differential analyses for RNA-seq: transcript-level estimates improve gene-level inferences. F1000Res. 2015;4: 1521.
}
