% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{addPerCellQC}
\alias{addPerFeatureQC}
\alias{aggregateAcrossCells}
\alias{aggregateAcrossFeatures}
\alias{calculateAverage}
\alias{calculateCPM}
\alias{calculateFPKM}
\alias{calculateTPM}
\alias{computeLibraryFactors}
\alias{computeMedianFactors}
\alias{isOutlier}
\alias{librarySizeFactors}
\alias{logNormCounts}
\alias{makePerCellDF}
\alias{makePerFeatureDF}
\alias{medianSizeFactors}
\alias{mockSCE}
\alias{normalizeCounts}
\alias{numDetectedAcrossCells}
\alias{numDetectedAcrossFeatures}
\alias{perCellQCMetrics}
\alias{perFeatureQCMetrics}
\alias{quickPerCellQC}
\alias{readSparseCounts}
\alias{sumCountsAcrossCells}
\alias{sumCountsAcrossFeatures}
\alias{uniquifyFeatureNames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{scuttle}{\code{\link[scuttle:addPerCellQC]{addPerCellQC()}}, \code{\link[scuttle:addPerFeatureQC]{addPerFeatureQC()}}, \code{\link[scuttle:aggregateAcrossCells]{aggregateAcrossCells()}}, \code{\link[scuttle:aggregateAcrossFeatures]{aggregateAcrossFeatures()}}, \code{\link[scuttle:calculateAverage]{calculateAverage()}}, \code{\link[scuttle:calculateCPM]{calculateCPM()}}, \code{\link[scuttle:calculateFPKM]{calculateFPKM()}}, \code{\link[scuttle:calculateTPM]{calculateTPM()}}, \code{\link[scuttle:computeLibraryFactors]{computeLibraryFactors()}}, \code{\link[scuttle:computeMedianFactors]{computeMedianFactors()}}, \code{\link[scuttle:isOutlier]{isOutlier()}}, \code{\link[scuttle:librarySizeFactors]{librarySizeFactors()}}, \code{\link[scuttle:logNormCounts]{logNormCounts()}}, \code{\link[scuttle:makePerCellDF]{makePerCellDF()}}, \code{\link[scuttle:makePerFeatureDF]{makePerFeatureDF()}}, \code{\link[scuttle:medianSizeFactors]{medianSizeFactors()}}, \code{\link[scuttle:mockSCE]{mockSCE()}}, \code{\link[scuttle:normalizeCounts]{normalizeCounts()}}, \code{\link[scuttle:numDetectedAcrossCells]{numDetectedAcrossCells()}}, \code{\link[scuttle:numDetectedAcrossFeatures]{numDetectedAcrossFeatures()}}, \code{\link[scuttle:perCellQCMetrics]{perCellQCMetrics()}}, \code{\link[scuttle:perFeatureQCMetrics]{perFeatureQCMetrics()}}, \code{\link[scuttle:quickPerCellQC]{quickPerCellQC()}}, \code{\link[scuttle:readSparseCounts]{readSparseCounts()}}, \code{\link[scuttle:sumCountsAcrossCells]{sumCountsAcrossCells()}}, \code{\link[scuttle:sumCountsAcrossFeatures]{sumCountsAcrossFeatures()}}, \code{\link[scuttle:uniquifyFeatureNames]{uniquifyFeatureNames()}}}
}}

