% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{addPerCellQC}
\alias{addPerFeatureQC}
\alias{aggregateAcrossCells}
\alias{aggregateAcrossFeatures}
\alias{calculateAverage}
\alias{calculateCPM}
\alias{calculateFPKM}
\alias{calculateTPM}
\alias{computeLibraryFactors}
\alias{computeMedianFactors}
\alias{isOutlier}
\alias{librarySizeFactors}
\alias{logNormCounts}
\alias{makePerCellDF}
\alias{makePerFeatureDF}
\alias{medianSizeFactors}
\alias{mockSCE}
\alias{normalizeCounts}
\alias{numDetectedAcrossCells}
\alias{numDetectedAcrossFeatures}
\alias{perCellQCMetrics}
\alias{perFeatureQCMetrics}
\alias{quickPerCellQC}
\alias{readSparseCounts}
\alias{sumCountsAcrossCells}
\alias{sumCountsAcrossFeatures}
\alias{uniquifyFeatureNames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{scuttle}{\code{\link[scuttle:addPerCellQCMetrics]{addPerCellQC}}, \code{\link[scuttle:addPerCellQCMetrics]{addPerFeatureQC}}, \code{\link[scuttle]{aggregateAcrossCells}}, \code{\link[scuttle]{aggregateAcrossFeatures}}, \code{\link[scuttle]{calculateAverage}}, \code{\link[scuttle]{calculateCPM}}, \code{\link[scuttle]{calculateFPKM}}, \code{\link[scuttle]{calculateTPM}}, \code{\link[scuttle:librarySizeFactors]{computeLibraryFactors}}, \code{\link[scuttle:medianSizeFactors]{computeMedianFactors}}, \code{\link[scuttle]{isOutlier}}, \code{\link[scuttle]{librarySizeFactors}}, \code{\link[scuttle]{logNormCounts}}, \code{\link[scuttle]{makePerCellDF}}, \code{\link[scuttle]{makePerFeatureDF}}, \code{\link[scuttle]{medianSizeFactors}}, \code{\link[scuttle]{mockSCE}}, \code{\link[scuttle]{normalizeCounts}}, \code{\link[scuttle]{numDetectedAcrossCells}}, \code{\link[scuttle]{numDetectedAcrossFeatures}}, \code{\link[scuttle]{perCellQCMetrics}}, \code{\link[scuttle]{perFeatureQCMetrics}}, \code{\link[scuttle]{quickPerCellQC}}, \code{\link[scuttle]{readSparseCounts}}, \code{\link[scuttle]{sumCountsAcrossCells}}, \code{\link[scuttle]{sumCountsAcrossFeatures}}, \code{\link[scuttle]{uniquifyFeatureNames}}}
}}

