% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{rename}
\alias{rename}
\alias{rename}
\alias{rename,SingleCellExperiment-method}
\title{Rename variables of \code{colData(object)}.}
\usage{
rename(object, ...)

\S4method{rename}{SingleCellExperiment}(object, ...)
}
\arguments{
\item{object}{A \code{SingleCellExperiment} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::rename} to
act on \code{colData(object)}.}
}
\value{
An SingleCellExperiment object.
}
\description{
Rename variables of \code{colData(object)}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), 
colData = sc_example_cell_info)
example_sce <- rename(example_sce, Cell_Phase = Cell_Cycle)

}
