% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/verbs.R
\docType{methods}
\name{rename}
\alias{rename}
\alias{rename,SCESet-method}
\alias{rename.SCESet}
\title{Rename variables of \code{pData(object)}.}
\usage{
rename(object, ...)

\S4method{rename}{SCESet}(object, ...)

rename.SCESet(object, ...)
}
\arguments{
\item{object}{A \code{SCESet} object.}

\item{...}{Additional arguments to be passed to \code{dplyr::rename} to
act on \code{pData(object)}.}
}
\description{
Rename variables of \code{pData(object)}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- rename(example_sceset, Cell_Phase = Cell_Cycle)
}

