% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiffusionMap.R
\name{runDiffusionMap}
\alias{runDiffusionMap}
\title{Create a diffusion map from cell-level data}
\usage{
runDiffusionMap(object, ncomponents = 2, ntop = 500,
  feature_set = NULL, exprs_values = "logcounts",
  scale_features = TRUE, use_dimred = NULL, n_dimred = NULL,
  name = "DiffusionMap", ...)
}
\arguments{
\item{object}{A SingleCellExperiment object}

\item{ncomponents}{Numeric scalar indicating the number of diffusion components to obtain.}

\item{ntop}{Numeric scalar specifying the number of most variable features to use for constructing the diffusion map.}

\item{feature_set}{Character vector of row names, a logical vector or a numeric vector of indices indicating a set of features to use to construct the diffusion map. 
This will override any \code{ntop} argument if specified.}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{object} should be used to obtain the expression values for the calculations.}

\item{scale_features}{Logical scalar, should the expression values be standardised so that each feature has unit variance?}

\item{use_dimred}{String or integer scalar specifying the entry of \code{reducedDims(object)} to use as input to \code{\link[destiny]{DiffusionMap}}.
Default is to not use existing reduced dimension results.}

\item{n_dimred}{Integer scalar, number of dimensions of the reduced dimension slot to use when \code{use_dimred} is supplied.
Defaults to all available dimensions.}

\item{name}{String specifying the name to be used to store the result in the \code{reducedDims} of the output.}

\item{...}{Additional arguments to pass to \code{\link[destiny]{DiffusionMap}}.}
}
\value{
A SingleCellExperiment object containing the coordinates of the first \code{ncomponent} diffusion map components for each cell.
By default, this is stored in the \code{"DiffusionMap"} entry of the \code{reducedDims} slot.
}
\description{
Produce a diffusion map for the cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link[destiny]{DiffusionMap}} is used internally to compute the diffusion map.

Setting \code{use_dimred} allows users to easily construct a diffusion map from low-rank approximations of the original expression matrix (e.g., after PCA).
In such cases, arguments such as \code{ntop}, \code{feature_set}, \code{exprs_values} and \code{scale_features} will be ignored. 

The behaviour of \code{\link[destiny]{DiffusionMap}} seems to be non-deterministic, in a manner that is not responsive to any \code{\link{set.seed}} call.
The reason for this is unknown.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runDiffusionMap(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\references{
Haghverdi L, Buettner F, Theis FJ. Diffusion maps for high-dimensional single-cell analysis of differentiation data. Bioinformatics. 2015; doi:10.1093/bioinformatics/btv325
}
\seealso{
\code{\link[destiny]{destiny}},
\code{\link[scater]{plotDiffusionMap}}
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
