% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_dimensions.R
\name{runMDS}
\alias{runMDS}
\title{Run MDS for a SingleCellExperiment object}
\usage{
runMDS(object, ntop = 500, ncomponents = 2, feature_set = NULL,
  exprs_values = "logcounts", scale_features = TRUE, use_dimred = NULL,
  n_dimred = NULL, method = "euclidean")
}
\arguments{
\item{object}{a \code{SingleCellExperiment} object}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the diffusion map. Default is \code{500}, but any \code{ntop}
argument is overrided if the \code{feature_set} argument is non-NULL.}

\item{ncomponents}{numeric scalar indicating the number of MDS dimensions
to obtain.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the diffusion map. If character, entries must all be in
\code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{exprs_values}{Integer or character string indicating which values should be used #' as the expression values for this plot. 
Defaults to \code{"logcounts"}, but any other element of the \code{assays} slot of the \code{SingleCellExperiment} object can be used.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{use_dimred}{character(1), use named reduced dimension representation of cells
stored in \code{SingleCellExperiment} object instead of recomputing (e.g. "PCA").
 Default is \code{NULL}, no reduced dimension values are provided to \code{Rtsne}.}

\item{n_dimred}{integer(1), number of components of the reduced dimension slot
to use. Default is \code{NULL}, in which case (if \code{use_dimred} is not \code{NULL})
all components of the reduced dimension slot are used.}

\item{method}{string specifying the type of distance to be computed between cells.}
}
\value{
A \code{SingleCellExperiment} object containing the coordinates of the 
first \code{ncomponent} MDS dimensions for each cell in the #' \code{"MDS"} 
entry of the \code{reducedDims} slot.
}
\description{
Perform multi-dimensional scaling using data stored in a
\code{SingleCellExperiment} object.
}
\details{
The function \code{\link{cmdscale}} is used internally to
compute the multidimensional scaling components to plot.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

example_sce <- runMDS(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link[scater]{plotMDS}}
}
