% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runMDS.R
\docType{methods}
\name{runMDS}
\alias{runMDS}
\alias{runMDS,ANY-method}
\alias{runMDS,SingleCellExperiment-method}
\title{Perform MDS on cell-level data}
\usage{
runMDS(x, ...)

\S4method{runMDS}{ANY}(x, ncomponents = 2, ntop = 500,
  subset.row = NULL, feature_set = NULL, scale = FALSE,
  scale_features = NULL, transposed = FALSE, method = "euclidean")

\S4method{runMDS}{SingleCellExperiment}(x, ..., use.dimred = NULL,
  use_dimred = NULL, n.dimred = NULL, n_dimred = NULL,
  assay.type = "logcounts", exprs_values = NULL, alt.exp = NULL,
  name = "MDS")
}
\arguments{
\item{x}{A numeric matrix of log-expression values where rows are features and columns are cells.

Alternatively, a \linkS4class{SingleCellExperiment} object containing such a matrix.

Alternatively, if \code{transposed=TRUE}, a numeric matrix where rows are cells and columns are dimensions.}

\item{...}{For the generic, additional arguments to pass to specific methods.

For the SingleCellExperiment method, additional arguments to pass to the ANY method.}

\item{ncomponents}{Numeric scalar indicating the number of MDS?g dimensions to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{subset.row}{Vector specifying the subset of features to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{feature_set}{Deprecated, same as \code{subset.row}.}

\item{scale}{Logical scalar, should the expression values be standardised? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{scale_features}{Deprecated, same as \code{scale} but with a different default.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{method}{String specifying the type of distance to be computed between cells.}

\item{use.dimred}{String or integer scalar specifying the existing dimensionality reduction results to use, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_dimred}{Deprecated, same as \code{use.dimred}.}

\item{n.dimred}{Integer scalar or vector specifying the dimensions to use if \code{use.dimred} is specified, see \code{?"\link{scater-red-dim-args}"}.}

\item{n_dimred}{Deprecated, same as \code{n.dimred}.}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the expression values, see \code{?"\link{scater-red-dim-args}"}.}

\item{exprs_values}{Deprecated, same as \code{assay.type}.}

\item{alt.exp}{String or integer scalar specifying an alternative experiment to use to compute the PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{name}{String specifying the name to be used to store the result in the \code{reducedDims} of the output.}

\item{feature_set}{Character vector of row names, a logical vector or a numeric vector of indices indicating a set of features to use for MDS?g.
This will override any \code{ntop} argument if specified.}
}
\value{
For the ANY method, a matrix is returned containing the MDS coordinates for each cell (row) and dimension (column).

For the \linkS4class{SingleCellExperiment} method, a modified version of \code{x} is returned that contains the MDS coordinates in the \code{"MDS"} entry of the \code{\link{reducedDims}}.
}
\description{
Perform multi-dimensional scaling (MDS) on cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link{cmdscale}} is used internally to compute the MDS components.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runMDS(example_sce, scale_features=NULL)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link{cmdscale}}, to perform the underlying calculations.

\code{\link[scater]{plotMDS}}, to quickly visualize the results.

\code{?"\link{scater-red-dim-args}"}, for a full description of various options.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
