% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runMDS.R
\docType{methods}
\name{calculateMDS}
\alias{calculateMDS}
\alias{runMDS}
\alias{calculateMDS,ANY-method}
\alias{calculateMDS,SummarizedExperiment-method}
\alias{calculateMDS,SingleCellExperiment-method}
\title{Perform MDS on cell-level data}
\usage{
calculateMDS(x, ...)

\S4method{calculateMDS}{ANY}(x, ncomponents = 2, ntop = 500,
  subset.row = NULL, feature_set = NULL, scale = FALSE,
  scale_features = NULL, transposed = FALSE, method = "euclidean")

\S4method{calculateMDS}{SummarizedExperiment}(x, ...,
  assay.type = "logcounts")

\S4method{calculateMDS}{SingleCellExperiment}(x, ...,
  assay.type = "logcounts", exprs_values = NULL, use.dimred = NULL,
  use_dimred = NULL, n.dimred = NULL, n_dimred = NULL,
  alt.exp = NULL)

runMDS(x, ..., name = "MDS")
}
\arguments{
\item{x}{For \code{calculateMDS}, a numeric matrix of log-expression values where rows are features and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.

For \code{runMDS}, a \linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{...}{For the \code{calculateMDS} generic, additional arguments to pass to specific methods.
For the SummarizedExperiment and SingleCellExperiment methods, additional arguments to pass to the ANY method.

For \code{runMDS}, additional arguments to pass to \code{calculateMDS}.}

\item{ncomponents}{Numeric scalar indicating the number of MDS?g dimensions to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{subset.row}{Vector specifying the subset of features to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{feature_set}{Deprecated, same as \code{subset.row}.}

\item{scale}{Logical scalar, should the expression values be standardised? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{scale_features}{Deprecated, same as \code{scale} but with a different default.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows? 
See \code{?"\link{scater-red-dim-args}"} for details.}

\item{method}{String specifying the type of distance to be computed between cells.}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the expression values, see \code{?"\link{scater-red-dim-args}"}.}

\item{exprs_values}{Deprecated, same as \code{assay.type}.}

\item{use.dimred}{String or integer scalar specifying the existing dimensionality reduction results to use, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_dimred}{Deprecated, same as \code{use.dimred}.}

\item{n.dimred}{Integer scalar or vector specifying the dimensions to use if \code{use.dimred} is specified, see \code{?"\link{scater-red-dim-args}"}.}

\item{n_dimred}{Deprecated, same as \code{n.dimred}.}

\item{alt.exp}{String or integer scalar specifying an alternative experiment to use to compute the PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{name}{String specifying the name to be used to store the result in the \code{reducedDims} of the output.}
}
\value{
For \code{calculateMDS}, a matrix is returned containing the MDS coordinates for each cell (row) and dimension (column).

For \code{runMDS}, a modified \code{x} is returned that contains the MDS coordinates in \code{\link{reducedDim}(x, name)}.
}
\description{
Perform multi-dimensional scaling (MDS) on cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link{cmdscale}} is used internally to compute the MDS components.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runMDS(example_sce, scale_features=NULL)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link{cmdscale}}, to perform the underlying calculations.

\code{\link[scater]{plotMDS}}, to quickly visualize the results.

\code{?"\link{scater-red-dim-args}"}, for a full description of various options.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
