% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPCA.R
\docType{methods}
\name{runPCA,SingleCellExperiment-method}
\alias{runPCA,SingleCellExperiment-method}
\title{Perform PCA on expression data}
\usage{
\S4method{runPCA}{SingleCellExperiment}(x, ncomponents = 50,
  ntop = 500, exprs_values = "logcounts", feature_set = NULL,
  scale_features = TRUE, use_coldata = FALSE,
  selected_variables = NULL, detect_outliers = FALSE,
  BSPARAM = ExactParam(), BPPARAM = SerialParam())
}
\arguments{
\item{x}{A \linkS4class{SingleCellExperiment} object.}

\item{ncomponents}{Numeric scalar indicating the number of principal components to obtain.}

\item{ntop}{Numeric scalar specifying the number of most variable features to use for PCA.}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{x} contains the expression values of interest.}

\item{feature_set}{Character vector of row names, a logical vector or a numeric vector of indices indicating a set of features to use for PCA.
This will override any \code{ntop} argument if specified.}

\item{scale_features}{Logical scalar, should the expression values be standardised so that each feature has unit variance?
This will also remove features with standard deviations below 1e-8.}

\item{use_coldata}{Deprecated, use \code{\link{runColDataPCA}} instead.}

\item{selected_variables}{Deprecated, use \code{\link{runColDataPCA}} instead.}

\item{detect_outliers}{Deprecated, use \code{\link{runColDataPCA}} instead.}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying which algorithm should be used to perform the PCA.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether the PCA should be parallelized.}
}
\value{
A SingleCellExperiment object containing the first \code{ncomponent} principal coordinates for each cell,
stored in the \code{"PCA"} entry of the \code{reducedDims} slot.
The proportion of variance explained by each PC is stored as a numeric vector in the \code{"percentVar"} attribute of the reduced dimension matrix.
}
\description{
Perform a principal components analysis (PCA) on cells, 
based on the expression data in a SingleCellExperiment object.
}
\details{
Algorithms like \code{BSPARAM=IrlbaParam()} or \code{RandomParam()} involve
a random initialization, after which it converges towards the exact PCs.
This means that the result will change slightly across different runs.
For full reproducibility, users should call \code{\link{set.seed}} prior to running \code{runPCA} with such algorithms.

In the returned output, 
the vector of proportion of variance explained may not have length equal to the total number of available PCs.
This is because not all PCA methods are guaranteed to compute singular values for all components.
As such, the proportions of variance explained - while accurate - may not sum to unity.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts),
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runPCA(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link{runPCA}}, \code{\link[scater]{plotPCA}}
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
