% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{runPCA}
\alias{runPCA}
\title{Run PCA for a SingleCellExperiment object}
\usage{
runPCA(object, ntop = 500, ncomponents = 2, exprs_values = "logcounts",
  feature_set = NULL, scale_features = TRUE, pca_data_input = "logcounts",
  selected_variables = NULL, detect_outliers = FALSE)
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the PCA. Default is \code{500}, but any \code{ntop} argument is
overrided if the \code{feature_set} argument is non-NULL.}

\item{exprs_values}{character string indicating which values should be used
as the expression values for this plot. Valid arguments are \code{"tpm"}
(transcripts per million), \code{"norm_tpm"} (normalised TPM
values), \code{"fpkm"} (FPKM values), \code{"norm_fpkm"} (normalised FPKM
values), \code{"counts"} (counts for each feature), \code{"norm_counts"},
\code{"cpm"} (counts-per-million), \code{"norm_cpm"} (normalised
counts-per-million), \code{"logcounts"} (log-transformed count data; default),
\code{"norm_exprs"} (normalised
expression values) or \code{"stand_exprs"} (standardised expression values)
or any other named element of the \code{assays} slot of the \code{SingleCellExperiment}
object that can be accessed with the \code{assay} function.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the PCA. If character, entries must all be in
\code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{pca_data_input}{character argument defining which data should be used
as input for the PCA. Possible options are \code{"logcounts"} (default), which
uses log-count data to produce a PCA at the cell level; \code{"coldata"} or
\code{"pdata"} (for backwards compatibility) which uses numeric variables
from \code{colData(object)} to do PCA at the cell level; and
\code{"rowdata"} which uses numeric variables from \code{rowData(object)} to
do PCA at the feature level.}

\item{selected_variables}{character vector indicating which variables in
\code{colData(object)} to use for the phenotype-data based PCA. Ignored if
the argument \code{pca_data_input} is anything other than \code{"pdata"}.}

\item{detect_outliers}{logical, should outliers be detected based on PCA
coordinates generated from column-level metadata? Only an option when 
\code{pca_data_input} argument is \code{"pdata"} or \code{"coldata"}. 
Default is \code{FALSE}.}
}
\description{
Perform a principal components analysis (PCA) based on the data stored in 
a \code{\link{SingleCellExperiment}} object.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

example_sce <- runPCA(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link[scater]{plotPCA}}
}
