% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runPCA.R
\docType{methods}
\name{calculatePCA}
\alias{calculatePCA}
\alias{runPCA}
\alias{calculatePCA,ANY-method}
\alias{calculatePCA,SummarizedExperiment-method}
\alias{calculatePCA,SingleCellExperiment-method}
\alias{runPCA,SingleCellExperiment-method}
\title{Perform PCA on expression data}
\usage{
calculatePCA(x, ...)

\S4method{calculatePCA}{ANY}(x, ncomponents = 50, ntop = 500,
  subset_row = NULL, feature_set = NULL, scale = FALSE,
  scale_features = NULL, transposed = FALSE, BSPARAM = ExactParam(),
  BPPARAM = SerialParam())

\S4method{calculatePCA}{SummarizedExperiment}(x, ...,
  exprs_values = "logcounts")

\S4method{calculatePCA}{SingleCellExperiment}(x, ...,
  exprs_values = "logcounts", dimred = NULL, use_dimred = NULL,
  n_dimred = NULL)

\S4method{runPCA}{SingleCellExperiment}(x, ..., use_coldata = FALSE,
  altexp = NULL, name = "PCA")
}
\arguments{
\item{x}{For \code{calculatePCA}, a numeric matrix of log-expression values where rows are features and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.

For \code{runPCA}, a \linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{...}{For the \code{calculatePCA} generic, additional arguments to pass to specific methods.
For the SummarizedExperiment and SingleCellExperiment methods, additional arguments to pass to the ANY method.

For \code{runPCA}, additional arguments to pass to \code{calculatePCA}.}

\item{ncomponents}{Numeric scalar indicating the number of principal components to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{subset_row}{Vector specifying the subset of features to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{feature_set}{Deprecated, same as \code{subset_row}.}

\item{scale}{Logical scalar, should the expression values be standardised? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{scale_features}{Deprecated, same as \code{scale} but with a different default.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{BSPARAM}{A \linkS4class{BiocSingularParam} object specifying which algorithm should be used to perform the PCA.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether the PCA should be parallelized.}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{x} contains the expression values, see \code{?"\link{scater-red-dim-args}"}.}

\item{dimred}{String or integer scalar specifying the existing dimensionality reduction results to use, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_dimred}{Deprecated, same as \code{dimred}.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if \code{dimred} is specified, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_coldata}{Deprecated, use \code{\link{runColDataPCA}} instead.}

\item{altexp}{String or integer scalar specifying an alternative experiment to use to compute the PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{name}{String specifying the name to be used to store the result in the \code{\link{reducedDims}} of the output.}
}
\value{
A SingleCellExperiment object containing the first \code{ncomponents} principal coordinates for each cell.
By default, this is stored in the \code{"PCA"} entry of the \code{\link{reducedDims}}.
The proportion of variance explained by each PC is stored as a numeric vector in the \code{"percentVar"} attribute of the reduced dimension matrix.
}
\description{
Perform a principal components analysis (PCA) on cells, 
based on the expression data in a SingleCellExperiment object.
}
\details{
Algorithms like \code{BSPARAM=IrlbaParam()} or \code{RandomParam()} involve a random initialization, after which it converges towards the exact PCs.
This means that the result will change slightly across different runs.
For full reproducibility, users should call \code{\link{set.seed}} prior to running \code{runPCA} with such algorithms.
}
\examples{
example_sce <- mockSCE()
example_sce <- logNormCounts(example_sce)

example_sce <- runPCA(example_sce, scale_features=NULL)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link[BiocSingular]{runPCA}}, for the underlying calculations. 

\code{\link[scater]{plotPCA}}, to conveniently visualize the results.

\code{?"\link{scater-red-dim-args}"}, for a full description of various options.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
