% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduced_dimensions.R
\name{runTSNE}
\alias{runTSNE}
\title{Run t-SNE for a SingleCellExperiment object}
\usage{
runTSNE(object, ntop = 500, ncomponents = 2, exprs_values = "logcounts",
  feature_set = NULL, use_dimred = NULL, n_dimred = NULL,
  scale_features = TRUE, rand_seed = NULL,
  perplexity = floor(ncol(object)/5), ...)
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the t-SNE Default is \code{500}, but any \code{ntop} argument is
overrided if the \code{feature_set} argument is non-NULL.}

\item{ncomponents}{numeric scalar indicating the number of t-SNE
components to obtain.}

\item{exprs_values}{Integer or character string indicating which values should be used #' as the expression values for this plot. 
Defaults to \code{"logcounts"}, but any other element of the \code{assays} slot of the \code{SingleCellExperiment} object can be used.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the t-SNE calculation. If character, entries must all be
in \code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{use_dimred}{character(1), use named reduced dimension representation of cells
stored in \code{SingleCellExperiment} object instead of recomputing (e.g. "PCA").
 Default is \code{NULL}, no reduced dimension values are provided to \code{Rtsne}.}

\item{n_dimred}{integer(1), number of components of the reduced dimension slot
to use. Default is \code{NULL}, in which case (if \code{use_dimred} is not \code{NULL})
all components of the reduced dimension slot are used.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{rand_seed}{(optional) numeric scalar that can be passed to
\code{set.seed} to make plots reproducible.}

\item{perplexity}{numeric scalar value defining the "perplexity parameter"
for the t-SNE plot. Passed to \code{\link[Rtsne]{Rtsne}} - see documentation
for that package for more details.}

\item{...}{Additional arguments to pass to \code{\link[Rtsne]{Rtsne}}.}
}
\value{
A \code{SingleCellExperiment} object containing the coordinates of
the first \code{ncomponent} t-SNE dimensions for each cell in the \code{"TSNE"}
entry of the \code{reducedDims} slot.
}
\description{
Perform t-stochastic neighbour embedding (t-SNE) based on the data stored in 
a \code{\link{SingleCellExperiment}} object.
}
\details{
The function \code{\link[Rtsne]{Rtsne}} is used internally to
compute the t-SNE. Note that the algorithm is not deterministic, so different
runs of the function will produce differing plots (see \code{\link{set.seed}}
to set a random seed for replicable results). The value of the
\code{perplexity} parameter can have a large effect on the resulting plot, so
it can often be worthwhile to try multiple values to find the most appealing
visualisation and to ensure that the conclusions are robust.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

example_sce <- runTSNE(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\references{
L.J.P. van der Maaten. Barnes-Hut-SNE. In Proceedings of the International Conference on Learning Representations, 2013.
}
\seealso{
\code{\link[Rtsne]{Rtsne}},
\code{\link[scater]{plotTSNE}}
}
