% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{runTSNE}
\alias{runTSNE}
\title{Run t-SNE for a SingleCellExperiment object}
\usage{
runTSNE(object, ntop = 500, ncomponents = 2, exprs_values = "logcounts",
  feature_set = NULL, use_dimred = NULL, n_dimred = NULL,
  scale_features = TRUE, rand_seed = NULL,
  perplexity = floor(ncol(object)/5), ...)
}
\arguments{
\item{object}{a \code{\link{SingleCellExperiment}} object}

\item{ntop}{numeric scalar indicating the number of most variable features to
use for the t-SNE Default is \code{500}, but any \code{ntop} argument is
overrided if the \code{feature_set} argument is non-NULL.}

\item{exprs_values}{character string indicating which values should be used
as the expression values for this plot. Valid arguments are \code{"tpm"}
(transcripts per million), \code{"norm_tpm"} (normalised TPM
values), \code{"fpkm"} (FPKM values), \code{"norm_fpkm"} (normalised FPKM
values), \code{"counts"} (counts for each feature), \code{"norm_counts"},
\code{"cpm"} (counts-per-million), \code{"norm_cpm"} (normalised
counts-per-million), \code{"logcounts"} (log-transformed count data; default),
\code{"norm_exprs"} (normalised
expression values) or \code{"stand_exprs"} (standardised expression values),
or any other named element of the \code{assayData} slot of the \code{SingleCellExperiment}
object that can be accessed with the \code{assay} function.}

\item{feature_set}{character, numeric or logical vector indicating a set of
features to use for the t-SNE calculation. If character, entries must all be
in \code{featureNames(object)}. If numeric, values are taken to be indices for
features. If logical, vector is used to index features and should have length
equal to \code{nrow(object)}.}

\item{use_dimred}{character(1), use named reduced dimension representation of cells
stored in \code{SingleCellExperiment} object instead of recomputing (e.g. "PCA").
 Default is \code{NULL}, no reduced dimension values are provided to \code{Rtsne}.}

\item{n_dimred}{integer(1), number of components of the reduced dimension slot
to use. Default is \code{NULL}, in which case (if \code{use_dimred} is not \code{NULL})
all components of the reduced dimension slot are used.}

\item{scale_features}{logical, should the expression values be standardised
so that each feature has unit variance? Default is \code{TRUE}.}

\item{rand_seed}{(optional) numeric scalar that can be passed to
\code{set.seed} to make plots reproducible.}

\item{perplexity}{numeric scalar value defining the "perplexity parameter"
for the t-SNE plot. Passed to \code{\link[Rtsne]{Rtsne}} - see documentation
for that package for more details.}

\item{...}{Additional arguments to pass to \code{\link[Rtsne]{Rtsne}}.}
}
\description{
Perform t-stochastic neighbour embedding (t-SNE) based on the data stored in 
a \code{\link{SingleCellExperiment}} object.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
assays = list(counts = sc_example_counts), colData = sc_example_cell_info)
example_sce <- normalize(example_sce)

example_sce <- runTSNE(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\seealso{
\code{\link[Rtsne]{Rtsne}},
\code{\link[scater]{plotTSNE}}
}
