% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runTSNE.R
\docType{methods}
\name{runTSNE}
\alias{runTSNE}
\alias{runTSNE,ANY-method}
\alias{runTSNE,SingleCellExperiment-method}
\title{Perform t-SNE on cell-level data}
\usage{
runTSNE(x, ...)

\S4method{runTSNE}{ANY}(x, ncomponents = 2, ntop = 500,
  subset.row = NULL, scale = FALSE, transposed = FALSE,
  perplexity = NULL, normalize = TRUE, theta = 0.5, ...,
  external.neighbors = FALSE, external_neighbors = NULL,
  BNPARAM = KmknnParam(), BPPARAM = SerialParam())

\S4method{runTSNE}{SingleCellExperiment}(x, ...,
  pca = is.null(use.dimred), use.dimred = NULL, use_dimred = NULL,
  n.dimred = NULL, n_dimred = NULL, assay.type = "logcounts",
  exprs_values = NULL, alt.exp = NULL, name = "TSNE")
}
\arguments{
\item{x}{A numeric matrix of log-expression values where rows are features and columns are cells.

Alternatively, a \linkS4class{SingleCellExperiment} object containing such a matrix.

Alternatively, if \code{transposed=TRUE}, a numeric matrix where rows are cells and columns are dimensions.}

\item{...}{For the generic, additional arguments to pass to specific methods.

For the ANY method, additional arguments to pass to \code{\link[Rtsne]{Rtsne}}.

For the SingleCellExperiment method, additional arguments to pass to the ANY method.}

\item{ncomponents}{Numeric scalar indicating the number of t-SNE dimensions to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{subset.row}{Vector specifying the subset of features to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{scale}{Logical scalar, should the expression values be standardised? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{normalize}{Logical scalar indicating if input values should be scaled for numerical precision, see \code{\link[Rtsne]{normalize_input}}.}

\item{theta}{Numeric scalar specifying the approximation accuracy of the Barnes-Hut algorithm, see \code{\link[Rtsne]{Rtsne}} for details.}

\item{external.neighbors}{Logical scalar indicating whether a nearest neighbors search should be computed externally with \code{\link{findKNN}}.}

\item{external_neighbors}{Deprecated, same as \code{external.neighbors}.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the neighbor search algorithm to use when \code{external.neighbors=TRUE}.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying how the neighbor search should be parallelized when \code{external.neighbors=TRUE}.}

\item{pca}{Logical scalar indicating whether a PCA step should be performed inside \code{\link[Rtsne]{Rtsne}}.}

\item{use.dimred}{String or integer scalar specifying the existing dimensionality reduction results to use, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_dimred}{Deprecated, same as \code{use.dimred}.}

\item{n.dimred}{Integer scalar or vector specifying the dimensions to use if \code{use.dimred} is specified, see \code{?"\link{scater-red-dim-args}"}.}

\item{n_dimred}{Deprecated, same as \code{n.dimred}.}

\item{assay.type}{Integer scalar or string indicating which assay of \code{x} contains the expression values, see \code{?"\link{scater-red-dim-args}"}.}

\item{exprs_values}{Deprecated, same as \code{assay.type}.}

\item{alt.exp}{String or integer scalar specifying an alternative experiment to use to compute the PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{name}{String specifying the name to be used to store the result in the \code{reducedDims} of the output.}

\item{feature_set}{Deprecated, same as \code{subset.row}.}

\item{scale_features}{Deprecated, same as \code{scale} but with a different default.}
}
\value{
For the ANY method, a matrix is returned containing the t-SNE coordinates for each cell (row) and dimension (column).

For the \linkS4class{SingleCellExperiment} method, a modified version of \code{x} is return that contains the t-SNE coordinates in the \code{"TSNE"} entry of the \code{\link{reducedDims}}.
}
\description{
Perform t-stochastic neighbour embedding (t-SNE) for the cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link[Rtsne]{Rtsne}} is used internally to compute the t-SNE. 
Note that the algorithm is not deterministic, so different runs of the function will produce differing results. 
Users are advised to test multiple random seeds, and then use \code{\link{set.seed}} to set a random seed for replicable results. 

The value of the \code{perplexity} parameter can have a large effect on the results.
By default, the function will set a \dQuote{reasonable} perplexity that scales with the number of cells in \code{x}.
(Specifically, it is the number of cells divided by 3, capped at a maximum of 50.)
However, it is often worthwhile to manually try multiple values to ensure that the conclusions are robust.

If \code{external.neighbors=TRUE}, the nearest neighbor search step will use a different algorithm to that in the \code{\link[Rtsne]{Rtsne}} function.
This can be parallelized or approximate to achieve greater speed for large data sets.
The neighbor search results are then used for t-SNE via the \code{\link[Rtsne]{Rtsne_neighbors}} function.

If \code{use.dimred} is specified, the PCA step of the \code{Rtsne} function is automatically turned off by default.
This presumes that the existing dimensionality reduction is sufficient such that an additional PCA is not required.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runTSNE(example_sce, scale_features=NULL)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\references{
van der Maaten LJP and Hinton GE (2008).
Visualizing High-Dimensional Data Using t-SNE.
\emph{J. Mach. Learn. Res.} 9, 2579-2605.
}
\seealso{
\code{\link[Rtsne]{Rtsne}}, for the underlying calculations.

\code{\link{plotTSNE}}, to quickly visualize the results.

\code{?"\link{scater-red-dim-args}"}, for a full description of various options.
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
