% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTSNE.R
\name{runTSNE}
\alias{runTSNE}
\title{Perform t-SNE on cell-level data}
\usage{
runTSNE(object, ncomponents = 2, ntop = 500, feature_set = NULL,
  exprs_values = "logcounts", scale_features = TRUE,
  use_dimred = NULL, n_dimred = NULL, rand_seed = NULL,
  perplexity = min(50, floor(ncol(object)/5)), pca = TRUE,
  initial_dims = 50, ...)
}
\arguments{
\item{object}{A SingleCellExperiment object.}

\item{ncomponents}{Numeric scalar indicating the number of t-SNE dimensions to obtain.}

\item{ntop}{Numeric scalar specifying the number of most variable features to use for t-SNE.}

\item{feature_set}{Character vector of row names, a logical vector or a numeric vector of indices indicating a set of features to use for t-SNE.
This will override any \code{ntop} argument if specified.}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{object} should be used to obtain the expression values for the calculations.}

\item{scale_features}{Logical scalar, should the expression values be standardised so that each feature has unit variance?}

\item{use_dimred}{String or integer scalar specifying the entry of \code{reducedDims(object)} to use as input to \code{\link[Rtsne]{Rtsne}}.
Default is to not use existing reduced dimension results.}

\item{n_dimred}{Integer scalar, number of dimensions of the reduced dimension slot to use when \code{use_dimred} is supplied.
Defaults to all available dimensions.}

\item{rand_seed}{Deprecated, numeric scalar that can be passed to \code{set.seed} to make the results reproducible.}

\item{perplexity}{Numeric scalar defining the perplexity parameter, see \code{?\link[Rtsne]{Rtsne}} for more details.}

\item{pca}{Logical scalar passed to \code{\link[Rtsne]{Rtsne}}, indicating whether an initial PCA step should be performed.
This is ignored if \code{use_dimred} is specified.}

\item{initial_dims}{Integer scalar passed to \code{\link[Rtsne]{Rtsne}}, specifying the number of principal components to be retained if \code{pca=TRUE}.}

\item{...}{Additional arguments to pass to \code{\link[Rtsne]{Rtsne}}.}
}
\value{
A SingleCellExperiment object containing the coordinates of the first \code{ncomponent} t-SNE dimensions for each cell.
This is stored in the \code{"TSNE"} entry of the \code{reducedDims} slot.
}
\description{
Perform t-stochastic neighbour embedding (t-SNE) for the cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link[Rtsne]{Rtsne}} is used internally to compute the t-SNE. 
Note that the algorithm is not deterministic, so different runs of the function will produce differing results. 
Users are advised to test multiple random seeds, and then use \code{\link{set.seed}} to set a random seed for replicable results. 

The value of the \code{perplexity} parameter can have a large effect on the results.
By default, the function will try to provide a reasonable setting, by scaling the perplexity with the number of cells until it reaches a maximum of 50.
However, it is often worthwhile to manually try multiple values to ensure that the conclusions are robust.

Setting \code{use_dimred} allows users to easily perform t-SNE on low-rank approximations of the original expression matrix (e.g., after PCA).
In such cases, arguments such as \code{ntop}, \code{feature_set}, \code{exprs_values} and \code{scale_features} will be ignored.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- normalize(example_sce)

example_sce <- runTSNE(example_sce)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\references{
L.J.P. van der Maaten. Barnes-Hut-SNE. In Proceedings of the International Conference on Learning Representations, 2013.
}
\seealso{
\code{\link[Rtsne]{Rtsne}},
\code{\link[scater]{plotTSNE}}
}
\author{
Aaron Lun, based on code by Davis McCarthy
}
