% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/runUMAP.R
\docType{methods}
\name{calculateUMAP}
\alias{calculateUMAP}
\alias{runUMAP}
\alias{calculateUMAP,ANY-method}
\alias{calculateUMAP,SummarizedExperiment-method}
\alias{calculateUMAP,SingleCellExperiment-method}
\title{Perform UMAP on cell-level data}
\usage{
calculateUMAP(x, ...)

\S4method{calculateUMAP}{ANY}(x, ncomponents = 2, ntop = 500,
  subset_row = NULL, feature_set = NULL, scale = FALSE,
  scale_features = NULL, transposed = FALSE, pca = if (transposed)
  NULL else 50, n_neighbors = 15, ..., external_neighbors = FALSE,
  BNPARAM = KmknnParam(), BPPARAM = SerialParam())

\S4method{calculateUMAP}{SummarizedExperiment}(x, ...,
  exprs_values = "logcounts")

\S4method{calculateUMAP}{SingleCellExperiment}(x, ..., pca = if
  (!is.null(dimred)) NULL else 50, exprs_values = "logcounts",
  dimred = NULL, use_dimred = NULL, n_dimred = NULL)

runUMAP(x, ..., altexp = NULL, name = "UMAP")
}
\arguments{
\item{x}{For \code{calculateUMAP}, a numeric matrix of log-expression values where rows are features and columns are cells.
Alternatively, a \linkS4class{SummarizedExperiment} or \linkS4class{SingleCellExperiment} containing such a matrix.

For \code{runTSNE}, a \linkS4class{SingleCellExperiment} object containing such a matrix.}

\item{...}{For the \code{calculateUMAP} generic, additional arguments to pass to specific methods.
For the ANY method, additional arguments to pass to \code{\link[uwot]{umap}}.
For the SummarizedExperiment and SingleCellExperiment methods, additional arguments to pass to the ANY method.

For \code{runUMAP}, additional arguments to pass to \code{calculateUMAP}.}

\item{ncomponents}{Numeric scalar indicating the number of UMAP dimensions to obtain.}

\item{ntop}{Numeric scalar specifying the number of features with the highest variances to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{subset_row}{Vector specifying the subset of features to use for PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{feature_set}{Deprecated, same as \code{subset_row}.}

\item{scale}{Logical scalar, should the expression values be standardised? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{scale_features}{Deprecated, same as \code{scale} but with a different default.}

\item{transposed}{Logical scalar, is \code{x} transposed with cells in rows? See \code{?"\link{scater-red-dim-args}"} for details.}

\item{pca}{Integer scalar specifying how many PCs should be used as input into the UMAP algorithm.
By default, no PCA is performed if the input is a dimensionality reduction result.}

\item{n_neighbors}{Integer scalar, number of nearest neighbors to identify when constructing the initial graph.}

\item{external_neighbors}{Logical scalar indicating whether a nearest neighbors search should be computed externally with \code{\link{findKNN}}.}

\item{BNPARAM}{A \linkS4class{BiocNeighborParam} object specifying the neighbor search algorithm to use when \code{external_neighbors=TRUE}.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying how the neighbor search should be parallelized when \code{external_neighbors=TRUE}.}

\item{exprs_values}{Integer scalar or string indicating which assay of \code{x} contains the expression values, see \code{?"\link{scater-red-dim-args}"}.}

\item{dimred}{String or integer scalar specifying the existing dimensionality reduction results to use, see \code{?"\link{scater-red-dim-args}"}.}

\item{use_dimred}{Deprecated, same as \code{dimred}.}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if \code{dimred} is specified, see \code{?"\link{scater-red-dim-args}"}.}

\item{altexp}{String or integer scalar specifying an alternative experiment to use to compute the PCA, see \code{?"\link{scater-red-dim-args}"}.}

\item{name}{String specifying the name to be used to store the result in the \code{reducedDims} of the output.}
}
\value{
For \code{calculateUMAP}, a matrix is returned containing the UMAP coordinates for each cell (row) and dimension (column).

For \code{runUMAP}, a modified \code{x} is returned that contains the UMAP coordinates in \code{\link{reducedDim}(x, name)}.
}
\description{
Perform uniform manifold approximation and projection (UMAP) for the cells, based on the data in a SingleCellExperiment object.
}
\details{
The function \code{\link[uwot]{umap}} is used internally to compute the UMAP. 
Note that the algorithm is not deterministic, so different runs of the function will produce differing results. 
Users are advised to test multiple random seeds, and then use \code{\link{set.seed}} to set a random seed for replicable results. 

If \code{external_neighbors=TRUE}, the nearest neighbor search is conducted using a different algorithm to that in the \code{\link[uwot]{umap}} function.
This can be parallelized or approximate to achieve greater speed for large data sets.
The neighbor search results are then used directly to create the UMAP embedding.
}
\examples{
## Set up an example SingleCellExperiment
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info
)
example_sce <- logNormCounts(example_sce)

example_sce <- runUMAP(example_sce, scale_features=NULL)
reducedDimNames(example_sce)
head(reducedDim(example_sce))
}
\references{
McInnes L, Healy J, Melville J (2018).
UMAP: uniform manifold approximation and projection for dimension reduction.
arXiv.
}
\seealso{
\code{\link[uwot]{umap}}, for the underlying calculations.

\code{\link{plotUMAP}}, to quickly visualize the results.

\code{?"\link{scater-red-dim-args}"}, for a full description of various options.
}
\author{
Aaron Lun
}
