% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{scater-vis-var}
\alias{scater-vis-var}
\title{Variable selection for visualization}
\description{
A number of \pkg{scater} functions accept a SingleCellExperiment object and extract (meta)data from it for use in a plot.
These values are then used on the x- or y-axes (e.g., \code{\link{plotColData}}) or for tuning visual parameters, e.g.,
\code{colour_by}, \code{shape_by}, \code{size_by}.
This page describes how the selection of these values can be controlled by the user,
by passing appropriate values to the arguments of the desired plotting function.
}
\section{When plotting by cells}{

Here, we assume that each visual feature of interest (e.g., point or line) corresponds to a cell in the SingleCellExperiment object \code{sce}.
We will also assume that the user wants to change the colour of each feature according to the cell (meta)data.
To do so, the user can pass to \code{colour_by}:
\itemize{
\item An unnamed character string.
This is initially assumed to be the name of a column-level metadata field.
The function will first search the column names of \code{colData(sce)}, and extract metadata for all cells if a matching field is found.
If no match is found, the function will assume that the string represents a gene name.
It will search \code{rownames(sce)} and extract gene expression values for any matching row across all cells.
Otherwise, an error is raised.
\item A named character string, where the name is either \code{"exprs"} or \code{"metadata"}.
This forces the function to only search for the string in \code{rownames(sce)} or \code{colnames(colData(sce))}, respectively.
Adding an explicit name is useful when the same field exists in both the row names and column metadata names.
\item A character vector of length greater than 1.
This will search for nested fields in \code{colData(sce)}.
For example, supplying a character vector \code{c("A", "B", "C")} will retrieve \code{colData(sce)$A$B$C}, where both \code{A} and \code{B} contain nested DataFrames.
See \code{\link{calculateQCMetrics}} with \code{compact=TRUE} for an example of how these can be constructed. 
The concatenated name \code{"A:B:C"} will be used in the legend.
\item A data frame with one column and number of rows equal to the number of cells.
This should contain values to use for visualization (in this case, for colouring by).
In this manner, the user can use new information without manually adding it to the SingleCellExperiment object.
The column name of the data frame will be used in the legend.
}

The same logic applies for other visualization parameters such as \code{shape_by} and \code{size_by}.
Other arguments may also use the same scheme, but this depends on the context; see the documentation for each function for details.
In particular, if an argument explicitly refers to a metadata field, any names for the character string will be ignored.
Similarly, a character vector of length > 1 is not allowed for an argument that explicitly refers to expression values.
}

\section{When plotting by features}{

Here, we assume that each visual feature of interest (e.g., point or line) corresponds to a feature in the SingleCellExperiment object \code{sce}.
The scheme is mostly the same as described above, with a few differences:
\itemize{
\item \code{rowData} is searched instead of \code{colData}, as we are extracting metadata for each feature.
\item When extracting expression values, the name of a single cell must be specified.
Visualization will then use the expression profile for all features in that cell.
(This tends to be a rather unusual choice for colouring.)
\item Character strings named with \code{"exprs"} will search for the string in \code{colnames(sce)}.
\item A data frame input should have number of rows equal to the number of features.
}
}

\seealso{
\code{\link{plotColData}}, 
\code{\link{plotRowData}}, 
\code{\link{plotReducedDim}}, 
\code{\link{plotExpression}}, 
\code{\link{plotPlatePosition}},
and most other plotting functions.
}
