% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCESet-methods.R
\docType{methods}
\name{setSpike}
\alias{setSpike}
\alias{setSpike,SCESet,NULL-method}
\alias{setSpike,SCESet,character-method}
\title{Set spike-in features in an SCESet object}
\usage{
\S4method{setSpike}{SCESet,`NULL`}(object) <- value
}
\arguments{
\item{object}{a \code{SCESet} object.}

\item{value}{a character vector containing the names of the feature control
sets that are spike-ins. If \code{NULL}, all spike-in information is removed.}
}
\value{
A \code{SCESet} object containing spike-in information in 
\code{featureControlInfo} and updated \code{isSpike} results.
}
\description{
Specify which feature control sets in the SCESet object are spike-ins, i.e., 
RNA of the same type and quantity added to each cell during the scRNA-seq protocol.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
pd <- new("AnnotatedDataFrame", data = sc_example_cell_info)
example_sceset <- newSCESet(countData = sc_example_counts, phenoData = pd)
example_sceset <- calculateQCMetrics(example_sceset,
                            feature_controls = list(ERCC = 1:40, Mito=41:50))
setSpike(example_sceset) <- "ERCC"
featureControlInfo(example_sceset)
isSpike(example_sceset)
}
\author{
Aaron Lun
}

