% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumCountsAcrossFeatures.R
\name{sumCountsAcrossFeatures}
\alias{sumCountsAcrossFeatures}
\title{Sum counts across a feature set}
\usage{
sumCountsAcrossFeatures(object, ids, exprs_values = "counts",
  BPPARAM = SerialParam())
}
\arguments{
\item{object}{A \linkS4class{SingleCellExperiment} object or a count matrix.}

\item{ids}{A factor specifying the set to which each feature in \code{object} belongs.}

\item{exprs_values}{A string or integer scalar specifying the assay of \code{object} containing counts, if \code{object} is a SingleCellExperiment.}

\item{BPPARAM}{A \linkS4class{BiocParallelParam} object specifying whether summation should be parallelized.}
}
\value{
A count matrix where counts for all features in the same set are summed together within each cell.
}
\description{
Create a count matrix where counts for all features in a set are summed together.
}
\details{
This function provides a convenient method for aggregating counts across multiple rows for each cell.
For example, genes with multiple mapping locations in the reference will often manifest as multiple rows with distinct Ensembl/Entrez IDs.
These counts can be aggregated into a single feature by setting the shared identifier (usually the gene symbol) as \code{ids}.

It is theoretically possible to aggregate transcript-level counts to gene-level counts with this function.
However, it is often better to do so with dedicated functions (e.g., from the \pkg{tximport} or \pkg{tximeta} packages) that account for differences in length across isoforms.

Any \code{NA} values in \code{ids} are implicitly ignored and will not be considered or reported.
This may be useful, e.g., to remove undesirable feature sets by setting their entries in \code{ids} to \code{NA}.
}
\examples{
data("sc_example_counts")
data("sc_example_cell_info")
example_sce <- SingleCellExperiment(
    assays = list(counts = sc_example_counts), 
    colData = sc_example_cell_info)

ids <- sample(LETTERS, nrow(example_sce), replace=TRUE)
out <- sumCountsAcrossFeatures(example_sce, ids)
dimnames(out)
}
\author{
Aaron Lun
}
